Boostrap for eigenvalues.

Boostrap for eigenvalues.

Usage

Boot4Eigs(
  data,
  center = FALSE,
  scale = FALSE,
  design = NULL,
  niter = 100,
  CI.perc = c(0.025, 0.975),
  suppressProgressBar = TRUE
)

Arguments

data

the data-set;

center

whether to center the data (default to FALSE), see expo.scale;

scale

whether to scale the data ( default to FALSE), see expo.scale;

design

whether there is a design;

niter

number of bootstrap iterations;

CI.perc

vector of the two levels of confidence for the left and right limits of the bootstrap confidence interval (default to c(0.025, 0.0975));

suppressProgressBar

whether to show a progress bar or not (default to FALSE).

Value

list containing four objects

Examples

Boot4Eigs(matrix(rnorm(100), 10, 10))
#> $BootMatEigs
#>        Dimension 1 Dimension 2 Dimension 3  Dimension 4  Dimension 5
#>   [1,]    31.23051    20.12630   14.994277 1.130448e+01 8.362102e+00
#>   [2,]    32.95875    20.53138   18.607352 1.365131e+01 3.120605e+00
#>   [3,]    38.74701    21.34357   18.285424 7.741186e+00 3.472760e+00
#>   [4,]    46.83114    20.24093   14.237414 5.885391e+00 2.805122e+00
#>   [5,]    35.91116    20.50169   14.969895 1.054954e+01 7.019230e+00
#>   [6,]    40.20320    21.93853   17.685467 1.017281e+01 4.818644e-15
#>   [7,]    31.21205    22.02385   15.138277 1.166358e+01 7.442403e+00
#>   [8,]    42.46889    22.49285   13.891907 7.719880e+00 2.048191e+00
#>   [9,]    36.24200    21.79107   15.753977 1.107296e+01 3.123265e+00
#>  [10,]    41.36325    23.58391   16.431537 8.621302e+00 7.192327e-15
#>  [11,]    35.21767    20.52763   17.011466 8.812187e+00 6.062227e+00
#>  [12,]    37.48161    21.03972   15.566746 9.775351e+00 4.690028e+00
#>  [13,]    34.08367    28.06736   12.104744 8.278763e+00 5.796972e+00
#>  [14,]    33.65686    23.17188   17.788217 1.128313e+01 3.210573e+00
#>  [15,]    36.99788    23.13889   16.286657 1.130968e+01 2.266896e+00
#>  [16,]    34.28573    29.96831   15.971148 7.914207e+00 1.860604e+00
#>  [17,]    34.53962    19.27189   16.701787 1.129283e+01 5.855495e+00
#>  [18,]    40.37155    23.60899   15.815459 8.096920e+00 2.107080e+00
#>  [19,]    39.65797    21.92617   13.684060 8.962654e+00 3.237760e+00
#>  [20,]    36.51573    22.65176   17.556754 8.200962e+00 5.074795e+00
#>  [21,]    44.72511    21.35457   15.322677 8.597638e+00 1.741384e-15
#>  [22,]    36.49542    20.54759   13.767604 8.800044e+00 4.746030e+00
#>  [23,]    31.98182    19.10086   15.838482 1.105690e+01 7.402406e+00
#>  [24,]    37.17360    20.94233   15.638563 8.677377e+00 5.281683e+00
#>  [25,]    28.53188    24.36693   21.299400 1.002370e+01 4.397431e+00
#>  [26,]    35.41611    20.89252   17.096681 9.844347e+00 4.696194e+00
#>  [27,]    34.16289    26.76958   15.606513 1.119721e+01 2.263811e+00
#>  [28,]    44.96628    18.53246   14.118042 7.847520e+00 4.535691e+00
#>  [29,]    34.73473    27.76707   19.393747 6.101358e+00 2.003097e+00
#>  [30,]    32.87646    24.13895   16.664704 1.151148e+01 3.358998e+00
#>  [31,]    27.81018    22.15518   15.218121 1.197392e+01 8.273027e+00
#>  [32,]    44.47736    16.95205   14.619693 1.063861e+01 3.312283e+00
#>  [33,]    33.46787    25.04480   19.179664 6.243454e+00 3.304819e+00
#>  [34,]    32.79130    24.36401   16.425103 9.315880e+00 4.412238e+00
#>  [35,]    29.03920    21.01363   17.916586 1.087481e+01 7.325031e+00
#>  [36,]    43.93089    24.69412   11.878167 8.102708e+00 1.394116e+00
#>  [37,]    38.90295    25.07155   16.700437 7.232246e+00 1.461926e+00
#>  [38,]    33.07074    22.11359   17.161160 1.140896e+01 5.697506e+00
#>  [39,]    29.15856    23.80879   20.761036 9.445088e+00 5.180621e+00
#>  [40,]    43.95126    21.20683   15.139217 5.630694e+00 4.072001e+00
#>  [41,]    31.46225    23.03726   15.544374 1.232629e+01 4.151617e+00
#>  [42,]    36.29654    23.72161   13.646551 9.488536e+00 6.846768e+00
#>  [43,]    32.39227    22.60582   18.940224 1.111845e+01 4.943234e+00
#>  [44,]    37.07604    31.32033   12.901798 5.701008e+00 3.000826e+00
#>  [45,]    43.07587    22.63405   15.315298 6.192111e+00 1.987080e+00
#>  [46,]    41.56885    22.23731   15.778672 1.041517e+01 5.430219e-15
#>  [47,]    34.02161    20.10811   13.874731 1.115320e+01 8.788023e+00
#>  [48,]    36.61450    26.31351   17.639758 6.016182e+00 3.416047e+00
#>  [49,]    33.78364    20.58479   15.859004 1.155551e+01 6.387915e+00
#>  [50,]    45.00750    27.10227   17.890230 3.485861e-15 2.471348e-15
#>  [51,]    37.07901    22.72450   11.928437 1.082978e+01 5.843409e+00
#>  [52,]    36.89595    29.11784   15.174351 6.773708e+00 2.038156e+00
#>  [53,]    28.64145    25.68935   19.975102 9.738771e+00 4.329519e+00
#>  [54,]    49.91049    23.43068   16.658824 6.405342e-15 7.542714e-16
#>  [55,]    30.29221    26.97651   15.541348 1.102287e+01 4.658634e+00
#>  [56,]    36.91447    27.03947   17.469630 6.624451e+00 1.951985e+00
#>  [57,]    41.19815    37.73362   11.068233 3.428005e-14 2.539956e-15
#>  [58,]    37.42457    25.66467   14.174080 7.354107e+00 4.704886e+00
#>  [59,]    29.92696    27.40548   18.607668 8.597627e+00 4.128149e+00
#>  [60,]    40.89616    18.55594   14.119753 9.228247e+00 5.885666e+00
#>  [61,]    38.34902    25.92399   19.826536 5.900454e+00 3.932585e-15
#>  [62,]    39.41112    19.73911   15.372808 1.177832e+01 3.698637e+00
#>  [63,]    42.51743    25.78256   15.581434 6.118573e+00 2.089490e-15
#>  [64,]    36.32613    24.88049   18.320246 6.762983e+00 3.103303e+00
#>  [65,]    38.67866    23.14308   19.152615 9.025645e+00 3.841658e-15
#>  [66,]    45.73249    24.24235   15.622269 4.402886e+00 2.745938e-15
#>  [67,]    38.17230    25.12837   14.797415 1.190192e+01 1.806227e-16
#>  [68,]    30.83133    27.96570   21.816910 9.386061e+00 2.622534e-14
#>  [69,]    35.62995    20.89381   15.631645 1.340310e+01 2.666239e+00
#>  [70,]    32.51490    19.80514   19.208229 8.769025e+00 5.259860e+00
#>  [71,]    36.63389    27.89393   16.387039 7.084461e+00 2.000686e+00
#>  [72,]    34.88965    29.84129   17.782551 7.486509e+00 5.119385e-15
#>  [73,]    34.91790    25.35793   20.560661 6.215902e+00 2.947607e+00
#>  [74,]    33.10791    27.58927   15.557671 8.258740e+00 5.486412e+00
#>  [75,]    51.28107    27.74064    7.401451 2.759731e+00 8.171075e-01
#>  [76,]    36.26325    22.95524   17.751861 8.025180e+00 3.566014e+00
#>  [77,]    30.04294    28.83478   15.250708 1.031926e+01 5.552312e+00
#>  [78,]    40.09905    22.95665   16.101933 6.650398e+00 4.191972e+00
#>  [79,]    31.47489    23.02556   18.248336 1.312107e+01 4.130155e+00
#>  [80,]    42.34878    24.54112   10.962743 5.484498e+00 4.939858e+00
#>  [81,]    34.77640    23.02087   16.634312 8.258438e+00 4.803459e+00
#>  [82,]    32.98327    19.37245   15.662749 1.181168e+01 8.346849e+00
#>  [83,]    34.77866    21.72182   15.241474 1.021841e+01 7.052101e+00
#>  [84,]    34.10666    25.01983   13.327048 9.785191e+00 7.761269e+00
#>  [85,]    34.42626    24.22158   13.465588 9.951144e+00 7.935436e+00
#>  [86,]    31.42791    26.56504   17.511252 8.346027e+00 4.127613e+00
#>  [87,]    37.03221    20.29830   15.521404 1.055103e+01 4.372773e+00
#>  [88,]    39.28998    25.85633   13.810486 1.104320e+01 3.231293e-15
#>  [89,]    36.47933    27.88553   14.018201 8.974093e+00 2.642847e+00
#>  [90,]    34.07621    23.69308   16.738422 8.959791e+00 4.283161e+00
#>  [91,]    30.94520    22.12674   17.314282 9.169828e+00 6.713099e+00
#>  [92,]    33.70469    21.47557   16.001000 9.166243e+00 7.701368e+00
#>  [93,]    28.21806    25.17885   19.143735 1.258560e+01 4.873755e+00
#>  [94,]    42.00944    27.26741   12.498791 5.399935e+00 2.824426e+00
#>  [95,]    32.07691    25.41462   16.114420 9.239761e+00 5.799075e+00
#>  [96,]    33.99843    23.58935   18.714317 1.181769e+01 1.880214e+00
#>  [97,]    33.77709    20.13963   13.631302 1.003700e+01 8.269945e+00
#>  [98,]    37.35508    28.56236   16.709036 4.942178e+00 2.431344e+00
#>  [99,]    41.18341    26.30357   13.522952 8.990071e+00 2.579128e-15
#> [100,]    41.66487    22.76118   18.430328 7.143624e+00 1.081524e-15
#>          Dimension 6   Dimension 7   Dimension 8   Dimension 9  Dimension 10
#>   [1,]  2.562936e+00  1.419398e+00  1.118497e-15 -2.275252e-15 -5.647093e-15
#>   [2,]  1.130596e+00  1.396464e-15  8.658677e-16 -4.207082e-16 -5.089127e-15
#>   [3,]  4.100519e-01  5.602665e-15  1.465310e-16 -9.903514e-16 -1.709693e-15
#>   [4,]  1.631245e-15  1.587638e-15  1.552140e-15 -9.434341e-16 -5.477321e-15
#>   [5,]  1.048488e+00  5.167815e-15  1.637089e-15 -1.693317e-15 -1.928970e-15
#>   [6,]  2.968059e-16 -1.590750e-15 -2.216110e-15 -2.306699e-15 -3.064379e-15
#>   [7,]  2.519836e+00  1.863102e-15  1.444293e-15 -4.989559e-16 -1.628080e-15
#>   [8,]  1.378286e+00  5.418581e-15 -3.791413e-16 -1.169569e-15 -4.629346e-15
#>   [9,]  2.016734e+00  5.517881e-15  3.162282e-15  2.757226e-15 -4.334440e-16
#>  [10,]  1.590155e-15  9.442590e-16  6.664072e-16  3.788212e-16 -1.637593e-15
#>  [11,]  1.982516e+00  3.863049e-01  4.936396e-15  1.256542e-15 -6.988283e-16
#>  [12,]  1.446545e+00  2.339734e-15  4.486484e-17  1.583800e-17 -5.390123e-15
#>  [13,]  1.668491e+00  2.419306e-15  3.896211e-16 -2.683794e-16 -1.296316e-15
#>  [14,]  8.893421e-01  9.499218e-15  3.199358e-16 -7.587708e-16 -1.660886e-15
#>  [15,]  3.784379e-15  3.164019e-15  3.950275e-16 -1.477545e-15 -4.321027e-15
#>  [16,]  3.117915e-15  1.758278e-15 -6.877460e-16 -1.114636e-15 -3.982335e-15
#>  [17,]  2.338377e+00  9.275330e-16  1.986411e-16  1.536095e-16 -3.552622e-15
#>  [18,]  7.597171e-15  6.973958e-16 -5.130602e-16 -5.989256e-16 -1.269670e-15
#>  [19,]  2.531382e+00  5.531690e-15  1.590386e-15  9.737349e-16 -1.336026e-15
#>  [20,]  2.939611e-15  5.942991e-16 -1.272374e-15 -2.232781e-15 -4.549428e-15
#>  [21,]  1.604736e-15 -2.891525e-16 -1.183266e-15 -1.327792e-15 -9.258156e-15
#>  [22,]  4.475400e+00  1.167915e+00  3.990336e-15  5.781485e-16  2.735464e-16
#>  [23,]  4.619535e+00  6.512851e-15  1.820917e-15 -1.300812e-15 -2.556630e-15
#>  [24,]  2.286442e+00  4.844607e-15 -6.362867e-16 -1.193315e-15 -1.836397e-15
#>  [25,]  1.380662e+00  3.751940e-15  1.421069e-15  1.985835e-16 -7.356647e-16
#>  [26,]  1.520646e+00  5.335069e-01  2.261983e-15  7.282759e-16 -1.560943e-15
#>  [27,]  2.040922e-15  4.744908e-16 -1.714523e-15 -2.560627e-15 -3.086391e-15
#>  [28,]  8.038478e-15  3.786844e-15  1.689284e-15 -2.503248e-16 -1.921887e-15
#>  [29,]  7.952133e-15  9.938688e-16  8.695844e-16  2.190662e-16  2.847315e-18
#>  [30,]  1.449402e+00  1.985760e-15  1.832625e-15 -5.937934e-16 -1.058900e-15
#>  [31,]  2.813499e+00  1.756077e+00  7.673509e-16  3.231407e-16 -4.289106e-15
#>  [32,]  1.278763e-14  5.643656e-16 -1.877063e-16 -3.624362e-16 -9.116920e-16
#>  [33,]  2.759391e+00  4.173953e-15  2.450864e-15  1.069427e-15 -1.343987e-15
#>  [34,]  2.691470e+00  2.019168e-15  3.681360e-16 -3.958907e-16 -2.626400e-15
#>  [35,]  2.551025e+00  1.279723e+00  6.465074e-15  8.883763e-17 -6.169306e-16
#>  [36,]  2.285293e-15  8.278440e-16 -4.912917e-16 -1.227306e-15 -2.113787e-15
#>  [37,]  6.308849e-01  2.713813e-15 -9.618937e-16 -2.007212e-15 -2.268568e-15
#>  [38,]  5.480424e-01  1.119294e-15  6.413757e-16 -8.458073e-16 -3.011971e-15
#>  [39,]  1.645902e+00  1.075581e-15 -9.029485e-16 -3.189526e-15 -5.382207e-15
#>  [40,]  7.669489e-15  7.238435e-16  2.059425e-16 -3.321181e-15 -5.004126e-15
#>  [41,]  3.478208e+00  1.632090e-15  1.030986e-15 -1.362148e-15 -1.489606e-15
#>  [42,]  1.627831e-15  1.321015e-15 -7.880203e-16 -1.194770e-15 -8.777588e-15
#>  [43,]  1.564478e-15  3.086897e-16 -9.577076e-16 -1.457930e-15 -2.742788e-15
#>  [44,]  4.872517e-15  9.382023e-16 -3.385924e-16 -1.222022e-15 -5.703033e-15
#>  [45,]  7.955858e-01  3.259728e-15  3.617183e-16 -7.670782e-17 -2.322551e-16
#>  [46,]  2.131151e-15 -6.145570e-16 -1.072573e-15 -1.185297e-15 -4.412762e-15
#>  [47,]  2.054327e+00 -1.394545e-18 -5.083687e-16 -6.779619e-16 -5.629215e-15
#>  [48,]  4.342927e-15  2.711460e-15  1.827713e-15  1.465105e-15  7.230624e-16
#>  [49,]  1.829147e+00  4.236659e-15 -1.064056e-15 -1.271861e-15 -1.665847e-15
#>  [50,]  1.199491e-15  7.895212e-16  1.259347e-16 -3.450617e-16 -8.057296e-15
#>  [51,]  1.594863e+00  2.433232e-15  7.174288e-16 -1.353060e-15 -6.001634e-15
#>  [52,]  8.208673e-16 -4.743713e-16 -9.259977e-16 -1.573606e-15 -5.715114e-15
#>  [53,]  1.625809e+00  1.404731e-15  1.284116e-15 -9.020761e-16 -5.265351e-15
#>  [54,]  3.951928e-16  1.915092e-16 -5.231118e-16 -6.422502e-16 -1.117557e-15
#>  [55,]  1.508426e+00  1.601929e-15  1.263014e-15 -6.748919e-16 -4.788691e-15
#>  [56,]  4.793518e-15 -6.738414e-16 -8.714981e-16 -2.595412e-15 -3.427648e-15
#>  [57,]  9.850840e-17  2.804867e-18 -6.836061e-16 -1.537599e-15 -1.555868e-15
#>  [58,]  6.776823e-01  2.304805e-15  1.635470e-15  1.228284e-15  3.420149e-16
#>  [59,]  1.334108e+00  1.850429e-15  1.710273e-15  7.334730e-16  4.483949e-16
#>  [60,]  1.314232e+00  2.205346e-15  8.143622e-16 -1.114459e-15 -2.177240e-15
#>  [61,]  6.594775e-16  3.980140e-16 -5.962849e-16 -2.008609e-15 -4.119144e-15
#>  [62,]  5.880775e-15  2.050130e-15  6.864585e-16  1.188558e-17 -1.081720e-18
#>  [63,]  8.979652e-16  8.640875e-16  3.830870e-16 -8.194760e-16 -3.854928e-15
#>  [64,]  6.068489e-01  3.584341e-16 -5.658741e-17 -2.133758e-16 -1.381214e-15
#>  [65,]  1.851744e-15  1.600979e-15  1.596617e-15  3.454683e-16 -8.667492e-15
#>  [66,]  1.554395e-15  1.434172e-15  1.196308e-15 -6.119339e-16 -5.714077e-15
#>  [67,] -1.431856e-15 -1.700821e-15 -2.042950e-15 -2.288768e-15 -6.953270e-15
#>  [68,]  1.313039e-14  3.170407e-15  1.312683e-15 -3.608209e-16 -1.852366e-15
#>  [69,]  1.775259e+00  1.092062e-15  6.765597e-16 -4.093059e-16 -3.813079e-15
#>  [70,]  2.830036e+00  1.612805e+00  5.210312e-15  6.504707e-17 -1.990075e-15
#>  [71,]  2.358109e-15  2.063582e-16 -2.005812e-15 -6.103167e-15 -6.270416e-15
#>  [72,] -1.429511e-15 -1.665142e-15 -1.811999e-15 -2.019251e-15 -2.611872e-15
#>  [73,]  5.508886e-15  1.544059e-15 -2.290959e-16 -7.488019e-16 -1.990243e-15
#>  [74,]  7.576486e-15  6.644604e-17 -5.025384e-16 -9.449967e-16 -2.459034e-15
#>  [75,]  2.169216e-15  2.085869e-15  1.013937e-15 -4.465959e-16 -8.712990e-15
#>  [76,]  1.438462e+00  3.539356e-15  1.379004e-16 -1.453773e-15 -4.007875e-15
#>  [77,]  2.604351e-15  2.437363e-15  1.107020e-15 -1.214982e-15 -5.474038e-15
#>  [78,]  9.067893e-16  1.827074e-16 -1.676433e-15 -3.817478e-15 -8.482416e-15
#>  [79,]  2.149190e-15  4.643753e-16 -6.280574e-16 -1.771534e-15 -2.690640e-15
#>  [80,]  1.723008e+00  1.048432e-14 -4.562669e-16 -9.248452e-16 -1.341529e-15
#>  [81,]  2.506524e+00  5.477609e-15  2.042748e-15  1.131836e-15 -6.002611e-16
#>  [82,]  1.823011e+00  3.257615e-15  2.114617e-15  1.880166e-15 -1.434208e-15
#>  [83,]  9.875212e-01  2.559207e-15 -1.818923e-15 -1.889217e-15 -4.311460e-15
#>  [84,]  2.680605e-15  1.897511e-15  1.259343e-15  7.343649e-16 -4.027380e-15
#>  [85,]  5.252440e-16  2.189130e-16 -5.801243e-16 -6.466176e-16 -7.015764e-15
#>  [86,]  2.022165e+00  2.727751e-15  4.786453e-16  2.660328e-16 -2.162313e-15
#>  [87,]  2.224284e+00  5.019017e-15 -1.154878e-15 -1.218156e-15 -2.059419e-15
#>  [88,]  1.320268e-15 -2.220355e-15 -3.864477e-15 -3.924470e-15 -4.154547e-15
#>  [89,]  2.839163e-15  2.054927e-15 -4.497938e-16 -1.085254e-15 -1.866024e-15
#>  [90,]  2.249336e+00  4.912646e-15  3.407084e-15  8.133239e-17 -2.181108e-15
#>  [91,]  1.893071e+00  1.299229e+00  5.385554e-01  1.332268e-15  2.622706e-16
#>  [92,]  1.951133e+00  2.380452e-15 -1.015764e-16 -1.371271e-16 -1.732120e-15
#>  [93,]  8.409049e-15  1.585701e-15  7.022217e-16  6.658149e-16 -1.145166e-16
#>  [94,]  8.172730e-16  3.629609e-16  3.008811e-16 -4.049326e-16 -5.861736e-15
#>  [95,]  1.355211e+00  3.491489e-15  7.105427e-18 -1.093318e-15 -2.577524e-15
#>  [96,]  2.720624e-15  1.341264e-15  1.714681e-16  1.422273e-16  3.445408e-17
#>  [97,]  3.367867e+00  7.771651e-01  1.639714e-15 -8.194429e-16 -1.195625e-15
#>  [98,]  4.893829e-15 -1.708099e-16 -8.691145e-16 -1.595908e-15 -1.653379e-15
#>  [99,] -1.531502e-16 -7.163967e-16 -2.177460e-15 -2.307315e-15 -3.700238e-15
#> [100,]  2.998867e-16 -4.659270e-16 -6.575391e-16 -1.553782e-15 -9.462201e-15
#> 
#> $BootMatEigsCI
#>                      LowCI       HighCI
#> Dimension 1   2.821806e+01 4.683114e+01
#> Dimension 2   1.853246e+01 2.996831e+01
#> Dimension 3   1.096274e+01 2.076104e+01
#> Dimension 4   6.405342e-15 1.312107e+01
#> Dimension 5   7.542714e-16 8.346849e+00
#> Dimension 6  -1.429511e-15 3.478208e+00
#> Dimension 7  -1.700821e-15 1.419398e+00
#> Dimension 8  -2.216110e-15 5.210312e-15
#> Dimension 9  -3.924470e-15 1.465105e-15
#> Dimension 10 -9.258156e-15 3.420149e-16
#> 
#> $FixedEigs
#>  [1] 25.2608911 17.1801627 12.7154616  9.4792985  6.1474342  3.3859837
#>  [7]  1.2764143  1.0553026  0.2161115  0.0112315
#> 
#> $BootstrappedEstEigs
#>   Dimension 1   Dimension 2   Dimension 3   Dimension 4   Dimension 5 
#>  3.651652e+01  2.381502e+01  1.605682e+01  8.802147e+00  3.739158e+00 
#>   Dimension 6   Dimension 7   Dimension 8   Dimension 9  Dimension 10 
#>  9.626201e-01  1.023212e-01  5.385554e-03 -7.852116e-16 -3.140566e-15 
#>