Core function to generate two-table analysis plots.

Core function to generate two-table analysis plots.

Usage

TTAplotInference(
  res,
  tab1,
  tab2,
  leDim = c(1, 1),
  leDim4CirCor = c(1, 2),
  color.obs = NULL,
  color.tab = NULL,
  title.plot = "Results",
  alpha.points = 0.25,
  DESIGN = NULL,
  sigbar.p = FALSE,
  sigbar.q = FALSE,
  tab1.name = "J-set",
  tab2.name = "K-set",
  TI = FALSE,
  mean.cex = 3,
  mean.textcex = 3,
  only.mean = FALSE,
  only.ind = FALSE,
  score.constraints = NULL,
  mean.constraints = NULL,
  scale.mean.constraints = 3,
  save2pptx = FALSE,
  title4pptx = "TTA Results"
)

Arguments

res

the result of a two-table analysis;

tab1

the result of a two-table analysis;

tab2

the result of a two-table analysis;

leDim

vector containning the dimensions to be plotted;

leDim4CirCor

vector containning the dimensions to be correlated with the latent variables for the circle of correlation plot;

color.obs

list containing the colors for the observations;

color.tab

list containing the colors for the two-tables;

title.plot

plot of the title;

alpha.points

opacity for points;

DESIGN

design for the observations;

sigbar.p

whether to plot only significant bars or not for the contribution plot of the first block (default to FALSE);

sigbar.q

whether to plot only significant bars or not for the contribution plot of the second block (default to FALSE);

tab1.name

the name of the first table (default to "1");

tab2.name

the name of the second table (default to "2");

TI

whether to plot toerance intervals (default to FALSE);

mean.cex

size expansion for the mean symbols;

mean.textcex

size expansion for the mean symbols;

only.mean

whether to plot only means;

only.ind

whether to plot only observations on the latent variable plot (default to FALSE);

score.constraints

constraints on the plot (use with care);

mean.constraints

constraints on the plot (use with care);

scale.mean.constraints

constraints on the plot (use with care).

save2pptx

Default: FALSE

title4pptx

Title of the PPTX, Default: 'TTA Results'.

Value

a list of plots

Examples

if (FALSE) {
  # An example will be added very soon
}