graph4epPCA run a PCA (with the ExPosition package) and generates the standard graphs and tables. Note: Still Under Development.

graph4epPCA graph4epPCA run a principal component analysis (with the ExPosition package) and generates the standard graphs and tables. Note that the parameters data, scale, center, DESIGN, make_design_nominal, k are passed to the ExPosition package unchanged except for scale which defaults now to 'SS1'.

graph4epPCA graph4epPCA run a principal component analysis (with the ExPosition package) and generates the standard graphs and tables. Note that the parameters data, scale, center, DESIGN, make_design_nominal, k are passed to the ExPosition package unchanged except for scale which defaults now to 'SS1'.

graph4epPCA graph4epPCA run a principal component analysis (with the ExPosition package) and generates the standard graphs and tables. Note that the parameters data, scale, center, DESIGN, make_design_nominal, k are passed to the ExPosition package unchanged except for scale which defaults now to 'SS1'.

Usage

OTAplot(
  resPCA,
  data,
  DESIGN = NULL,
  make_design_nominal = TRUE,
  k = 0,
  graphs = 12,
  printGraphs = FALSE,
  col4I = NULL,
  col4J = NULL,
  biplot = FALSE,
  rotation = FALSE,
  nfactor4rotation = "Kaiser",
  save2pptx = FALSE,
  title4pptx = "PCA Results"
)

OTAplotInference(
  resPCA,
  data,
  DESIGN = NULL,
  make_design_nominal = TRUE,
  k = 0,
  graphs = 12,
  printGraphs = FALSE,
  col4I = NULL,
  col4J = NULL,
  biplot = FALSE,
  rotation = FALSE,
  nfactor4rotation = "Kaiser",
  niter.boot = 100,
  niter.perm = 100,
  save2pptx = FALSE,
  title4pptx = "PCA Inference Results"
)

graph4epPCA(
  data,
  scale = TRUE,
  center = TRUE,
  DESIGN = NULL,
  make_design_nominal = TRUE,
  k = 0,
  graphs = 12,
  printGraphs = FALSE,
  col4I = NULL,
  col4J = NULL,
  biplot = FALSE,
  rotation = FALSE,
  nfactor4rotation = "Kaiser",
  inferences = FALSE,
  save2pptx = "",
  title4pptx = "PCA Results"
)

Arguments

resPCA

Output from epPCA

data

A data frame or a matrix with numerical data suitable for a PCA. Passed to ExPosition::epPCA.

DESIGN

Default: NULL. A design vector (could be factor or character) or (Boolean) matrix used to assigne oobservations to groups. Passed to ExPosition::epPCA.

make_design_nominal

if TRUE (Default) transform the vector from DESIGN into a Boolean matrix. Passed to ExPosition::epPCA.

k

number of factor to keep; when equql to 0 (Default), all factors are kept. Passed to ExPosition::epPCA.

graphs

do we want graphs? Current Default is 12 which indicates that the graphs are generated for the first 2 components. Note that current version is creating output only for the first two components,

printGraphs

(Default: FALSE) do we want to print the graphics as .png?

col4I

a color vector for plotting the observations (if NULL Default) use colors from ExPosition::epPCA.

col4J

a color vector for plotting the variables (if NULL Default) use colors from ExPosition::epPCA.

biplot

De we want to create biplots (Default: FALSE)?

rotation

Do we want to rotate (with varimax) the variables (Default: FALSE)

nfactor4rotation

number of factors to keep for the rotation, could be a number (note if the number is too big nfactor4rotation will default to 2), or a name (currently only 'Kaiser' which is the default). When 'Kaiser' is chosen, we keep only the components with an eigenvalue larger than average (when scaled == 'SS1' this is the familiar rule: "keep only the components with eigenvalue larger than 1").

save2pptx

Default: '' Not yet implemented,

title4pptx

PARAM_DESCRIPTION, Default: 'PCA Results'. Not yet implemented,.

scale

scale (i.e., normalize) the columns of data. Default: TRUE. Values could be TRUE, FALSE, 'Z', 'SS1'. TRUE defaults to SS1. Passed to ExPosition::epPCA.

center

Default: TRUE. do we center the columns of data Passed to ExPosition::epPCA.

inferences

Run inferences from InPosition Remains to be done, Default is FALSE.

Value

A list made of two lists

A list made of two lists

A list made of two lists

Details

Work in Progress

Work in Progress

Work in Progress

See also

Author

Hervé Abdi

Examples

if (FALSE) {
if(interactive()){
 # Example from data4PCCAR
 data("sixBeers12Descriptors10Judges", package = 'data4PCCAR')
 df <- sixBeers12Descriptors10Judges$ratingsIntensity
 res4graph <- graph4epPCA(data = df, scale = FALSE)
 }
}
if (FALSE) {
if(interactive()){
 # Example from data4PCCAR
 data("sixBeers12Descriptors10Judges", package = 'data4PCCAR')
 df <- sixBeers12Descriptors10Judges$ratingsIntensity
 res4graph <- graph4epPCA(data = df, scale = FALSE)
 }
}
if (FALSE) {
if(interactive()){
 # Example from data4PCCAR
 data("sixBeers12Descriptors10Judges", package = 'data4PCCAR')
 df <- sixBeers12Descriptors10Judges$ratingsIntensity
 res4graph <- graph4epPCA(data = df, scale = FALSE)
 }
}