Boostrap for eigenvalues.

Boostrap for eigenvalues.

Usage

Boot4Eigs(
  data,
  center = FALSE,
  scale = FALSE,
  design = NULL,
  niter = 100,
  CI.perc = c(0.025, 0.975),
  suppressProgressBar = TRUE
)

Arguments

data

the data-set;

center

whether to center the data (default to FALSE), see expo.scale;

scale

whether to scale the data ( default to FALSE), see expo.scale;

design

whether there is a design;

niter

number of bootstrap iterations;

CI.perc

vector of the two levels of confidence for the left and right limits of the bootstrap confidence interval (default to c(0.025, 0.0975));

suppressProgressBar

whether to show a progress bar or not (default to FALSE).

Value

list containing four objects

Examples

Boot4Eigs(matrix(rnorm(100), 10, 10))
#> $BootMatEigs
#>        Dimension 1 Dimension 2 Dimension 3 Dimension 4  Dimension 5
#>   [1,]    43.14252    22.62405   14.728263    8.915980 7.053243e+00
#>   [2,]    39.79944    19.85417   17.445286   10.086525 6.102998e+00
#>   [3,]    37.74162    32.58238   14.824908    8.978680 1.976165e+00
#>   [4,]    27.08109    20.27917   17.572887   11.161853 8.755615e+00
#>   [5,]    28.93144    21.54911   21.444049   14.686543 6.012466e+00
#>   [6,]    36.46665    25.30026   16.143336   10.666889 7.134391e+00
#>   [7,]    38.39893    19.80760   15.642592   13.026953 1.970482e+00
#>   [8,]    27.68840    26.31958   15.691833   13.622981 6.099195e+00
#>   [9,]    30.12236    17.24558    9.570729    7.933907 4.832554e+00
#>  [10,]    44.55551    28.50582   11.403533    8.000592 6.990699e+00
#>  [11,]    31.66866    21.21683   14.557964   13.087691 6.199120e+00
#>  [12,]    47.61035    22.75415   14.034623    7.587657 3.842517e+00
#>  [13,]    36.37357    24.47816   11.422278    4.991755 3.196470e+00
#>  [14,]    32.81506    26.64825   22.564682   15.290321 6.944457e+00
#>  [15,]    37.25639    24.84935   16.274497    8.768936 5.927355e+00
#>  [16,]    31.39292    20.12761   17.475237    9.709714 7.825822e+00
#>  [17,]    40.89255    29.94201   20.036021    9.463413 3.936280e+00
#>  [18,]    42.77657    24.57310   19.905746    7.093632 4.228323e+00
#>  [19,]    61.78624    22.73218   12.389591    6.143645 4.226068e+00
#>  [20,]    47.79544    32.90313   13.273382    4.381192 2.703442e+00
#>  [21,]    29.74443    21.71369   16.772534    8.558381 5.694375e+00
#>  [22,]    54.29954    28.99047   17.125175    7.844812 4.904896e+00
#>  [23,]    36.73160    23.99005   19.145940   11.008662 7.657483e+00
#>  [24,]    38.00934    23.74312   16.985216    9.293287 5.270130e+00
#>  [25,]    34.56232    22.17045   15.343739    6.632465 4.246256e+00
#>  [26,]    52.19915    28.57231   13.659430    7.476243 5.842307e+00
#>  [27,]    38.71474    22.07132   18.746330   11.300816 4.233652e+00
#>  [28,]    44.28715    23.24842   12.268133    8.580042 3.427171e+00
#>  [29,]    33.97025    21.35729   14.738532    7.924285 5.129795e+00
#>  [30,]    33.70003    16.40953   11.978146    8.928748 4.050215e+00
#>  [31,]    25.41293    22.75654   18.850677    9.853303 5.893258e+00
#>  [32,]    29.76958    20.88918   15.799438   12.150799 1.040045e+01
#>  [33,]    32.18745    24.94465   17.478162    8.413021 6.188141e+00
#>  [34,]    31.94020    18.84162   14.181052   11.246076 7.043486e+00
#>  [35,]    50.13632    33.22633   17.546647   13.064856 3.198705e-01
#>  [36,]    28.61629    20.50229   14.625504   12.393503 7.771429e+00
#>  [37,]    47.35795    35.59948   13.030472    8.949425 5.439331e+00
#>  [38,]    44.99646    20.68810   10.648798    5.358217 1.498827e+00
#>  [39,]    41.56680    19.67536   16.751639   13.498523 4.950560e+00
#>  [40,]    35.61182    20.81367   16.501185    5.599763 3.863969e+00
#>  [41,]    26.19947    20.54028   18.212527   13.046712 8.585027e+00
#>  [42,]    34.18341    23.13412   18.303894   10.984169 4.806607e+00
#>  [43,]    36.98382    23.52772   14.318316    9.245114 5.326327e+00
#>  [44,]    58.72207    23.42724   10.950533    7.294241 6.326553e+00
#>  [45,]    35.92895    14.74545   10.074454    6.755387 5.756540e+00
#>  [46,]    44.04778    37.05012   18.425598    9.567238 5.209773e+00
#>  [47,]    41.27282    35.94755   16.426939    9.747477 7.678761e+00
#>  [48,]    48.17873    19.15885    9.387026    7.302888 5.692768e+00
#>  [49,]    51.62786    23.09010   16.248057    8.479227 6.319471e+00
#>  [50,]    38.45072    17.38979    9.467152    7.465687 5.547988e+00
#>  [51,]    35.79602    23.51278   21.693524   13.033065 7.638033e+00
#>  [52,]    29.22778    18.40854   15.031401    8.390702 7.354653e+00
#>  [53,]    34.25263    16.36476   14.301935    9.867195 6.315355e+00
#>  [54,]    33.38038    21.88885   17.535490   12.632269 7.362877e+00
#>  [55,]    59.36504    25.43005   17.758119    8.699999 4.316168e+00
#>  [56,]    31.58317    21.29622   12.533620    8.143643 5.300727e+00
#>  [57,]    29.78597    21.67835   15.011034    8.370634 7.357626e+00
#>  [58,]    26.87970    15.58779   13.952536    9.697207 4.546029e+00
#>  [59,]    50.82920    27.68487   19.356818   11.871155 3.772752e+00
#>  [60,]    47.69737    31.08419   14.634203    6.000049 1.413200e-15
#>  [61,]    64.50370    23.67650   18.413982    8.243202 5.558725e+00
#>  [62,]    45.89477    32.67206   26.707550    4.944636 4.630601e-15
#>  [63,]    64.09540    22.65259   14.054791    8.214259 5.901690e+00
#>  [64,]    35.51839    17.86429   11.861558    9.674697 4.852104e-15
#>  [65,]    59.96310    34.37961   14.529594    8.350793 6.254529e+00
#>  [66,]    38.94882    30.04922   18.003822   14.816969 5.333507e+00
#>  [67,]    35.67216    17.54336   15.634164    4.641213 6.278655e-15
#>  [68,]    40.77041    33.93587   20.514754    8.261074 7.400743e+00
#>  [69,]    44.85258    23.40933   13.679988    7.742616 4.085229e+00
#>  [70,]    32.02792    20.72229   11.463590    9.103073 7.155973e+00
#>  [71,]    36.84791    11.94076    9.962199    9.431037 5.335227e+00
#>  [72,]    33.64637    19.23742   15.766628   11.155991 4.953316e+00
#>  [73,]    35.70591    17.04684   14.173405   11.692843 7.984496e+00
#>  [74,]    38.66071    32.66051   20.223052    9.813483 4.390176e+00
#>  [75,]    36.97127    19.17263   17.273728   13.084319 4.790780e+00
#>  [76,]    42.32484    25.40113   15.414047    8.689821 6.661255e+00
#>  [77,]    47.42516    23.03854   14.057748    8.812692 5.925548e+00
#>  [78,]    30.20634    16.16357   11.064316    9.025458 7.436882e+00
#>  [79,]    50.52323    23.37701   14.136528    7.130738 5.919554e+00
#>  [80,]    34.10923    21.06813   13.481835   10.566294 3.740409e+00
#>  [81,]    32.15217    18.79414   13.871007    8.982613 4.453073e+00
#>  [82,]    28.59116    20.96856   12.828641    7.975450 3.594327e+00
#>  [83,]    43.63212    35.00252   15.489318    7.603896 6.116478e+00
#>  [84,]    69.43981    15.85948    9.502175    4.351884 2.357674e+00
#>  [85,]    38.94678    22.31293   16.768338    9.266356 4.090564e+00
#>  [86,]    28.53007    19.93443   17.399604   16.087355 7.095143e+00
#>  [87,]    27.96808    26.28701   12.238482    7.350695 4.907624e+00
#>  [88,]    47.11370    21.83267    9.007976    7.986060 5.611466e+00
#>  [89,]    26.80284    21.60581   17.645593   12.841274 9.957708e+00
#>  [90,]    29.17219    19.23482   12.316012    9.965450 6.967179e+00
#>  [91,]    81.59360    23.88080   18.322090    7.343345 4.719222e+00
#>  [92,]    39.80112    21.03986   10.110607    8.342291 6.703362e+00
#>  [93,]    28.82613    26.25913   13.077772    9.760796 5.490310e+00
#>  [94,]    44.67281    29.59320   16.492714    8.705291 4.856488e+00
#>  [95,]    28.93455    21.67283   13.405859   10.562006 5.370793e+00
#>  [96,]    36.70390    21.11766   20.787085   10.412362 6.239109e+00
#>  [97,]    47.92961    21.89191   15.217655    5.377876 4.591781e+00
#>  [98,]    31.80043    20.36281   17.369348    8.091190 4.383506e+00
#>  [99,]    45.95321    21.27429    8.667837    5.880722 2.019549e+00
#> [100,]    35.93174    21.01536   17.524871    9.180275 5.280543e+00
#>         Dimension 6   Dimension 7   Dimension 8   Dimension 9
#>   [1,] 4.229297e+00  1.076165e+00  6.534077e-01 -1.332270e-15
#>   [2,] 5.031398e-15  1.005678e-15  9.408865e-16 -3.251545e-16
#>   [3,] 1.141882e-15 -1.616552e-16 -3.172026e-16 -1.591992e-15
#>   [4,] 4.522963e+00  2.271611e+00  2.486617e-01  1.443290e-15
#>   [5,] 5.318564e+00  3.866543e+00  7.176207e-01  9.467149e-16
#>   [6,] 5.155645e+00  9.323412e-01  2.413377e-01  3.067830e-15
#>   [7,] 8.385070e-01  6.832972e-16 -2.229692e-15 -2.685278e-15
#>   [8,] 5.109323e+00  3.663111e+00  8.549288e-01  1.910683e-01
#>   [9,] 9.820892e-01 -2.411535e-16 -5.845149e-16 -1.427237e-15
#>  [10,] 1.558580e+00  7.984245e-01  8.340474e-16 -4.929642e-16
#>  [11,] 4.112767e+00  2.237820e+00  2.017897e-01  9.246883e-16
#>  [12,] 3.241324e+00  1.812883e+00  4.253257e-15  4.415938e-16
#>  [13,] 2.427913e+00  3.523231e-15  2.545348e-15 -5.510111e-16
#>  [14,] 4.830155e+00  4.968992e-15  6.228349e-16 -6.526368e-16
#>  [15,] 3.674809e+00  1.841307e+00  2.284975e-15  1.236973e-15
#>  [16,] 4.157574e+00  2.838974e+00  1.245820e-16 -1.594897e-16
#>  [17,] 1.121210e-15 -3.797693e-16 -8.487434e-16 -2.699046e-15
#>  [18,] 3.040799e+00  5.072866e-15 -3.802968e-16 -7.295365e-16
#>  [19,] 1.497333e+00  9.111412e-01  7.125468e-15 -1.652605e-15
#>  [20,] 1.347676e-15  1.284838e-15 -8.638453e-16 -1.171924e-15
#>  [21,] 4.315245e+00  1.309485e+00  4.978391e-01  2.161861e-15
#>  [22,] 8.233148e-01  5.817995e-15  1.132211e-15  5.112708e-16
#>  [23,] 4.111567e+00  1.125978e+00  1.055201e-15 -5.984981e-16
#>  [24,] 2.007652e+00  4.092403e-15  1.332046e-17 -3.062258e-16
#>  [25,] 1.757807e+00 -1.425657e-16 -2.095129e-15 -2.408358e-15
#>  [26,] 1.971486e+00  4.443591e-01  1.419196e-15  1.318123e-15
#>  [27,] 6.980559e-01  2.257214e-15  8.362583e-16  6.795695e-16
#>  [28,] 1.577712e+00  1.467202e+00  5.237117e-15 -7.788635e-16
#>  [29,] 3.939044e+00  9.207981e-15  2.996095e-15  5.522577e-16
#>  [30,] 1.264959e+00  6.429666e-16 -1.530690e-16 -1.638638e-15
#>  [31,] 3.387123e+00  1.044287e+00  1.523951e-15 -2.688289e-15
#>  [32,] 3.619878e+00  1.140278e+00 -1.186023e-15 -1.475363e-15
#>  [33,] 5.791075e+00  1.312621e+00  2.995867e-01 -3.645073e-16
#>  [34,] 5.016043e+00  1.195330e+00  2.433050e-01 -1.675743e-15
#>  [35,] 4.598952e-15  3.060030e-15  1.102143e-15  3.226855e-16
#>  [36,] 5.168944e+00  3.319740e+00  8.468087e-01  3.275807e-01
#>  [37,] 2.597020e+00  7.826720e-01  2.815087e-15 -1.152783e-15
#>  [38,] 1.010799e-15  9.269514e-16 -1.091722e-16 -1.335642e-15
#>  [39,] 4.216350e+00  3.441342e+00 -4.636820e-16 -1.500326e-15
#>  [40,] 2.107882e-15 -4.337188e-16 -1.275536e-15 -1.431671e-15
#>  [41,] 4.658791e+00  2.673913e+00  1.004129e-15 -4.189172e-15
#>  [42,] 2.358913e+00  3.384966e-15  3.032981e-16 -2.376255e-16
#>  [43,] 2.936251e+00  4.958793e-15 -1.529675e-15 -1.892449e-15
#>  [44,] 2.168967e+00  7.997871e-15  1.761323e-15 -1.987361e-15
#>  [45,] 4.004018e+00  5.994470e-01  1.251666e-15  3.325404e-16
#>  [46,] 2.876046e-01  8.421104e-15  3.715266e-15 -1.162121e-15
#>  [47,] 1.570338e+00  6.753945e-15  3.925803e-15 -5.353754e-16
#>  [48,] 3.687446e+00  2.387724e+00  4.682631e-01  0.000000e+00
#>  [49,] 2.384529e-15  1.142927e-15  9.043935e-17 -6.734796e-16
#>  [50,] 2.459494e+00  1.405319e+00  3.023625e-15  1.753681e-15
#>  [51,] 4.400070e+00  2.233546e-01  1.355883e-15  1.331093e-15
#>  [52,] 7.037500e+00  1.124015e+00  1.088230e-15  7.592741e-16
#>  [53,] 4.345216e+00  3.047487e+00  5.788150e-15  1.103343e-15
#>  [54,] 4.693138e+00  1.157583e+00  3.940455e-01  7.326124e-15
#>  [55,] 7.335674e-01  4.581800e-15  2.923914e-16 -9.407517e-16
#>  [56,] 4.465119e+00  1.208138e+00  5.758227e-17 -1.633435e-15
#>  [57,] 4.512285e+00  3.660731e+00  6.970771e-01  4.233719e-15
#>  [58,] 2.904670e+00  1.954121e+00  2.383139e-15 -3.671931e-16
#>  [59,] 2.747785e+00  2.507780e-01  2.331183e-15 -5.974524e-17
#>  [60,] 1.153052e-15  9.580052e-16  7.055222e-16 -1.282765e-15
#>  [61,] 5.288814e-01 -1.753423e-17 -1.549254e-15 -3.436975e-15
#>  [62,] 2.527682e-15  2.510142e-15  3.207599e-16 -3.508470e-16
#>  [63,] 4.513952e+00  4.000189e-01  2.907894e-15  1.896160e-15
#>  [64,] 3.339754e-16 -1.145022e-16 -6.290726e-16 -8.494061e-16
#>  [65,] 3.350083e+00  2.565123e-01  1.396902e-15  8.881784e-16
#>  [66,] 3.154597e+00  5.125597e-01  2.331831e-16 -2.604565e-16
#>  [67,] 2.924378e-15  2.225689e-15  1.572825e-15 -1.073406e-15
#>  [68,] 4.018477e+00  7.176886e-16 -6.887765e-16 -3.354957e-15
#>  [69,] 2.895505e+00  1.228828e+00 -2.432642e-16 -3.875071e-15
#>  [70,] 6.488647e+00  1.486501e+00 -1.353215e-16 -1.749578e-15
#>  [71,] 4.345093e+00  3.062956e-15  9.495238e-16  8.197843e-16
#>  [72,] 2.431549e+00  6.687102e-01  4.098029e-16  1.453086e-16
#>  [73,] 3.377697e+00  1.947806e+00  1.104029e-14  5.022557e-15
#>  [74,] 1.892771e+00  6.713073e-01  1.890842e-15  8.855308e-16
#>  [75,] 3.136926e+00  3.584107e-16 -1.340927e-15 -4.102164e-15
#>  [76,] 1.181128e+00  4.478240e-15  2.404689e-15 -9.640523e-16
#>  [77,] 5.664032e+00  1.353014e+00  3.050195e-01  3.315249e-15
#>  [78,] 4.501529e+00  2.236899e+00  1.680452e-15 -7.246595e-16
#>  [79,] 3.647570e+00  1.545978e+00  3.804558e-16 -7.504558e-16
#>  [80,] 2.024782e+00  2.315996e-15  1.663091e-16 -5.985568e-16
#>  [81,] 1.828627e+00  1.402375e+00  5.613186e-16 -1.482572e-15
#>  [82,] 1.568448e+00  3.763454e-16 -2.438717e-16 -1.930367e-15
#>  [83,] 5.126760e+00  8.578631e-01  2.717412e-01  3.657001e-15
#>  [84,] 3.015502e-15  1.358633e-15  8.319880e-16 -2.249819e-15
#>  [85,] 3.064419e+00  4.770300e-15  1.179951e-15  5.917740e-16
#>  [86,] 2.919259e+00  6.057230e-01  4.252645e-15  9.657692e-16
#>  [87,] 1.420469e+00  2.109948e-15  1.079332e-15  6.033062e-17
#>  [88,] 4.155275e+00  2.114203e+00  1.118911e-15 -2.460753e-16
#>  [89,] 4.706896e+00  2.793843e+00  1.499660e+00  8.666327e-16
#>  [90,] 4.398321e+00  1.870671e+00  2.217508e-15  1.640548e-15
#>  [91,] 8.650525e-15  2.403654e-15  1.018770e-15 -9.052824e-16
#>  [92,] 3.403278e+00  1.087615e+00  1.605050e-15 -1.763489e-15
#>  [93,] 1.412044e+00  1.009277e+00  2.951548e-15  5.253169e-16
#>  [94,] 2.375801e+00  2.076237e-15  1.545942e-15  1.498910e-15
#>  [95,] 3.458426e+00  1.952762e+00  3.466546e-15  1.782312e-15
#>  [96,] 3.889940e+00  2.432308e+00  1.951648e-15  1.334185e-15
#>  [97,] 4.807903e-15  1.923315e-15  1.323540e-15 -4.164820e-15
#>  [98,] 2.149910e+00  1.177799e-15  1.037026e-15  5.135399e-16
#>  [99,] 3.513113e-15  2.449924e-15  7.308011e-16  7.120158e-16
#> [100,] 1.111913e+00  3.777664e-15 -3.534280e-16 -6.584561e-16
#> 
#> $BootMatEigsCI
#>                     LowCI       HighCI
#> Dimension 1  2.619947e+01 6.450370e+01
#> Dimension 2  1.474545e+01 3.559948e+01
#> Dimension 3  9.007976e+00 2.169352e+01
#> Dimension 4  4.381192e+00 1.481697e+01
#> Dimension 5  4.630601e-15 8.755615e+00
#> Dimension 6  1.010799e-15 5.791075e+00
#> Dimension 7 -3.797693e-16 3.660731e+00
#> Dimension 8 -2.095129e-15 8.468087e-01
#> Dimension 9 -4.164820e-15 7.326124e-15
#> 
#> $FixedEigs
#> [1] 27.4321832 21.0362649 15.8008790 12.3135816  8.2191499  5.3689934  3.9265463
#> [8]  2.5779650  0.4419551
#> 
#> $BootstrappedEstEigs
#> Dimension 1 Dimension 2 Dimension 3 Dimension 4 Dimension 5 Dimension 6 
#> 39.65667120 23.31395805 15.30661623  9.27730536  5.28631241  2.71075094 
#> Dimension 7 Dimension 8 Dimension 9 
#>  0.86961002  0.08441093  0.00518649 
#> 
#> $BootMatTaus
#>        Dimension 1 Dimension 2 Dimension 3 Dimension 4  Dimension 5
#>   [1,]   0.4212194   0.2208885  0.14379850  0.08705062 6.886390e-02
#>   [2,]   0.4266279   0.2128256  0.18700379  0.10812195 6.542075e-02
#>   [3,]   0.3927174   0.3390334  0.15425940  0.09342694 2.056283e-02
#>   [4,]   0.2946997   0.2206804  0.19123029  0.12146464 9.527967e-02
#>   [5,]   0.2821854   0.2101812  0.20915651  0.14324655 5.864314e-02
#>   [6,]   0.3573731   0.2479425  0.15820463  0.10453548 6.991701e-02
#>   [7,]   0.4281530   0.2208573  0.17441691  0.14525220 2.197113e-02
#>   [8,]   0.2790032   0.2652103  0.15811939  0.13727251 6.145879e-02
#>   [9,]   0.4261358   0.2439703  0.13539548  0.11223963 6.836532e-02
#>  [10,]   0.4376204   0.2799817  0.11200451  0.07858112 6.866204e-02
#>  [11,]   0.3394915   0.2274467  0.15606294  0.14030145 6.645523e-02
#>  [12,]   0.4719340   0.2255487  0.13911712  0.07521207 3.808865e-02
#>  [13,]   0.4388166   0.2953084  0.13780020  0.06022133 3.856273e-02
#>  [14,]   0.3007991   0.2442711  0.20683910  0.14015869 6.365635e-02
#>  [15,]   0.3778820   0.2520407  0.16506806  0.08894107 6.011964e-02
#>  [16,]   0.3356532   0.2152045  0.18684526  0.10381627 8.367370e-02
#>  [17,]   0.3921784   0.2871577  0.19215469  0.09075849 3.775075e-02
#>  [18,]   0.4209540   0.2418180  0.19588767  0.06980672 4.160991e-02
#>  [19,]   0.5633001   0.2072474  0.11295488  0.05601111 3.852871e-02
#>  [20,]   0.4729572   0.3255912  0.13134603  0.04335385 2.675176e-02
#>  [21,]   0.3356932   0.2450590  0.18929347  0.09658920 6.426625e-02
#>  [22,]   0.4763610   0.2543286  0.15023638  0.06882126 4.302986e-02
#>  [23,]   0.3539669   0.2311820  0.18450134  0.10608582 7.379193e-02
#>  [24,]   0.3988023   0.2491179  0.17821257  0.09750718 5.529535e-02
#>  [25,]   0.4079929   0.2617124  0.18112607  0.07829332 5.012518e-02
#>  [26,]   0.4738258   0.2593586  0.12399034  0.06786387 5.303220e-02
#>  [27,]   0.4042685   0.2304740  0.19575362  0.11800580 4.420880e-02
#>  [28,]   0.4668891   0.2450921  0.12933451  0.09045350 3.613032e-02
#>  [29,]   0.3901971   0.2453192  0.16929324  0.09102181 5.892307e-02
#>  [30,]   0.4414950   0.2149768  0.15692246  0.11697312 5.306078e-02
#>  [31,]   0.2914390   0.2609751  0.21618214  0.11299902 6.758469e-02
#>  [32,]   0.3174758   0.2227713  0.16849211  0.12958143 1.109150e-01
#>  [33,]   0.3331527   0.2581869  0.18090582  0.08707806 6.404969e-02
#>  [34,]   0.3560498   0.2100349  0.15808168  0.12536437 7.851647e-02
#>  [35,]   0.4386609   0.2907092  0.15352200  0.11430918 2.798664e-03
#>  [36,]   0.3058208   0.2191069  0.15630200  0.13244872 8.305285e-02
#>  [37,]   0.4163104   0.3129450  0.11454721  0.07867188 4.781563e-02
#>  [38,]   0.5408853   0.2486837  0.12800513  0.06440908 1.801682e-02
#>  [39,]   0.3992946   0.1890034  0.16091784  0.12966810 4.755555e-02
#>  [40,]   0.4322326   0.2526224  0.20028042  0.06796621 4.689829e-02
#>  [41,]   0.2789649   0.2187074  0.19392208  0.13891787 9.141106e-02
#>  [42,]   0.3645409   0.2467084  0.19519758  0.11713810 5.125893e-02
#>  [43,]   0.4005285   0.2548012  0.15506495  0.10012303 5.768322e-02
#>  [44,]   0.5392808   0.2151467  0.10056547  0.06698749 5.810062e-02
#>  [45,]   0.4614307   0.1893738  0.12938486  0.08675853 7.393047e-02
#>  [46,]   0.3844009   0.3233330  0.16079852  0.08349242 4.546522e-02
#>  [47,]   0.3664009   0.3191256  0.14583073  0.08653357 6.816847e-02
#>  [48,]   0.5004870   0.1990246  0.09751367  0.07586337 5.913723e-02
#>  [49,]   0.4881388   0.2183157  0.15362455  0.08017066 5.975027e-02
#>  [50,]   0.4678491   0.2115903  0.11519157  0.09083874 6.750514e-02
#>  [51,]   0.3367552   0.2211993  0.20408435  0.12261007 7.185568e-02
#>  [52,]   0.3376023   0.2126321  0.17362370  0.09691876 8.495163e-02
#>  [53,]   0.3870591   0.1849239  0.16161367  0.11150055 7.136432e-02
#>  [54,]   0.3370236   0.2209998  0.17704635  0.12754119 7.433899e-02
#>  [55,]   0.5104345   0.2186535  0.15268847  0.07480463 3.711143e-02
#>  [56,]   0.3736298   0.2519350  0.14827311  0.09633954 6.270776e-02
#>  [57,]   0.3270534   0.2380308  0.16482292  0.09191055 8.078760e-02
#>  [58,]   0.3559185   0.2064005  0.18474785  0.12840233 6.019472e-02
#>  [59,]   0.4362521   0.2376112  0.16613390  0.10188664 3.238043e-02
#>  [60,]   0.4797765   0.3126684  0.14720196  0.06035307 1.421504e-17
#>  [61,]   0.5334191   0.1957949  0.15227608  0.06816790 4.596838e-02
#>  [62,]   0.4163961   0.2964285  0.24231344  0.04486191 4.201272e-17
#>  [63,]   0.5348740   0.1890351  0.11728678  0.06854773 4.924942e-02
#>  [64,]   0.4740909   0.2384483  0.15832525  0.12913555 6.476473e-17
#>  [65,]   0.4718375   0.2705262  0.11433043  0.06571070 4.921563e-02
#>  [66,]   0.3514618   0.2711546  0.16246079  0.13370364 4.812788e-02
#>  [67,]   0.4853956   0.2387147  0.21273608  0.06315358 8.543447e-17
#>  [68,]   0.3548297   0.2953479  0.17854237  0.07189712 6.440955e-02
#>  [69,]   0.4581746   0.2391292  0.13974277  0.07909178 4.173112e-02
#>  [70,]   0.3621102   0.2342878  0.12960825  0.10292007 8.090600e-02
#>  [71,]   0.4732450   0.1533575  0.12794650  0.12112467 6.852138e-02
#>  [72,]   0.3829544   0.2189554  0.17945174  0.12697465 5.637738e-02
#>  [73,]   0.3884075   0.1854348  0.15417775  0.12719429 8.685504e-02
#>  [74,]   0.3569384   0.3015410  0.18671107  0.09060383 4.053268e-02
#>  [75,]   0.3915218   0.2030361  0.18292697  0.13856156 5.073386e-02
#>  [76,]   0.4246403   0.2548466  0.15464737  0.08718398 6.683161e-02
#>  [77,]   0.4449651   0.2161584  0.13189639  0.08268482 5.559627e-02
#>  [78,]   0.3746058   0.2004535  0.13721484  0.11192980 9.222897e-02
#>  [79,]   0.4753758   0.2199556  0.13301135  0.06709350 5.569740e-02
#>  [80,]   0.4013291   0.2478875  0.15862720  0.12432296 4.400963e-02
#>  [81,]   0.3945825   0.2306483  0.17022981  0.11023774 5.464966e-02
#>  [82,]   0.3785576   0.2776315  0.16985596  0.10559792 4.759022e-02
#>  [83,]   0.3824001   0.3067687  0.13575131  0.06664198 5.360596e-02
#>  [84,]   0.6840618   0.1562341  0.09360732  0.04287105 2.322579e-02
#>  [85,]   0.4123561   0.2362422  0.17753782  0.09810922 4.330959e-02
#>  [86,]   0.3081946   0.2153407  0.18795837  0.17378287 7.664494e-02
#>  [87,]   0.3488494   0.3278812  0.15265214  0.09168615 6.121342e-02
#>  [88,]   0.4816300   0.2231892  0.09208599  0.08163924 5.736444e-02
#>  [89,]   0.2739074   0.2207973  0.18032641  0.13122941 1.017613e-01
#>  [90,]   0.3475998   0.2291916  0.14675083  0.11874283 8.301708e-02
#>  [91,]   0.6005753   0.1757763  0.13486101  0.05405120 3.473616e-02
#>  [92,]   0.4398491   0.2325151  0.11173407  0.09219210 7.408002e-02
#>  [93,]   0.3358301   0.3059241  0.15235862  0.11371519 6.396319e-02
#>  [94,]   0.4186913   0.2773592  0.15457624  0.08158943 4.551693e-02
#>  [95,]   0.3389818   0.2539074  0.15705593  0.12373886 6.292136e-02
#>  [96,]   0.3613216   0.2078871  0.20463281  0.10250167 6.141922e-02
#>  [97,]   0.5044753   0.2304197  0.16017096  0.05660396 4.833005e-02
#>  [98,]   0.3778694   0.2419617  0.20639171  0.09614377 5.208712e-02
#>  [99,]   0.5483964   0.2538831  0.10344023  0.07017936 2.410089e-02
#> [100,]   0.3990434   0.2333881  0.19462411  0.10195241 5.864357e-02
#>         Dimension 6   Dimension 7   Dimension 8   Dimension 9
#>   [1,] 4.129248e-02  1.050707e-02  6.379507e-03 -1.300753e-17
#>   [2,] 5.393379e-17  1.078031e-17  1.008578e-17 -3.485476e-18
#>   [3,] 1.188176e-17 -1.682091e-18 -3.300627e-18 -1.656534e-17
#>   [4,] 4.921943e-02  2.471994e-02  2.705967e-03  1.570606e-17
#>   [5,] 5.187511e-02  3.771268e-02  6.999380e-03  9.233871e-18
#>   [6,] 5.052530e-02  9.136941e-03  2.365108e-03  3.006473e-17
#>   [7,] 9.349461e-03  7.618852e-18 -2.486135e-17 -2.994120e-17
#>   [8,] 5.148429e-02  3.691149e-02  8.614724e-03  1.925307e-03
#>   [9,] 1.389345e-02 -3.411558e-18 -8.269033e-18 -2.019087e-17
#>  [10,] 1.530824e-02  7.842056e-03  8.191942e-18 -4.841852e-18
#>  [11,] 4.408930e-02  2.398967e-02  2.163207e-03  9.912758e-18
#>  [12,] 3.212937e-02  1.797006e-02  4.216008e-17  4.377265e-18
#>  [13,] 2.929073e-02  4.250483e-17  3.070749e-17 -6.647487e-18
#>  [14,] 4.427560e-02  4.554826e-17  5.709215e-18 -5.982394e-18
#>  [15,] 3.727265e-02  1.867590e-02  2.317592e-17  1.254630e-17
#>  [16,] 4.445279e-02  3.035432e-02  1.332031e-18 -1.705264e-18
#>  [17,] 1.075292e-17 -3.642163e-18 -8.139841e-18 -2.588509e-17
#>  [18,] 2.992377e-02  4.992086e-17 -3.742410e-18 -7.179193e-18
#>  [19,] 1.365106e-02  8.306800e-03  6.496230e-17 -1.506667e-17
#>  [20,] 1.333586e-17  1.271404e-17 -8.548134e-18 -1.159671e-17
#>  [21,] 4.870151e-02  1.477874e-02  5.618572e-03  2.439859e-17
#>  [22,] 7.222807e-03  5.104032e-17  9.932700e-18  4.485296e-18
#>  [23,] 3.962143e-02  1.085058e-02  1.016853e-17 -5.767473e-18
#>  [24,] 2.106472e-02  4.293838e-17  1.397612e-19 -3.212988e-18
#>  [25,] 2.075013e-02 -1.682925e-18 -2.473207e-17 -2.842960e-17
#>  [26,] 1.789571e-02  4.033568e-03  1.288242e-17  1.196495e-17
#>  [27,] 7.289265e-03  2.357037e-17  8.732407e-18  7.096226e-18
#>  [28,] 1.663274e-02  1.546771e-02  5.521134e-17 -8.211024e-18
#>  [29,] 4.524558e-02  1.057669e-16  3.441446e-17  6.343473e-18
#>  [30,] 1.657189e-02  8.423332e-18 -2.005316e-18 -2.146735e-17
#>  [31,] 3.884399e-02  1.197603e-02  1.747688e-17 -3.082966e-17
#>  [32,] 3.860396e-02  1.216042e-02 -1.264827e-17 -1.573391e-17
#>  [33,] 5.993990e-02  1.358614e-02  3.100840e-03 -3.772794e-18
#>  [34,] 5.591578e-02  1.332481e-02  2.712215e-03 -1.868016e-17
#>  [35,] 4.023790e-17  2.677331e-17  9.643054e-18  2.823293e-18
#>  [36,] 5.524023e-02  3.547788e-02  9.049800e-03  3.500838e-03
#>  [37,] 2.282967e-02  6.880249e-03  2.474664e-17 -1.013379e-17
#>  [38,] 1.215043e-17  1.114253e-17 -1.312317e-18 -1.605524e-17
#>  [39,] 4.050267e-02  3.305786e-02 -4.454174e-18 -1.441228e-17
#>  [40,] 2.558408e-17 -5.264191e-18 -1.548161e-17 -1.737667e-17
#>  [41,] 4.960555e-02  2.847110e-02  1.069170e-17 -4.460517e-17
#>  [42,] 2.515607e-02  3.609818e-17  3.234452e-18 -2.534102e-18
#>  [43,] 3.179910e-02  5.370289e-17 -1.656612e-17 -2.049490e-17
#>  [44,] 1.991896e-02  7.344935e-17  1.617531e-17 -1.825115e-17
#>  [45,] 5.142307e-02  7.698617e-03  1.607498e-17  4.270771e-18
#>  [46,] 2.509899e-03  7.349021e-17  3.242279e-17 -1.014172e-17
#>  [47,] 1.394073e-02  5.995838e-17  3.485145e-17 -4.752814e-18
#>  [48,] 3.830568e-02  2.480400e-02  4.864379e-03  0.000000e+00
#>  [49,] 2.254560e-17  1.080631e-17  8.550995e-19 -6.367715e-18
#>  [50,] 2.992589e-02  1.709922e-02  3.678995e-17  2.133791e-17
#>  [51,] 4.139417e-02  2.101234e-03  1.275563e-17  1.252241e-17
#>  [52,] 8.128829e-02  1.298320e-02  1.256986e-17  8.770173e-18
#>  [53,] 4.910149e-02  3.443699e-02  6.540683e-17  1.246791e-17
#>  [54,] 4.738408e-02  1.168749e-02  3.978464e-03  7.396791e-17
#>  [55,] 6.307384e-03  3.939539e-17  2.514050e-18 -8.088803e-18
#>  [56,] 5.282249e-02  1.429231e-02  6.812000e-19 -1.932359e-17
#>  [57,] 4.954543e-02  4.019526e-02  7.653989e-03  4.648673e-17
#>  [58,] 3.846122e-02  2.587485e-02  3.155554e-17 -4.862065e-18
#>  [59,] 2.358344e-02  2.152354e-03  2.000786e-17 -5.127759e-19
#>  [60,] 1.159828e-17  9.636346e-18  7.096680e-18 -1.290303e-17
#>  [61,] 4.373632e-03 -1.450009e-19 -1.281170e-17 -2.842237e-17
#>  [62,] 2.293326e-17  2.277413e-17  2.910204e-18 -3.183180e-18
#>  [63,] 3.766879e-02  3.338145e-03  2.426628e-17  1.582339e-17
#>  [64,] 4.457825e-18 -1.528349e-18 -8.396711e-18 -1.133767e-17
#>  [65,] 2.636112e-02  2.018443e-03  1.099194e-17  6.988896e-18
#>  [66,] 2.846608e-02  4.625177e-03  2.104171e-18 -2.350277e-18
#>  [67,] 3.979239e-17  3.028524e-17  2.140163e-17 -1.460597e-17
#>  [68,] 3.497329e-02  6.246130e-18 -5.994505e-18 -2.919859e-17
#>  [69,] 2.957794e-02  1.255263e-02 -2.484974e-18 -3.958433e-17
#>  [70,] 7.336115e-02  1.680650e-02 -1.529955e-18 -1.978086e-17
#>  [71,] 5.580489e-02  3.933815e-17  1.219492e-17  1.052865e-17
#>  [72,] 2.767527e-02  7.611089e-03  4.664273e-18  1.653865e-18
#>  [73,] 3.674246e-02  2.118816e-02  1.200959e-16  5.463518e-17
#>  [74,] 1.747517e-02  6.197903e-03  1.745736e-17  8.175740e-18
#>  [75,] 3.321972e-02  3.795532e-18 -1.420028e-17 -4.344149e-17
#>  [76,] 1.185012e-02  4.492967e-17  2.412597e-17 -9.672227e-18
#>  [77,] 5.314261e-02  1.269461e-02  2.861836e-03  3.110522e-17
#>  [78,] 5.582600e-02  2.774105e-02  2.084023e-17 -8.986912e-18
#>  [79,] 3.432018e-02  1.454619e-02  3.579729e-18 -7.061079e-18
#>  [80,] 2.382357e-02  2.725000e-17  1.956792e-18 -7.042617e-18
#>  [81,] 2.244154e-02  1.721043e-02  6.888696e-18 -1.819464e-17
#>  [82,] 2.076684e-02  4.982953e-18 -3.228952e-18 -2.555877e-17
#>  [83,] 4.493189e-02  7.518474e-03  2.381591e-03  3.205065e-17
#>  [84,] 2.970616e-17  1.338409e-17  8.196036e-18 -2.216330e-17
#>  [85,] 3.244509e-02  5.050641e-17  1.249295e-17  6.265514e-18
#>  [86,] 3.153515e-02  6.543294e-03  4.593899e-17  1.043267e-17
#>  [87,] 1.771769e-02  2.631765e-17  1.346265e-17  7.525115e-19
#>  [88,] 4.247820e-02  2.161290e-02  1.143831e-17 -2.515558e-18
#>  [89,] 4.810140e-02  2.855125e-02  1.532555e-02  8.856419e-18
#>  [90,] 5.240798e-02  2.228989e-02  2.642260e-17  1.954787e-17
#>  [91,] 6.367279e-17  1.769226e-17  7.498729e-18 -6.663394e-18
#>  [92,] 3.761022e-02  1.201942e-02  1.773768e-17 -1.948862e-17
#>  [93,] 1.645059e-02  1.175828e-02  3.438611e-17  6.120046e-18
#>  [94,] 2.226695e-02  1.945932e-17  1.448918e-17  1.404838e-17
#>  [95,] 4.051708e-02  2.287752e-02  4.061221e-17  2.088062e-17
#>  [96,] 3.829346e-02  2.394419e-02  1.921247e-17  1.313402e-17
#>  [97,] 5.060480e-17  2.024354e-17  1.393071e-17 -4.383614e-17
#>  [98,] 2.554636e-02  1.399523e-17  1.232249e-17  6.102151e-18
#>  [99,] 4.192478e-17  2.923691e-17  8.721234e-18  8.497053e-18
#> [100,] 1.234845e-02  4.195320e-17 -3.925028e-18 -7.312547e-18
#> 
#> $BootMatTausCI
#>                     LowCI       HighCI
#> Dimension 1  2.789649e-01 5.633001e-01
#> Dimension 2  1.562341e-01 3.255912e-01
#> Dimension 3  9.360732e-02 2.127361e-01
#> Dimension 4  4.335385e-02 1.432466e-01
#> Dimension 5  4.201272e-17 9.527967e-02
#> Dimension 6  1.075292e-17 5.993990e-02
#> Dimension 7 -3.642163e-18 3.691149e-02
#> Dimension 8 -2.473207e-17 8.614724e-03
#> Dimension 9 -4.383614e-17 7.396791e-17
#> 
#> $BootstrappedEstInertia
#>   [1] 102.42293  93.28841  96.10376  91.89385 102.52633 102.04086  89.68506
#>   [8]  99.24041  70.68721 101.81316  93.28265 100.88351  82.89014 109.09292
#>  [15]  98.59264  93.52786 104.27027 101.61817 109.68619 101.05659  88.60598
#>  [22] 113.98820 103.77128  95.30874  84.71304 110.16528  95.76492  94.85583
#>  [29]  87.05919  76.33162  87.19812  93.76960  96.61470  89.70711 114.29402
#>  [36]  93.57209 113.75634  83.19040 104.10057  82.39040  93.91672  93.77111
#>  [43]  92.33754 108.88960  77.86424 114.58811 112.64388  96.26369 105.76471
#>  [50]  82.18616 106.29685  86.57459  88.49458  99.04463 116.30295  84.53064
#>  [57]  91.07370  75.52205 116.51336  99.41581 120.92498 110.21902 119.83270
#>  [64]  74.91893 127.08422 110.81949  73.49089 114.90132  97.89407  88.44799
#>  [71]  77.86223  87.85999  91.92899 108.31201  94.42964  99.67222 106.58175
#>  [78]  80.63499 106.28062  84.99069  81.48401  75.52659 114.10069 101.51102
#>  [85]  94.44939  92.57158  80.17236  97.82135  97.85362  83.92465 135.85906
#>  [92]  90.48813  85.83546 106.69630  85.35723 101.58237  95.00882  84.15719
#>  [99]  83.79561  90.04470
#>