Boostrap for eigenvalues.

Boostrap for eigenvalues.

Usage

Boot4Eigs(
  data,
  center = FALSE,
  scale = FALSE,
  design = NULL,
  niter = 100,
  CI.perc = c(0.025, 0.975),
  suppressProgressBar = TRUE
)

Arguments

data

the data-set;

center

whether to center the data (default to FALSE), see expo.scale;

scale

whether to scale the data ( default to FALSE), see expo.scale;

design

whether there is a design;

niter

number of bootstrap iterations;

CI.perc

vector of the two levels of confidence for the left and right limits of the bootstrap confidence interval (default to c(0.025, 0.0975));

suppressProgressBar

whether to show a progress bar or not (default to FALSE).

Value

list containing four objects

Examples

Boot4Eigs(matrix(rnorm(100), 10, 10))
#> $BootMatEigs
#>        Dimension 1 Dimension 2 Dimension 3 Dimension 4  Dimension 5
#>   [1,]    28.05501    18.53865   17.547311    9.650822 7.199955e+00
#>   [2,]    35.75514    22.97285   13.756252   12.748677 2.815851e+00
#>   [3,]    29.70457    24.76704   17.703075    9.291292 7.303648e+00
#>   [4,]    34.35199    19.68398   17.933987   10.477285 4.150156e+00
#>   [5,]    29.15108    23.98236   17.441232   11.209696 8.215636e+00
#>   [6,]    41.36535    21.91679   15.362565    7.670714 3.684581e+00
#>   [7,]    35.36477    26.09406   12.590241    8.738844 4.964267e+00
#>   [8,]    34.75128    29.62186   19.551243    6.075612 4.044149e-15
#>   [9,]    37.97899    33.02400   14.283796    4.713211 2.388097e-15
#>  [10,]    37.03591    27.99026   11.551153   10.158453 3.264229e+00
#>  [11,]    36.91335    27.94340   12.961001    9.163644 3.018606e+00
#>  [12,]    37.63856    20.39903   14.438110   12.228489 5.295813e+00
#>  [13,]    30.51457    28.97669   14.351663   10.235552 4.273162e+00
#>  [14,]    36.57471    24.09248   13.681605    8.970058 4.944525e+00
#>  [15,]    39.32937    25.64963   16.650181    6.589919 1.780907e+00
#>  [16,]    33.50670    24.66423   14.257474   11.619997 5.951606e+00
#>  [17,]    36.98405    18.39394   14.912809    9.144629 6.100778e+00
#>  [18,]    30.85611    23.13635   20.895549   11.680593 3.431402e+00
#>  [19,]    41.19632    31.98264   10.484235    6.336808 1.306544e-15
#>  [20,]    33.13409    25.17798   14.774357    9.987135 4.628747e+00
#>  [21,]    30.10402    20.53750   13.498693    9.759805 7.006755e+00
#>  [22,]    33.66018    23.76275   15.522754   10.112971 5.941974e+00
#>  [23,]    30.02477    22.61437   14.151146   10.617156 7.054427e+00
#>  [24,]    40.83953    32.80877   13.491149    2.860550 1.342181e-15
#>  [25,]    33.50731    21.94373   17.487065    8.834814 5.570295e+00
#>  [26,]    31.98496    25.32607   19.863041   10.670922 2.155014e+00
#>  [27,]    30.30447    23.31082   20.239946    9.936939 6.207824e+00
#>  [28,]    32.63209    25.74562   15.630269    8.574588 7.417438e+00
#>  [29,]    37.65197    17.87827   17.016619   10.646200 4.036040e+00
#>  [30,]    40.96970    20.05783   18.149158    8.798459 2.024855e+00
#>  [31,]    39.94803    29.86686   15.006884    5.178218 2.144758e-15
#>  [32,]    28.79720    24.36419   15.886117   12.262563 8.689924e+00
#>  [33,]    36.61034    28.78139   15.878381    6.323163 2.406723e+00
#>  [34,]    46.29736    20.83035   12.737176    7.680445 2.454679e+00
#>  [35,]    35.92368    30.15290   14.235684    6.867937 2.819797e+00
#>  [36,]    28.60886    25.53146   18.351879    9.131222 8.376585e+00
#>  [37,]    39.12276    28.05232   14.027892    6.700565 2.096471e+00
#>  [38,]    37.60347    26.65274   11.040245    9.503253 5.200293e+00
#>  [39,]    33.41361    23.85831   20.983159    8.127523 3.617394e+00
#>  [40,]    35.64410    32.84935   15.837426    5.669123 3.511772e-15
#>  [41,]    33.70577    22.77272   14.966383   10.045795 5.881943e+00
#>  [42,]    35.98117    24.10563   14.175663    7.755372 7.178877e+00
#>  [43,]    35.65780    30.00760   15.010237    9.324361 2.553650e-14
#>  [44,]    45.72090    32.13913    8.384958    3.755015 2.084985e-15
#>  [45,]    39.77705    29.31057   16.184621    4.727761 4.077779e-15
#>  [46,]    37.71676    23.26484   16.227195    8.136863 4.654343e+00
#>  [47,]    35.02777    26.42198   14.194895    7.341414 5.697684e+00
#>  [48,]    38.92946    20.67793   11.527847   10.409573 4.161594e+00
#>  [49,]    29.57342    17.88045   14.441621   11.911927 7.214660e+00
#>  [50,]    37.32938    19.65719   16.716148    9.411550 4.125426e+00
#>  [51,]    32.18407    22.81804   14.926523    9.347228 7.669519e+00
#>  [52,]    39.31961    27.72716   14.975752    5.379542 2.597938e+00
#>  [53,]    39.28496    22.67144   12.784156    9.798571 5.460872e+00
#>  [54,]    39.51579    27.30372   15.722932    7.457559 1.943152e-15
#>  [55,]    30.37152    26.16979   16.924071   13.715908 2.818715e+00
#>  [56,]    35.05748    24.82552   11.842291    8.647709 7.576945e+00
#>  [57,]    35.72429    26.03241   15.999186   12.244108 8.135707e-15
#>  [58,]    34.77437    20.96262   19.065013   10.284711 4.913292e+00
#>  [59,]    36.94551    32.12189   11.514820    6.535400 2.882381e+00
#>  [60,]    28.69405    20.27689   18.829739   12.882422 6.461647e+00
#>  [61,]    32.91933    20.19430   20.018098   12.371331 4.496940e+00
#>  [62,]    31.94761    23.93200   14.167651    8.020114 6.569039e+00
#>  [63,]    31.87620    24.08716   15.072894    9.868695 6.024184e+00
#>  [64,]    37.09546    27.75036   13.828436   11.325741 5.275753e-15
#>  [65,]    42.59646    25.24395   15.961828    6.197758 1.115152e-15
#>  [66,]    47.47013    28.38412   10.121218    4.024532 1.201513e-15
#>  [67,]    30.96484    21.60701   15.184958   10.491942 5.997754e+00
#>  [68,]    49.03006    22.00045   14.694595    4.274896 9.591676e-15
#>  [69,]    36.29598    21.54609   16.473736   10.970750 3.059409e+00
#>  [70,]    27.93938    26.15684   13.600507   11.760519 9.596299e+00
#>  [71,]    29.27991    20.42784   14.604082   12.796186 7.820405e+00
#>  [72,]    27.02226    22.91262   18.009417   11.687194 7.585340e+00
#>  [73,]    36.79326    21.86803   17.748311    9.137053 4.453345e+00
#>  [74,]    40.49726    19.16311   14.493325    9.232861 6.613444e+00
#>  [75,]    33.33961    26.03493   13.786372    9.255191 5.539266e+00
#>  [76,]    31.22294    25.25002   17.070347    6.921960 5.314895e+00
#>  [77,]    32.80483    23.17935   18.381834    7.923388 4.079141e+00
#>  [78,]    29.99086    26.10124   19.385847   10.120644 4.401412e+00
#>  [79,]    31.27455    25.76730   13.215575   11.487221 5.485443e+00
#>  [80,]    36.23579    31.47715   19.820697    2.466359 3.552714e-14
#>  [81,]    31.53798    21.82333   17.731228   10.320309 4.899213e+00
#>  [82,]    40.21599    21.34301   14.500778   10.677083 3.263141e+00
#>  [83,]    33.91791    23.39557   18.735005    9.384420 2.871838e+00
#>  [84,]    47.01117    20.76954   17.286662    4.932635 2.082491e-15
#>  [85,]    40.27326    18.21449   15.623669    8.024934 6.307683e+00
#>  [86,]    29.75501    19.80662   15.227766   10.214018 6.730373e+00
#>  [87,]    39.90740    20.27600   12.629431    9.715936 4.992487e+00
#>  [88,]    33.13242    20.42835   14.599755   11.524555 8.128856e+00
#>  [89,]    31.69322    24.81471   20.129970   10.187797 3.174300e+00
#>  [90,]    40.83117    17.86245   14.233021    8.255897 5.344498e+00
#>  [91,]    33.40426    24.87064   12.775330   11.311432 7.638335e+00
#>  [92,]    36.98145    23.63257   14.467622    7.921448 5.155298e+00
#>  [93,]    41.86743    21.10142   14.168690    9.096552 3.765907e+00
#>  [94,]    32.99376    28.60933   23.187378    5.209534 6.785498e-15
#>  [95,]    34.58939    24.77737   14.784777   11.450912 4.397549e+00
#>  [96,]    35.42286    20.89418   15.312546   13.655602 4.714819e+00
#>  [97,]    35.89157    24.73992   12.640074    8.858354 5.275142e+00
#>  [98,]    31.39052    24.86247   17.809369    7.393558 5.900102e+00
#>  [99,]    32.90569    25.03090   16.133283    8.969762 4.652281e+00
#> [100,]    29.89046    25.47887   19.905503   11.442022 3.283142e+00
#>          Dimension 6   Dimension 7   Dimension 8   Dimension 9  Dimension 10
#>   [1,]  6.013860e+00  2.763814e+00  2.305827e-01  6.529501e-16 -2.406631e-15
#>   [2,]  1.951232e+00  3.551279e-15  4.773100e-16 -1.574070e-15 -4.360608e-15
#>   [3,]  1.230374e+00  3.313639e-15 -5.377549e-16 -8.476460e-16 -5.722070e-15
#>   [4,]  2.332325e+00  1.070274e+00  2.155294e-15 -3.684368e-16 -3.033980e-15
#>   [5,]  3.578545e-14 -5.683447e-18 -7.101699e-17 -2.453888e-15 -4.698812e-15
#>   [6,]  5.277836e-15 -6.869532e-16 -1.596711e-15 -3.662337e-15 -5.111016e-15
#>   [7,]  2.247820e+00  1.094929e-15  6.589635e-17 -5.684994e-16 -2.914992e-15
#>   [8,]  2.597763e-15  7.550178e-16  1.411130e-16 -4.430320e-16 -1.563438e-15
#>   [9,]  1.503476e-15  5.688855e-16 -5.933330e-16 -1.626768e-15 -2.466019e-15
#>  [10,]  4.389795e-15  1.694239e-15  2.061835e-16 -1.315684e-15 -1.693245e-15
#>  [11,]  1.925653e-15  1.019169e-15  2.152666e-16 -3.003032e-16 -1.324971e-15
#>  [12,]  3.407691e-16  1.164301e-16 -1.255616e-15 -2.641218e-15 -9.247744e-15
#>  [13,]  1.648364e+00  3.531421e-15  2.771849e-15  1.393206e-15 -5.645148e-15
#>  [14,]  1.736623e+00  4.946231e-15  1.135489e-15  1.074657e-16 -1.024608e-15
#>  [15,]  1.991629e-15  6.648159e-16  1.346000e-16 -2.769630e-15 -7.855220e-15
#>  [16,] -2.364761e-16 -1.146441e-15 -1.294247e-15 -2.742267e-15 -8.618359e-15
#>  [17,]  3.619373e+00  8.444278e-01  2.483834e-15  6.901176e-16 -5.536946e-16
#>  [18,]  1.593003e-15  1.204744e-15 -1.605897e-15 -4.416556e-15 -6.008329e-15
#>  [19,] -1.521564e-16 -1.083982e-15 -1.634492e-15 -1.679845e-15 -7.409544e-15
#>  [20,]  2.297688e+00  1.661437e-15 -3.218756e-16 -3.480972e-15 -4.109035e-15
#>  [21,]  5.032324e+00  3.791589e+00  2.693151e-01  1.347712e-15 -6.866848e-16
#>  [22,]  9.993747e-01  2.471893e-15 -1.781380e-16 -5.011333e-16 -5.616037e-15
#>  [23,]  5.057393e+00  4.807429e-01  1.722407e-15  3.392432e-16 -1.509916e-15
#>  [24,] -7.323815e-16 -9.099898e-16 -1.195159e-15 -2.750826e-15 -8.120391e-15
#>  [25,]  2.656782e+00  4.432050e-15  1.082534e-16 -2.994780e-17 -7.429697e-17
#>  [26,]  2.411502e-15  2.340084e-15  7.307123e-16  2.548881e-16 -2.096084e-15
#>  [27,]  3.952141e-15  2.337386e-15  2.173636e-15  2.119825e-15 -3.790832e-15
#>  [28,]  4.195815e-15  1.907982e-15  3.373139e-16  1.714798e-16 -1.141340e-15
#>  [29,]  2.770896e+00 -3.671274e-16 -1.098154e-15 -1.167649e-15 -1.527286e-14
#>  [30,]  2.525539e-15  6.344617e-16  2.500426e-16 -3.364610e-16 -9.416656e-15
#>  [31,]  1.828169e-15  1.665553e-15  8.203267e-16  3.871920e-16 -5.394079e-15
#>  [32,]  2.740004e-14  5.657267e-16  4.268067e-16 -2.393115e-16 -1.866897e-15
#>  [33,]  5.746524e-15  5.446844e-16 -3.338222e-17 -1.343529e-16 -1.040658e-15
#>  [34,]  3.114078e-15  1.285150e-16 -6.916794e-16 -1.467327e-15 -2.628543e-15
#>  [35,]  5.019797e-15  1.562594e-15  4.470225e-16 -1.316960e-15 -1.836300e-15
#>  [36,]  2.771754e-14  2.022750e-15  3.128521e-16 -1.393769e-16 -2.040940e-15
#>  [37,]  8.521011e-16  7.771270e-16 -1.883246e-15 -3.498613e-15 -8.061713e-15
#>  [38,]  4.974193e-15  3.156667e-15  2.665300e-16  2.188259e-16 -7.415802e-16
#>  [39,]  2.253307e-15  2.172204e-15  1.253529e-15  5.094865e-16 -5.940285e-15
#>  [40,]  2.576181e-15  2.536923e-15  2.367736e-15 -1.244748e-15 -2.471062e-15
#>  [41,]  2.627393e+00  6.211683e-15  3.335683e-15  1.802988e-15  3.311853e-16
#>  [42,]  8.032901e-01  3.512500e-15  1.815013e-15 -3.560261e-15 -5.138601e-15
#>  [43,]  1.923612e-14  1.196579e-15  1.927528e-16  1.785314e-16 -2.443872e-16
#>  [44,]  3.154436e-16 -2.233107e-16 -4.698648e-16 -8.804988e-16 -7.983715e-15
#>  [45,]  2.903544e-15  2.435562e-15 -8.092929e-16 -9.273807e-16 -1.394500e-15
#>  [46,]  6.946837e-15  7.099579e-16  5.571108e-16 -1.547419e-15 -1.557994e-15
#>  [47,]  1.316256e+00 -5.401016e-16 -1.095706e-15 -1.217742e-15 -8.910793e-15
#>  [48,]  3.038907e+00  1.254692e+00  3.097080e-15 -1.330179e-15 -2.945135e-15
#>  [49,]  5.354206e+00  2.741396e+00  8.823131e-01  4.408185e-15 -5.279101e-15
#>  [50,]  2.760301e+00  2.385196e-15  1.525891e-15  7.203205e-16 -4.779094e-15
#>  [51,]  3.054617e+00  2.058344e-15 -1.292209e-15 -2.826687e-15 -4.486974e-15
#>  [52,]  5.639922e-15  4.263090e-15  9.599975e-16  4.649768e-16 -7.393633e-18
#>  [53,]  3.925983e-15  1.441420e-15  1.212975e-15 -5.376137e-16 -2.830043e-15
#>  [54,]  8.786935e-16  2.161926e-16 -4.552750e-16 -4.644373e-16 -8.318430e-15
#>  [55,]  4.836765e-15  2.209824e-15 -7.532861e-16 -3.394037e-15 -3.533147e-15
#>  [56,]  2.050057e+00  1.599750e-15  1.551160e-15 -4.944089e-16 -2.319957e-15
#>  [57,]  2.783839e-15  1.519323e-15 -7.149704e-16 -8.353832e-16 -1.115590e-15
#>  [58,]  1.442031e-16 -5.650250e-16 -6.282661e-16 -1.002634e-15 -3.900103e-15
#>  [59,]  5.263978e-15  2.171836e-15 -1.061697e-15 -3.363997e-15 -3.481413e-15
#>  [60,]  2.855252e+00  2.017510e-15  1.275940e-15 -2.338507e-15 -4.103528e-15
#>  [61,]  2.571818e-15  7.150725e-16  3.856389e-16 -5.794285e-17 -1.051133e-15
#>  [62,]  3.980341e+00  1.383243e+00  1.566631e-15  1.291635e-15 -3.318242e-15
#>  [63,]  3.070868e+00  3.718222e-15  5.074929e-16 -3.172026e-15 -4.296888e-15
#>  [64,]  3.310556e-15  2.953638e-15  5.396236e-16 -2.014527e-15 -2.523004e-15
#>  [65,]  9.374179e-16  8.081644e-16 -2.117692e-15 -2.292100e-15 -2.484882e-15
#>  [66,] -1.911560e-16 -3.436549e-16 -8.334205e-16 -2.404766e-15 -3.630953e-15
#>  [67,]  4.075947e+00  1.677548e+00  2.164021e-15 -2.186369e-15 -5.712953e-15
#>  [68,]  1.992694e-15  5.951056e-16 -7.967227e-16 -2.226274e-15 -4.006008e-15
#>  [69,]  1.654036e+00  5.430834e-15  2.162430e-15  1.094566e-15  7.626392e-16
#>  [70,]  9.464547e-01  9.279700e-15  2.135275e-15  1.614241e-15 -5.937544e-16
#>  [71,]  4.083845e+00  9.877368e-01  7.288898e-16 -1.940257e-15 -4.050449e-15
#>  [72,]  2.783171e+00  1.527772e-15  1.183714e-15 -3.644473e-15 -5.390295e-15
#>  [73,]  3.831429e-15  2.646269e-15  1.668799e-15 -5.267395e-16 -1.906420e-15
#>  [74,]  8.288944e-15  5.590473e-16  2.865786e-16 -1.274767e-17 -1.811163e-15
#>  [75,]  2.044630e+00  2.368751e-15  1.726055e-16 -1.716596e-15 -3.581559e-15
#>  [76,]  3.134414e+00  1.085418e+00  6.777194e-15  1.406188e-15 -1.179765e-16
#>  [77,]  2.812013e+00  8.194434e-01  1.059904e-15 -1.105564e-15 -2.135519e-15
#>  [78,]  8.762298e-15  3.945219e-15  3.748518e-16 -1.657625e-15 -1.876935e-15
#>  [79,]  2.769908e+00  3.202403e-15 -9.293572e-17 -1.956544e-16 -1.729466e-15
#>  [80,]  4.160617e-15  1.975171e-15  3.908855e-16 -6.228891e-16 -1.421260e-15
#>  [81,]  3.687942e+00  3.380540e-15  2.058390e-15  1.612170e-15 -8.516899e-16
#>  [82,]  1.894782e-15  5.310306e-16  1.846505e-16 -6.032007e-16 -3.708439e-15
#>  [83,]  1.695261e+00  4.530801e-15  2.189511e-15 -1.776759e-15 -2.814528e-15
#>  [84,]  1.776819e-15  1.640791e-15  9.154330e-16  7.407133e-16 -4.303432e-15
#>  [85,]  1.555972e+00  4.631402e-15  9.121557e-16  4.978563e-16 -4.869765e-16
#>  [86,]  5.681012e+00  2.585207e+00  8.479392e-15  6.179944e-15 -1.635444e-15
#>  [87,]  2.478746e+00  1.413046e-15 -8.408025e-16 -2.734770e-15 -4.427744e-15
#>  [88,]  2.186060e+00  4.099178e-15 -4.897753e-16 -2.627554e-15 -2.909503e-15
#>  [89,]  2.423983e-15  2.111310e-15  1.350267e-15 -2.214816e-15 -2.419416e-15
#>  [90,]  3.472967e+00  2.138378e-15  6.526731e-16 -9.256847e-17 -1.816497e-15
#>  [91,]  1.189484e-15  8.109942e-16  1.738804e-18 -5.662094e-16 -4.996435e-15
#>  [92,]  1.841612e+00  2.213497e-15  1.000564e-15  9.591529e-16 -3.674537e-15
#>  [93,]  6.452603e-15  8.072813e-16 -2.245760e-15 -2.610955e-15 -4.062593e-15
#>  [94,] -6.628310e-16 -1.102864e-15 -1.350781e-15 -1.387803e-15 -1.495197e-15
#>  [95,]  3.565108e-15  1.780676e-16 -1.086709e-15 -1.699983e-15 -3.513667e-15
#>  [96,]  1.140540e-15  6.841413e-16  2.934610e-16 -4.746400e-16 -7.427817e-15
#>  [97,]  2.594939e+00  1.042031e-15  9.291445e-16 -2.299270e-15 -2.787588e-15
#>  [98,]  2.643976e+00  5.948439e-15  3.485904e-15  1.043678e-15 -8.596626e-16
#>  [99,]  2.308092e+00  3.603572e-15  9.745601e-16  6.150053e-17 -4.350953e-15
#> [100,]  1.468853e-15 -5.058020e-16 -6.002066e-16 -2.031432e-15 -4.123517e-15
#> 
#> $BootMatEigsCI
#>                      LowCI        HighCI
#> Dimension 1   2.793938e+01  4.701117e+01
#> Dimension 2   1.787827e+01  3.280877e+01
#> Dimension 3   1.012122e+01  2.089555e+01
#> Dimension 4   2.860550e+00  1.288242e+01
#> Dimension 5   1.201513e-15  8.376585e+00
#> Dimension 6  -6.628310e-16  5.354206e+00
#> Dimension 7  -1.102864e-15  2.741396e+00
#> Dimension 8  -2.117692e-15  2.305827e-01
#> Dimension 9  -3.662337e-15  2.119825e-15
#> Dimension 10 -9.416656e-15 -7.393633e-18
#> 
#> $FixedEigs
#>  [1] 31.0682100 23.6055006 18.4483918  9.2821913  7.5119515  6.5835363
#>  [7]  2.5235067  1.1977289  0.2820774  0.1041698
#> 
#> $BootstrappedEstEigs
#>   Dimension 1   Dimension 2   Dimension 3   Dimension 4   Dimension 5 
#>  3.535311e+01  2.436897e+01  1.556098e+01  9.029417e+00  4.209275e+00 
#>   Dimension 6   Dimension 7   Dimension 8   Dimension 9  Dimension 10 
#>  1.249569e+00  2.148553e-01  1.382211e-02 -8.306652e-16 -3.510552e-15 
#>