Boostrap for eigenvalues.

Boostrap for eigenvalues.

Usage

Boot4Eigs(
  data,
  center = FALSE,
  scale = FALSE,
  design = NULL,
  niter = 100,
  CI.perc = c(0.025, 0.975),
  suppressProgressBar = TRUE
)

Arguments

data

the data-set;

center

whether to center the data (default to FALSE), see expo.scale;

scale

whether to scale the data ( default to FALSE), see expo.scale;

design

whether there is a design;

niter

number of bootstrap iterations;

CI.perc

vector of the two levels of confidence for the left and right limits of the bootstrap confidence interval (default to c(0.025, 0.0975));

suppressProgressBar

whether to show a progress bar or not (default to FALSE).

Value

list containing four objects

Examples

Boot4Eigs(matrix(rnorm(100), 10, 10))
#> $BootMatEigs
#>        Dimension 1 Dimension 2  Dimension 3  Dimension 4  Dimension 5
#>   [1,]    37.86285    23.50438 1.253692e+01 1.013330e+01 5.962543e+00
#>   [2,]    35.61982    24.94154 1.288514e+01 1.012435e+01 4.823748e+00
#>   [3,]    35.53435    19.33826 1.266662e+01 9.925399e+00 9.211528e+00
#>   [4,]    36.02159    25.48733 1.259551e+01 9.802700e+00 4.009042e+00
#>   [5,]    31.67570    24.97045 1.406058e+01 8.907380e+00 7.665597e+00
#>   [6,]    40.31394    19.01556 1.339678e+01 8.050581e+00 6.483786e+00
#>   [7,]    37.78716    16.78243 1.670507e+01 1.067848e+01 8.046867e+00
#>   [8,]    37.75590    23.02999 1.322208e+01 7.789718e+00 6.326762e+00
#>   [9,]    32.48062    27.94993 1.428659e+01 7.840193e+00 5.189025e+00
#>  [10,]    30.20936    25.82751 1.534183e+01 1.043104e+01 5.011436e+00
#>  [11,]    35.57013    26.09610 1.507357e+01 1.326021e+01 2.489351e-15
#>  [12,]    35.84700    29.59315 1.170952e+01 1.110257e+01 1.747770e+00
#>  [13,]    31.57761    25.52143 1.596013e+01 9.187022e+00 5.599819e+00
#>  [14,]    32.63706    20.73827 1.393869e+01 1.154707e+01 5.717202e+00
#>  [15,]    39.95079    18.40725 1.291473e+01 9.896883e+00 6.786543e+00
#>  [16,]    31.33190    20.52870 1.457779e+01 9.932811e+00 7.139705e+00
#>  [17,]    33.60381    24.31179 1.532954e+01 9.581024e+00 7.173828e+00
#>  [18,]    31.83603    20.48028 1.156526e+01 9.610185e+00 8.816071e+00
#>  [19,]    32.92365    20.27826 1.484843e+01 9.256913e+00 8.363268e+00
#>  [20,]    36.21462    24.97883 1.013071e+01 9.377586e+00 6.164631e+00
#>  [21,]    37.96959    30.96153 1.658420e+01 4.484685e+00 3.425507e-15
#>  [22,]    43.33766    20.69353 1.355261e+01 1.241620e+01 7.318250e-15
#>  [23,]    32.54436    19.96451 1.617860e+01 9.719030e+00 7.865267e+00
#>  [24,]    44.71776    20.28072 1.023783e+01 7.988844e+00 5.600878e+00
#>  [25,]    29.27852    22.89755 1.674856e+01 1.220246e+01 4.748955e+00
#>  [26,]    41.57150    20.45599 1.777315e+01 1.019936e+01 4.982335e-15
#>  [27,]    41.84248    21.89168 1.154205e+01 8.459008e+00 6.264792e+00
#>  [28,]    36.61564    15.28175 1.435321e+01 1.222588e+01 8.123889e+00
#>  [29,]    34.71458    22.52490 1.511982e+01 1.133626e+01 6.304453e+00
#>  [30,]    29.37590    26.67881 1.843981e+01 1.148123e+01 4.024254e+00
#>  [31,]    32.73081    20.08593 1.529362e+01 1.184600e+01 6.708856e+00
#>  [32,]    37.85439    18.70746 1.128210e+01 9.848300e+00 7.840518e+00
#>  [33,]    33.28430    20.72747 2.031330e+01 9.317711e+00 6.357221e+00
#>  [34,]    44.56538    24.83396 1.235108e+01 8.249574e+00 2.108397e-15
#>  [35,]    57.47490    32.52510 2.949894e-15 1.922699e-15 1.002009e-15
#>  [36,]    43.09856    23.79505 1.636321e+01 6.743184e+00 4.560818e-15
#>  [37,]    42.80692    25.36361 1.608559e+01 5.743879e+00 5.458173e-15
#>  [38,]    32.03840    21.59127 1.402570e+01 1.148924e+01 8.166525e+00
#>  [39,]    45.09216    18.00296 1.340079e+01 7.615686e+00 5.888413e+00
#>  [40,]    36.16280    29.33931 2.044943e+01 4.048465e+00 4.601385e-15
#>  [41,]    27.01324    23.44532 2.186288e+01 1.119098e+01 6.487577e+00
#>  [42,]    31.19744    17.39097 1.307404e+01 1.125830e+01 9.378338e+00
#>  [43,]    36.32695    23.71043 1.500302e+01 1.192406e+01 3.035548e+00
#>  [44,]    34.06420    23.47348 1.490418e+01 1.209667e+01 3.705653e+00
#>  [45,]    38.27794    17.53561 1.403598e+01 1.276739e+01 7.383078e+00
#>  [46,]    36.15011    21.48981 1.667216e+01 8.817808e+00 6.870105e+00
#>  [47,]    36.53277    18.36320 1.566126e+01 1.098976e+01 6.402538e+00
#>  [48,]    34.81688    25.47324 1.659018e+01 1.148892e+01 1.630782e+00
#>  [49,]    37.21374    19.36813 1.889996e+01 9.536291e+00 4.981879e+00
#>  [50,]    31.11933    24.12950 2.023246e+01 1.089693e+01 3.621775e+00
#>  [51,]    34.58340    17.02207 1.545799e+01 9.381733e+00 6.814310e+00
#>  [52,]    35.94278    22.32331 1.327907e+01 8.907886e+00 5.533400e+00
#>  [53,]    34.91257    16.94712 1.361129e+01 1.173265e+01 8.938123e+00
#>  [54,]    30.25913    20.05330 1.339843e+01 1.207204e+01 8.487779e+00
#>  [55,]    45.99327    21.75829 1.109603e+01 8.298017e+00 2.854387e+00
#>  [56,]    44.79224    16.60021 1.470063e+01 1.105564e+01 2.851271e+00
#>  [57,]    33.87444    25.12308 1.670125e+01 1.096747e+01 3.333756e+00
#>  [58,]    36.29287    21.04458 1.835562e+01 8.256649e+00 6.050286e+00
#>  [59,]    34.23057    21.53742 1.113757e+01 9.227883e+00 7.775247e+00
#>  [60,]    33.84030    29.45127 1.679495e+01 9.913483e+00 2.312808e-14
#>  [61,]    33.25403    21.29352 1.541692e+01 9.194872e+00 6.745275e+00
#>  [62,]    30.91139    23.22999 1.705228e+01 9.907255e+00 8.899092e+00
#>  [63,]    36.27406    23.81281 1.815592e+01 6.673354e+00 5.083856e+00
#>  [64,]    39.52044    25.98104 1.306230e+01 7.145244e+00 4.290977e+00
#>  [65,]    33.36042    26.49019 1.709718e+01 7.916644e+00 5.135572e+00
#>  [66,]    42.66522    17.27537 1.202096e+01 8.871424e+00 7.186219e+00
#>  [67,]    28.94197    20.37999 1.688880e+01 1.087775e+01 7.293373e+00
#>  [68,]    37.86786    18.82730 1.259158e+01 1.039671e+01 6.356087e+00
#>  [69,]    42.04015    19.31059 1.552367e+01 1.312559e+01 2.157994e-15
#>  [70,]    36.48610    18.35596 1.666757e+01 1.151957e+01 4.939897e+00
#>  [71,]    33.96852    27.81515 1.512539e+01 9.395484e+00 3.695453e+00
#>  [72,]    35.39995    19.83241 1.393617e+01 8.530218e+00 8.010828e+00
#>  [73,]    31.79468    23.28745 1.702001e+01 1.270783e+01 5.190033e+00
#>  [74,]    39.82950    25.02996 1.299260e+01 9.188115e+00 2.959819e+00
#>  [75,]    36.60380    24.05577 1.907384e+01 1.026659e+01 3.267980e-15
#>  [76,]    36.69697    20.52645 1.176977e+01 1.013587e+01 5.975915e+00
#>  [77,]    32.13805    20.52689 1.374739e+01 1.108072e+01 8.163672e+00
#>  [78,]    29.62863    26.16405 1.708908e+01 8.334437e+00 6.333962e+00
#>  [79,]    40.49629    20.04405 1.485658e+01 1.073392e+01 3.869166e+00
#>  [80,]    35.21317    23.32016 1.390118e+01 9.603152e+00 6.504399e+00
#>  [81,]    46.10472    18.19561 1.619680e+01 9.502870e+00 1.171173e-15
#>  [82,]    34.16512    18.20029 1.495940e+01 1.034488e+01 6.192070e+00
#>  [83,]    33.12729    24.64495 1.085640e+01 9.282934e+00 7.420388e+00
#>  [84,]    36.03652    28.81475 1.629616e+01 8.852577e+00 2.232918e-15
#>  [85,]    33.94351    25.53207 1.396535e+01 9.593091e+00 4.981149e+00
#>  [86,]    34.84486    17.53807 1.494015e+01 1.043191e+01 6.572610e+00
#>  [87,]    40.57892    23.58974 1.585497e+01 9.976378e+00 7.778737e-15
#>  [88,]    44.24563    19.78257 1.414102e+01 7.114262e+00 4.716515e+00
#>  [89,]    38.90447    25.95374 1.509655e+01 7.907180e+00 2.138064e+00
#>  [90,]    30.39247    19.73463 1.515633e+01 1.273265e+01 9.547084e+00
#>  [91,]    27.04912    24.72395 1.496245e+01 1.168559e+01 7.124714e+00
#>  [92,]    33.31064    24.95588 1.856562e+01 6.568014e+00 3.835587e+00
#>  [93,]    47.05752    20.47508 1.261634e+01 9.851055e+00 1.688379e-15
#>  [94,]    38.80352    20.54680 1.388564e+01 9.741179e+00 4.662970e+00
#>  [95,]    51.41257    18.34665 1.356788e+01 6.672897e+00 9.734585e-15
#>  [96,]    40.85443    17.51003 1.314515e+01 9.631813e+00 5.756778e+00
#>  [97,]    36.89717    29.06545 1.384583e+01 5.573991e+00 4.617555e+00
#>  [98,]    37.12658    19.77122 1.735977e+01 8.468475e+00 5.646214e+00
#>  [99,]    29.95325    28.88087 1.511493e+01 1.028201e+01 5.768943e+00
#> [100,]    31.04501    23.50946 1.631821e+01 1.070490e+01 5.467388e+00
#>          Dimension 6   Dimension 7   Dimension 8   Dimension 9  Dimension 10
#>   [1,]  6.083115e-15  1.852205e-15  6.207303e-16  6.005105e-16 -6.901964e-16
#>   [2,]  1.605400e+00 -2.432435e-16 -1.863048e-15 -2.609877e-15 -3.000734e-15
#>   [3,]  3.323840e+00  7.075182e-15  2.160857e-15  8.659639e-16 -3.786575e-15
#>   [4,]  2.083825e+00  3.264652e-15  3.617972e-16 -5.255863e-16 -1.848208e-15
#>   [5,]  2.720303e+00  2.912889e-15  1.644666e-15 -8.043673e-16 -1.223110e-15
#>   [6,]  2.739357e+00  1.572713e-15 -1.515779e-15 -2.205984e-15 -2.824108e-15
#>   [7,]  1.252331e-15  7.950118e-16  1.076426e-16 -1.580007e-16 -4.513881e-15
#>   [8,]  1.875552e+00  5.746726e-15 -2.923863e-16 -4.171825e-16 -1.459976e-15
#>   [9,]  2.253648e+00  4.413248e-15  3.048818e-15  3.484787e-16 -1.788111e-15
#>  [10,]  1.898628e+00  1.280189e+00  2.264400e-15 -1.290356e-15 -2.376385e-15
#>  [11,] -1.312847e-15 -2.018473e-15 -2.190467e-15 -3.074217e-15 -6.621491e-15
#>  [12,]  4.300098e-15  2.799969e-16  2.614874e-16 -1.027113e-15 -2.571360e-15
#>  [13,]  2.153981e+00  2.957928e-15 -3.188755e-16 -1.576906e-15 -2.262526e-15
#>  [14,]  3.104389e+00  2.317316e+00  3.047493e-15  3.766219e-16 -3.246607e-15
#>  [15,]  2.043805e+00  4.384723e-15  4.886310e-17 -9.171456e-16 -5.658604e-15
#>  [16,]  4.326256e+00  1.590455e+00  5.723792e-01  3.101702e-15 -6.623222e-15
#>  [17,]  4.005290e-15  1.449000e-15 -1.843784e-16 -1.036929e-15 -3.065283e-15
#>  [18,]  5.892509e+00  1.799667e+00  4.578720e-15  1.712113e-15 -1.395001e-15
#>  [19,]  3.220103e+00  1.109376e+00  2.674150e-15 -3.832871e-17 -1.116676e-15
#>  [20,]  3.133623e+00  3.291660e-15  1.404370e-15 -2.993439e-15 -7.858435e-15
#>  [21,]  1.615755e-15 -8.593767e-17 -3.265078e-16 -7.363545e-16 -1.148789e-15
#>  [22,]  2.936351e-17 -4.265248e-16 -7.560045e-16 -9.720149e-16 -2.966461e-15
#>  [23,]  3.728238e+00  2.730764e-15  1.451751e-15 -2.022268e-16 -6.517263e-16
#>  [24,]  1.173975e+00  3.186729e-15  1.302646e-15 -9.801766e-16 -4.726776e-15
#>  [25,]  2.473413e+00  1.650542e+00 -1.568043e-15 -1.651604e-15 -6.380990e-15
#>  [26,]  1.670830e-15  6.817412e-16  2.524398e-16  8.454802e-17 -5.821777e-15
#>  [27,]  6.358809e-15  2.396706e-15  3.199160e-16  1.773518e-16 -1.386434e-15
#>  [28,]  3.399632e+00  7.983914e-16 -1.922178e-15 -1.957420e-15 -3.175904e-15
#>  [29,]  3.036314e-15  1.727784e-15 -2.217936e-15 -3.227013e-15 -4.288351e-15
#>  [30,]  1.068849e-15 -3.763678e-16 -1.379387e-15 -1.846875e-15 -4.397300e-15
#>  [31,]  3.334792e+00  1.306803e-15 -7.170229e-16 -9.293700e-16 -3.635680e-15
#>  [32,]  3.106333e+00  1.360898e+00  2.265313e-15  2.064557e-15 -2.124519e-15
#>  [33,]  4.667284e-15  2.742678e-15  1.864227e-15  1.261412e-15 -1.416684e-17
#>  [34,]  1.474477e-15  1.153906e-15  6.699485e-16  2.234446e-16 -5.921729e-15
#>  [35,]  4.894949e-16 -4.576820e-16 -1.300625e-15 -2.995628e-15 -1.067646e-14
#>  [36,]  2.369429e-15  1.944217e-15  1.925673e-15  3.774487e-16 -7.324151e-15
#>  [37,]  1.342674e-15  4.280044e-16  3.375446e-16  8.066318e-17 -8.120418e-16
#>  [38,]  2.688867e+00  1.215178e-15  6.410691e-16 -8.518347e-16 -1.400630e-15
#>  [39,]  5.372554e-15  1.709406e-16 -7.630811e-16 -1.012920e-15 -1.884108e-15
#>  [40,]  4.335717e-15  3.643249e-15  3.354798e-15  1.250070e-15  5.246364e-16
#>  [41,]  2.825467e-14  8.245967e-17 -4.670301e-17 -1.204567e-15 -1.533241e-15
#>  [42,]  5.569009e+00  2.131900e+00  1.672852e-15 -9.499417e-16 -4.293486e-15
#>  [43,]  8.993948e-17 -7.374398e-16 -1.445532e-15 -2.804876e-15 -3.050482e-15
#>  [44,]  1.755817e+00  5.739912e-15  1.751513e-15  1.018579e-15 -3.000850e-16
#>  [45,]  2.666606e-15  9.970458e-16 -6.632504e-17 -1.025047e-15 -2.084896e-15
#>  [46,]  8.529157e-15  7.421423e-16 -1.324955e-16 -2.858899e-16 -7.277665e-16
#>  [47,]  2.050467e+00  2.639357e-15 -7.210954e-16 -1.023304e-15 -1.696217e-15
#>  [48,]  4.382918e-15  2.698742e-15  1.076433e-15 -7.616366e-16 -1.793073e-15
#>  [49,]  3.694623e-15  1.143764e-15  1.450586e-16  4.601536e-17 -2.966142e-16
#>  [50,]  3.924790e-15  9.336509e-16 -5.863707e-16 -1.516262e-15 -2.881144e-15
#>  [51,]  4.136789e+00  2.603711e+00  2.958469e-15 -1.152876e-15 -4.600384e-15
#>  [52,]  4.013549e+00  4.143874e-15  1.803366e-15  1.092883e-15  4.992555e-17
#>  [53,]  3.858242e+00  5.731912e-15  2.612708e-15  1.785976e-15 -4.103003e-15
#>  [54,]  5.729319e+00  2.776253e-15  1.692323e-15  7.360440e-16 -2.034156e-15
#>  [55,]  6.716591e-15  1.872979e-15  1.025440e-15 -2.056790e-15 -3.189345e-15
#>  [56,]  2.807326e-15  2.623897e-15  6.249718e-16  2.689236e-16 -3.124706e-15
#>  [57,]  1.230158e-15  1.855742e-16 -4.716174e-16 -1.820131e-15 -6.400512e-15
#>  [58,]  3.164086e-15  2.058676e-15  4.074569e-16 -8.944870e-16 -2.404949e-15
#>  [59,]  3.934988e+00  2.156328e+00  3.819261e-15  8.142522e-17 -4.633316e-16
#>  [60,]  1.769991e-14  6.570701e-16  2.084145e-16  1.815456e-16 -1.604620e-16
#>  [61,]  4.095379e+00  8.675552e-16 -7.074222e-16 -1.398310e-15 -2.174784e-15
#>  [62,]  3.515408e-14  1.936021e-15 -3.640573e-17 -5.046996e-17 -5.357165e-16
#>  [63,]  2.005261e-15  1.357453e-15 -1.914038e-15 -3.830198e-15 -4.415652e-15
#>  [64,]  3.808972e-15  2.124010e-15  8.938524e-16 -8.796721e-16 -3.918992e-15
#>  [65,]  1.929955e-15  1.645641e-15  1.371966e-15 -8.440534e-16 -4.980563e-15
#>  [66,]  1.980805e+00  3.667628e-15 -1.001813e-15 -2.181925e-15 -2.600088e-15
#>  [67,]  3.908348e+00  1.709771e+00  4.348448e-15  2.102142e-15  1.200771e-15
#>  [68,]  3.017869e+00  9.425924e-01  4.438236e-17 -1.423895e-15 -4.992560e-15
#>  [69,]  7.946705e-16  4.094034e-16 -4.610379e-16 -7.702902e-16 -1.353049e-15
#>  [70,]  2.030894e+00  3.830013e-15  1.886426e-15  1.149832e-15  5.598921e-16
#>  [71,]  1.558464e-15  1.198537e-15 -1.417578e-16 -5.667162e-16 -4.412716e-15
#>  [72,]  4.290429e+00  4.644493e-15  2.297849e-15  1.013375e-15 -7.162520e-16
#>  [73,]  8.171279e-15  1.433856e-16 -1.362337e-15 -2.213306e-15 -2.247307e-15
#>  [74,]  4.065876e-15  6.397295e-16  2.332160e-16  2.967185e-17 -1.220163e-15
#>  [75,]  3.005026e-15  2.110530e-15  1.470658e-15  3.783024e-16 -2.112764e-15
#>  [76,]  4.895027e+00  3.109369e-15  9.812871e-16 -1.861886e-17 -2.612542e-15
#>  [77,]  4.343280e+00  3.564133e-15  8.659233e-16 -4.619157e-16 -1.064504e-15
#>  [78,]  2.449832e+00  1.290237e-15  2.140978e-16 -2.328844e-15 -2.891842e-15
#>  [79,]  5.336221e-15  1.757308e-15  1.143073e-15 -1.436917e-15 -5.681360e-15
#>  [80,]  1.457940e+00  4.974245e-15  1.274024e-15 -6.791366e-16 -1.118157e-15
#>  [81,]  6.174623e-16  2.506735e-16 -4.735755e-16 -2.155953e-15 -5.451869e-15
#>  [82,]  6.138246e+00  1.965128e-15  1.665478e-15  9.362003e-16 -8.983294e-16
#>  [83,]  3.406557e+00  1.261480e+00  2.297442e-15  1.963179e-16 -6.082396e-15
#>  [84,]  1.360534e-15 -6.760979e-17 -5.136034e-16 -8.974391e-16 -3.450957e-15
#>  [85,]  1.984836e+00  3.145339e-15  2.431896e-16 -9.945816e-16 -1.371976e-15
#>  [86,]  4.183661e+00  1.488736e+00  5.340450e-15 -1.478403e-15 -2.389066e-15
#>  [87,]  1.217047e-15 -4.656027e-16 -7.616921e-16 -1.637314e-15 -2.120389e-15
#>  [88,]  4.069544e-15  1.227617e-15  7.233056e-16  4.157805e-17 -1.864916e-15
#>  [89,]  2.135961e-15  6.477369e-16  4.650744e-16 -5.912167e-16 -1.232831e-15
#>  [90,]  2.436829e+00  4.556863e-15 -5.045983e-16 -1.009887e-15 -1.312235e-15
#>  [91,]  2.445425e+00  2.008753e+00  1.918520e-15 -1.640964e-15 -5.340641e-15
#>  [92,]  2.764251e+00  2.844301e-15  1.454928e-15  3.685824e-16 -1.654930e-15
#>  [93,]  2.604256e-16 -1.482477e-15 -1.920248e-15 -3.101300e-15 -4.153450e-15
#>  [94,]  2.359885e+00  1.275870e-15  1.099659e-15 -1.709395e-15 -1.969574e-15
#>  [95,]  4.064766e-15  1.580225e-15  5.685549e-16  3.750854e-16 -5.165499e-16
#>  [96,]  3.101800e+00  4.841046e-15  2.211324e-15  1.208969e-15 -2.974285e-15
#>  [97,]  2.321321e-15  1.183244e-15 -1.606775e-15 -2.364877e-15 -9.698261e-15
#>  [98,]  1.627741e+00 -1.952810e-16 -1.559671e-15 -2.036199e-15 -4.971447e-15
#>  [99,]  3.086790e-15  2.717319e-15 -4.736070e-17 -8.311539e-16 -5.664673e-15
#> [100,]  2.955042e+00  4.520771e-15  3.794738e-15  2.963291e-15 -1.432126e-15
#> 
#> $BootMatEigsCI
#>                      LowCI       HighCI
#> Dimension 1   2.704912e+01 4.705752e+01
#> Dimension 2   1.660021e+01 2.959315e+01
#> Dimension 3   1.013071e+01 2.031330e+01
#> Dimension 4   4.048465e+00 1.276739e+01
#> Dimension 5   1.171173e-15 9.211528e+00
#> Dimension 6   2.936351e-17 5.729319e+00
#> Dimension 7  -1.482477e-15 2.156328e+00
#> Dimension 8  -2.190467e-15 4.578720e-15
#> Dimension 9  -3.227013e-15 2.102142e-15
#> Dimension 10 -9.698261e-15 5.246364e-16
#> 
#> $FixedEigs
#>  [1] 29.443773847 20.319812259 10.790817460  9.666415034  8.147521585
#>  [6]  5.552814068  3.365355082  0.869018634  0.485438322  0.001863481
#> 
#> $BootstrappedEstEigs
#>   Dimension 1   Dimension 2   Dimension 3   Dimension 4   Dimension 5 
#>  3.640209e+01  2.239956e+01  1.476143e+01  9.631543e+00  4.953506e+00 
#>   Dimension 6   Dimension 7   Dimension 8   Dimension 9  Dimension 10 
#>  1.592027e+00  2.541171e-01  5.723792e-03 -6.256800e-16 -2.890710e-15 
#>