Boostrap for eigenvalues.

Boostrap for eigenvalues.

Usage

Boot4Eigs(
  data,
  center = FALSE,
  scale = FALSE,
  design = NULL,
  niter = 100,
  CI.perc = c(0.025, 0.975),
  suppressProgressBar = TRUE
)

Arguments

data

the data-set;

center

whether to center the data (default to FALSE), see expo.scale;

scale

whether to scale the data ( default to FALSE), see expo.scale;

design

whether there is a design;

niter

number of bootstrap iterations;

CI.perc

vector of the two levels of confidence for the left and right limits of the bootstrap confidence interval (default to c(0.025, 0.0975));

suppressProgressBar

whether to show a progress bar or not (default to FALSE).

Value

list containing four objects

Examples

Boot4Eigs(matrix(rnorm(100), 10, 10))
#> $BootMatEigs
#>        Dimension 1 Dimension 2 Dimension 3  Dimension 4  Dimension 5
#>   [1,]    35.43830    26.56667   11.761377 7.361424e+00 4.843705e+00
#>   [2,]    45.23072    17.49988   12.097696 6.875308e+00 4.617996e+00
#>   [3,]    37.08103    21.92555   14.499416 9.966452e+00 4.276608e+00
#>   [4,]    32.82010    26.68397   15.549761 7.391570e+00 4.784472e+00
#>   [5,]    23.63118    20.26657   16.515547 1.142986e+01 9.980221e+00
#>   [6,]    30.91753    23.24966   13.860418 1.074440e+01 6.956217e+00
#>   [7,]    38.60649    19.68749   15.806235 6.906663e+00 5.161248e+00
#>   [8,]    37.46600    24.16068   13.429745 9.291787e+00 5.651779e+00
#>   [9,]    35.08344    25.16796   15.019400 9.627215e+00 5.101984e+00
#>  [10,]    36.24043    23.31494   15.206600 8.657922e+00 6.580111e+00
#>  [11,]    37.11896    20.15965   14.260901 9.791799e+00 5.625728e+00
#>  [12,]    28.41998    22.20088   15.998077 1.339827e+01 7.984974e+00
#>  [13,]    29.75736    26.88442   12.937248 8.413649e+00 7.294641e+00
#>  [14,]    30.41584    20.11117   18.239765 1.102463e+01 6.769742e+00
#>  [15,]    27.38284    19.07817   17.235149 1.280270e+01 6.922733e+00
#>  [16,]    40.89256    22.01256   12.978524 1.108408e+01 3.032269e+00
#>  [17,]    38.37712    24.88435   14.050843 1.033326e+01 2.354421e+00
#>  [18,]    50.55927    14.71638   13.288651 8.953466e+00 2.482239e+00
#>  [19,]    35.92484    28.69736   15.727755 6.116539e+00 2.374829e+00
#>  [20,]    36.59762    29.44929   16.017095 6.275408e+00 1.660582e+00
#>  [21,]    28.77695    22.87612   19.313154 1.245055e+01 4.387464e+00
#>  [22,]    30.92200    23.89923   14.972961 1.132832e+01 6.988267e+00
#>  [23,]    40.38384    20.73335   11.487113 8.464404e+00 4.682438e+00
#>  [24,]    26.68497    25.42572   23.563264 9.148422e+00 4.798153e+00
#>  [25,]    37.70434    20.97247   17.699031 7.332148e+00 5.015491e+00
#>  [26,]    31.77901    21.65441   14.905461 1.193227e+01 8.504301e+00
#>  [27,]    40.81936    22.78232   11.839050 9.435013e+00 5.124257e+00
#>  [28,]    34.54029    21.52322   14.228403 1.115901e+01 4.605461e+00
#>  [29,]    27.40225    22.41346   16.515277 9.386937e+00 8.320440e+00
#>  [30,]    40.00959    18.83594   15.298721 8.300949e+00 5.650495e+00
#>  [31,]    35.70077    26.72976   14.054496 6.199044e+00 4.801843e+00
#>  [32,]    45.34573    26.58137   10.207044 7.865860e+00 2.946288e-15
#>  [33,]    30.49542    27.15363   15.524031 9.193632e+00 6.139775e+00
#>  [34,]    42.89227    25.17386   12.495354 5.520363e+00 3.918151e+00
#>  [35,]    50.79383    18.65013   12.323683 5.022169e+00 3.210185e+00
#>  [36,]    30.59468    22.24005   18.418035 9.303135e+00 6.330269e+00
#>  [37,]    30.97746    18.61783   16.968193 1.039101e+01 5.563343e+00
#>  [38,]    34.88676    23.97585   19.588621 6.968920e+00 4.579859e+00
#>  [39,]    29.05042    19.72496   16.001850 1.398007e+01 6.239115e+00
#>  [40,]    31.05553    26.87698   19.996796 7.283646e+00 4.787044e+00
#>  [41,]    40.05369    21.85693   13.463040 9.804356e+00 4.821980e+00
#>  [42,]    36.45949    25.92364   18.408623 5.071434e+00 4.136814e+00
#>  [43,]    34.22086    31.57545   14.077297 1.012639e+01 2.572020e-14
#>  [44,]    32.15929    21.21706   18.613742 8.896609e+00 5.262414e+00
#>  [45,]    54.53328    19.10962   13.681021 2.676077e+00 1.082308e-15
#>  [46,]    31.98216    29.53926   23.257744 3.396338e+00 1.824493e+00
#>  [47,]    32.99593    21.64097   15.854610 1.346590e+01 4.015842e+00
#>  [48,]    35.05701    25.27787   13.722619 9.617696e+00 6.324805e+00
#>  [49,]    42.79771    18.23657   11.039785 7.828037e+00 6.090129e+00
#>  [50,]    37.93319    27.90743   14.682041 7.102414e+00 2.374923e+00
#>  [51,]    31.09715    24.88248   16.784161 1.039820e+01 4.276915e+00
#>  [52,]    27.77664    22.61105   15.896434 1.272385e+01 5.107146e+00
#>  [53,]    38.88679    20.78140   14.137599 8.487171e+00 4.819992e+00
#>  [54,]    31.27026    27.25064   19.449223 8.941864e+00 3.088009e+00
#>  [55,]    33.87462    18.60775   17.942785 9.651693e+00 6.004792e+00
#>  [56,]    32.18991    29.53815   13.918195 7.159699e+00 6.073943e+00
#>  [57,]    37.77996    27.05161   18.748152 4.633421e+00 1.786863e+00
#>  [58,]    30.28857    25.97177   18.086094 8.370310e+00 4.261475e+00
#>  [59,]    31.20688    18.59511   15.283925 1.101277e+01 6.643763e+00
#>  [60,]    28.93260    21.86601   17.068247 1.014645e+01 6.121115e+00
#>  [61,]    35.98018    26.24723   13.187223 1.045159e+01 4.133776e+00
#>  [62,]    28.78425    18.45438   14.011300 1.371480e+01 7.824736e+00
#>  [63,]    32.90659    24.11712   16.149384 9.187077e+00 5.934385e+00
#>  [64,]    42.55075    20.73261   13.396725 9.367580e+00 2.222936e+00
#>  [65,]    48.89881    23.31520   11.578747 6.207252e+00 3.388082e-15
#>  [66,]    38.73183    28.72033   11.920526 6.447317e+00 4.180003e+00
#>  [67,]    30.56231    26.60678   15.537578 8.538563e+00 4.584358e+00
#>  [68,]    50.54899    18.31689   11.814446 7.720244e+00 1.599431e+00
#>  [69,]    32.29080    21.64582   14.210510 1.221581e+01 5.826949e+00
#>  [70,]    30.61148    25.58214   17.533860 8.459589e+00 5.302238e+00
#>  [71,]    35.02239    26.17501   15.706490 8.370778e+00 4.725333e+00
#>  [72,]    34.37706    22.07967   16.136034 8.691902e+00 5.836497e+00
#>  [73,]    30.55878    25.57713   13.892587 9.438984e+00 7.243861e+00
#>  [74,]    30.10731    22.77772   17.579003 9.625395e+00 6.863232e+00
#>  [75,]    30.75968    21.72815   14.441160 1.072578e+01 6.060257e+00
#>  [76,]    48.35363    21.94866   15.026573 4.671143e+00 4.604366e-15
#>  [77,]    35.47725    29.33717   18.634727 6.550855e+00 3.078806e-15
#>  [78,]    38.27372    24.02319   15.102002 8.507779e+00 4.093312e+00
#>  [79,]    42.25766    18.52016   13.303622 7.925310e+00 4.522136e+00
#>  [80,]    44.68292    19.58288   11.989117 8.413197e+00 5.331889e+00
#>  [81,]    46.33504    15.35024   14.531924 7.434970e+00 4.867962e+00
#>  [82,]    30.99826    30.13676   15.134091 6.843716e+00 3.800921e+00
#>  [83,]    33.60992    31.97194   14.951835 8.768033e+00 6.982722e-01
#>  [84,]    34.60600    22.03443   14.103283 1.213085e+01 4.663083e+00
#>  [85,]    41.82367    27.68893    9.058242 7.581854e+00 3.847312e+00
#>  [86,]    60.67069    18.44785   10.881461 5.609227e-15 8.211252e-16
#>  [87,]    36.82991    27.16849   11.818106 8.121522e+00 3.963686e+00
#>  [88,]    31.70095    27.16973   16.464650 8.397833e+00 6.266839e+00
#>  [89,]    32.18383    18.27143   16.133115 1.382941e+01 7.075407e+00
#>  [90,]    36.26854    21.65449   20.307621 8.093055e+00 3.676291e+00
#>  [91,]    32.88548    25.23461   17.721817 1.142586e+01 2.732240e+00
#>  [92,]    38.87865    25.76819   14.722724 1.063044e+01 5.039899e-15
#>  [93,]    28.96809    26.20424   16.468542 1.346748e+01 3.261157e+00
#>  [94,]    42.55935    20.75653   13.800530 6.637257e+00 6.246329e+00
#>  [95,]    28.65288    19.59362   16.923730 9.531663e+00 8.081563e+00
#>  [96,]    26.27299    21.29392   17.569794 1.161916e+01 7.421781e+00
#>  [97,]    29.08448    23.93732   16.982256 7.100485e+00 5.765132e+00
#>  [98,]    32.87346    25.75088   17.076897 8.176344e+00 4.263118e+00
#>  [99,]    27.04315    25.13347   16.222163 1.340018e+01 6.514692e+00
#> [100,]    31.80308    22.66416   15.544097 1.077996e+01 5.993946e+00
#>         Dimension 6   Dimension 7   Dimension 8   Dimension 9  Dimension 10
#>   [1,] 3.479060e+00  5.494632e-01  1.424049e-15  1.601338e-16 -3.777358e-15
#>   [2,] 3.678400e+00  2.337765e-15  8.637840e-16 -1.520284e-15 -2.780380e-15
#>   [3,] 2.250951e+00  2.996179e-15  3.015850e-16 -1.114340e-15 -3.430835e-15
#>   [4,] 2.770121e+00  1.801817e-15 -9.874231e-16 -1.598243e-15 -3.305707e-15
#>   [5,] 6.394062e+00  1.782565e+00  1.835197e-15 -1.583011e-15 -2.027336e-15
#>   [6,] 4.271767e+00  4.914562e-15  2.470641e-15  2.574576e-16 -3.672995e-15
#>   [7,] 3.831875e+00  3.039143e-15  3.364714e-16 -6.870671e-16 -1.034733e-15
#>   [8,] 1.167493e-15  4.593827e-16 -9.520145e-16 -1.106956e-15 -3.448908e-15
#>   [9,] 2.530763e-15  1.172576e-15 -1.400876e-15 -1.559924e-15 -4.102168e-15
#>  [10,] 2.887001e-15  9.478805e-16 -2.139892e-16 -4.700979e-16 -6.522187e-16
#>  [11,] 3.042970e+00  2.080330e-15  5.734538e-16 -9.925051e-17 -5.969196e-15
#>  [12,] 1.997814e+00  2.198533e-15  1.990611e-15  6.663218e-16 -7.970434e-15
#>  [13,] 4.155433e+00  5.572463e-01  5.054064e-15  2.885596e-15  5.343430e-16
#>  [14,] 2.311474e+00  1.127379e+00  3.207851e-15  3.816392e-16 -2.070747e-15
#>  [15,] 4.187411e+00  2.391004e+00  1.762119e-15  4.973650e-16 -5.172250e-16
#>  [16,] 5.535571e-15  1.043156e-15  8.184717e-16 -4.853394e-16 -2.128898e-15
#>  [17,] 1.823879e-15  7.047246e-16 -4.808288e-16 -2.882197e-15 -4.378539e-15
#>  [18,] 4.644308e-15  2.163067e-15 -3.080672e-16 -8.124681e-16 -2.743710e-15
#>  [19,] 1.158673e+00  2.784742e-15  2.414968e-15  1.804051e-15  3.557420e-16
#>  [20,] 2.215638e-15  5.267293e-16  2.085707e-16 -1.538873e-15 -1.658950e-15
#>  [21,] 2.195758e+00  6.042423e-16 -8.363277e-17 -4.883413e-15 -6.151345e-15
#>  [22,] 1.889227e+00  3.223692e-15  1.809127e-16 -3.449986e-16 -1.047109e-15
#>  [23,] 3.363560e+00  8.852912e-01  3.272207e-15  9.273628e-16 -2.037586e-15
#>  [24,] 3.794805e-01  4.263256e-14  3.085741e-16 -2.988708e-16 -4.228551e-15
#>  [25,] 1.276517e+00  5.784413e-15  2.032545e-15  1.391140e-15 -1.926136e-15
#>  [26,] 1.224551e+00  3.810167e-15  1.837419e-15 -2.799044e-16 -4.131829e-15
#>  [27,] 2.625426e-15  1.109123e-15  1.063187e-15 -1.583949e-15 -2.242303e-15
#>  [28,] 2.543015e+00  1.400600e+00  1.629772e-15  2.263986e-16 -5.408991e-15
#>  [29,] 3.829550e+00  2.132085e+00  2.789281e-15  1.505855e-15 -7.841363e-16
#>  [30,] 1.904307e+00  3.766224e-15  2.571413e-15 -1.885463e-16 -1.472684e-15
#>  [31,] 2.514087e+00  3.402116e-15  2.024745e-16 -2.358860e-16 -1.170630e-15
#>  [32,] 2.129149e-15  9.706323e-16  1.590643e-16 -7.037034e-17 -3.528331e-16
#>  [33,] 1.493509e+00  2.065691e-15  2.054418e-15  3.590106e-16 -2.924305e-15
#>  [34,] 6.208437e-15  5.061168e-16 -1.095005e-15 -1.411719e-15 -3.781207e-15
#>  [35,] 2.139269e-15  1.228683e-15 -5.820286e-16 -2.297134e-15 -2.951483e-15
#>  [36,] 3.113825e+00  1.512610e-15  3.846513e-16 -1.010710e-15 -1.610001e-15
#>  [37,] 4.658093e+00  1.864002e+00  9.600688e-01  3.337136e-16 -1.332268e-15
#>  [38,] 1.642114e-15  7.837229e-16  7.472277e-16 -7.171863e-17 -3.491028e-15
#>  [39,] 3.154855e+00  1.848730e+00  3.497687e-16 -2.186040e-15 -3.104284e-15
#>  [40,] 7.307740e-15  3.425225e-15  1.114561e-15 -6.526324e-16 -1.705264e-15
#>  [41,] 5.468688e-15 -1.717894e-16 -8.027018e-16 -8.138095e-16 -1.118353e-15
#>  [42,] 1.062430e-14 -1.744228e-16 -1.788136e-16 -5.439663e-16 -1.889951e-15
#>  [43,] 1.915259e-14  2.751415e-15  1.389578e-15  5.484573e-16 -5.035598e-16
#>  [44,] 3.850890e+00  1.657627e-15  1.470004e-15 -2.640182e-15 -4.295404e-15
#>  [45,] 5.393691e-16 -3.851424e-17 -7.390086e-17 -2.157831e-15 -2.967731e-15
#>  [46,] 4.691943e-15  1.423566e-15  1.151288e-15 -6.558614e-18 -2.268647e-15
#>  [47,] 2.026744e+00  1.333983e-15  5.924960e-16 -5.494518e-16 -4.951118e-15
#>  [48,] 1.961546e-15  1.343913e-15  8.463043e-16  4.464694e-16 -2.116811e-15
#>  [49,] 4.007770e+00  1.698055e-15  4.092459e-16 -1.747262e-15 -2.996521e-15
#>  [50,] 6.624520e-16 -7.262747e-17 -4.471197e-16 -8.448271e-16 -1.216251e-15
#>  [51,] 2.561103e+00  4.837624e-15  4.134717e-16  9.608568e-17 -6.025829e-16
#>  [52,] 3.486797e+00  2.398090e+00  1.334197e-15 -7.842416e-17 -4.892539e-15
#>  [53,] 2.887046e+00  5.004184e-15 -8.403127e-17 -7.330668e-16 -2.773877e-15
#>  [54,] 4.243879e-15  1.798763e-15 -6.839270e-16 -1.055721e-15 -2.341422e-15
#>  [55,] 2.464732e+00  1.453631e+00  3.676482e-15  3.188293e-16 -1.365782e-16
#>  [56,] 1.120105e+00  1.572016e-15  4.066454e-16 -4.044935e-16 -2.554483e-15
#>  [57,] 4.440892e-16  5.361543e-18 -5.325960e-16 -1.735254e-15 -3.364955e-15
#>  [58,] 2.062486e+00  9.592877e-01  2.793221e-15 -2.398191e-15 -2.794346e-15
#>  [59,] 5.022255e+00  2.235298e+00  2.754264e-15  2.894199e-16 -4.029644e-15
#>  [60,] 3.386635e+00  2.478944e+00  3.060358e-15  4.034527e-16 -3.091952e-15
#>  [61,] 2.156705e-15  1.793951e-15  1.596497e-15  1.098127e-15 -4.143603e-15
#>  [62,] 4.656283e+00  2.554246e+00  9.715399e-16 -4.264995e-16 -1.443659e-15
#>  [63,] 1.705436e+00  5.839546e-15  7.095521e-16 -5.529798e-16 -1.557870e-15
#>  [64,] 1.729391e+00  3.894367e-15  3.735353e-16 -8.414144e-17 -6.882475e-16
#>  [65,] 2.389488e-16 -1.365016e-17 -2.151938e-15 -2.577493e-15 -5.404576e-15
#>  [66,] 2.611024e-15  6.987729e-16 -1.855151e-15 -3.388527e-15 -4.033575e-15
#>  [67,] 3.103299e+00  1.067114e+00  1.655560e-15 -2.298614e-16 -1.754589e-15
#>  [68,] 3.789411e-15  1.564778e-15 -2.675730e-16 -1.523970e-15 -1.969395e-15
#>  [69,] 3.810102e+00  2.445804e-15  2.250852e-15 -1.637822e-16 -3.548978e-15
#>  [70,] 1.780852e+00  7.298473e-01  4.156301e-16 -4.933906e-16 -6.204103e-16
#>  [71,] 2.366627e-15  2.090559e-15  2.045077e-15 -1.813327e-15 -2.368161e-15
#>  [72,] 2.878838e+00  2.252766e-15  2.197207e-15  1.235236e-15 -1.192390e-15
#>  [73,] 3.288663e+00  3.708876e-15 -1.266224e-15 -1.876750e-15 -3.525488e-15
#>  [74,] 2.028402e+00  1.018937e+00  4.018681e-15 -3.264463e-15 -4.396076e-15
#>  [75,] 3.679587e+00  2.605381e+00  3.104282e-15  1.426711e-15  1.032365e-15
#>  [76,] 4.531648e-16  1.782684e-16  1.228171e-16 -1.160051e-15 -2.878115e-15
#>  [77,] 1.619859e-15 -9.049524e-16 -1.537827e-15 -2.039462e-15 -2.542176e-15
#>  [78,] 2.701529e-15  1.189217e-15  4.064345e-16  1.156293e-16 -1.721377e-15
#>  [79,] 3.471107e+00  1.397814e-15  1.273637e-15  1.269456e-15 -2.764063e-15
#>  [80,] 2.612079e-15  1.280306e-15  1.062624e-15 -6.549480e-16 -3.483492e-15
#>  [81,] 1.479871e+00  9.972084e-16  2.611305e-16 -8.762358e-17 -3.988071e-15
#>  [82,] 3.086248e+00  1.085899e-15  3.745796e-16  1.098361e-16 -4.705001e-15
#>  [83,] 7.105427e-15 -1.014624e-17 -7.316936e-16 -2.121936e-15 -5.637236e-15
#>  [84,] 2.462348e+00  3.533143e-16 -1.745968e-17 -1.809553e-15 -6.956188e-15
#>  [85,] 2.920837e-15 -4.554454e-16 -1.357179e-15 -2.286350e-15 -2.462703e-15
#>  [86,] 6.727586e-16 -1.043802e-15 -1.438515e-15 -2.971791e-15 -1.105472e-14
#>  [87,] 2.098288e+00  4.834436e-15  9.676644e-16  3.356427e-16 -2.516612e-15
#>  [88,] 6.405706e-15 -2.934681e-16 -4.464070e-16 -1.637773e-15 -1.937365e-15
#>  [89,] 2.506801e+00  5.921283e-15  3.318384e-15  2.972376e-15 -1.808590e-15
#>  [90,] 1.449644e-15 -6.505419e-16 -9.463393e-16 -1.307969e-15 -3.862097e-15
#>  [91,] 2.059391e-15 -5.633124e-16 -1.608765e-15 -4.347976e-15 -8.709467e-15
#>  [92,] 2.748901e-15  4.444215e-16 -4.283480e-16 -5.833523e-16 -8.192212e-16
#>  [93,] 1.630492e+00  4.832541e-15  9.539623e-16 -2.238787e-15 -3.286832e-15
#>  [94,] 5.207896e-15 -1.593025e-16 -2.959568e-16 -1.030059e-15 -2.057011e-15
#>  [95,] 5.558647e+00  1.657901e+00  3.511282e-15  3.393569e-15 -1.515183e-15
#>  [96,] 3.820101e+00  2.002258e+00  1.744369e-15 -1.238859e-15 -2.696155e-15
#>  [97,] 4.342206e+00  2.788125e+00  3.691088e-15  1.344253e-15 -5.175782e-16
#>  [98,] 1.859307e+00  4.761990e-15  2.753380e-15  1.306516e-16 -2.616509e-15
#>  [99,] 1.686343e+00  8.997628e-15  7.394089e-16  7.043680e-16  1.373826e-16
#> [100,] 3.214766e+00  6.471793e-16 -4.917436e-16 -7.211560e-16 -2.474457e-15
#> 
#> $BootMatEigsCI
#>                      LowCI       HighCI
#> Dimension 1   2.627299e+01 5.079383e+01
#> Dimension 2   1.535024e+01 3.013676e+01
#> Dimension 3   1.020704e+01 2.030762e+01
#> Dimension 4   2.676077e+00 1.371480e+01
#> Dimension 5   1.082308e-15 8.320440e+00
#> Dimension 6   4.440892e-16 5.022255e+00
#> Dimension 7  -9.049524e-16 2.554246e+00
#> Dimension 8  -1.855151e-15 4.018681e-15
#> Dimension 9  -4.347976e-15 2.885596e-15
#> Dimension 10 -8.709467e-15 3.557420e-16
#> 
#> $FixedEigs
#>  [1] 36.73523121 24.08633130 15.06463955 11.62201049  9.00726355  7.08326532
#>  [7]  3.74875144  1.53408229  0.57239325  0.00978756
#> 
#> $BootstrappedEstEigs
#>   Dimension 1   Dimension 2   Dimension 3   Dimension 4   Dimension 5 
#>  3.556058e+01  2.334718e+01  1.531565e+01  8.958240e+00  4.665633e+00 
#>   Dimension 6   Dimension 7   Dimension 8   Dimension 9  Dimension 10 
#>  1.758243e+00  3.848743e-01  9.600688e-03 -6.168120e-16 -2.760284e-15 
#>