Boostrap for eigenvalues.

Boostrap for eigenvalues.

Usage

Boot4Eigs(
  data,
  center = FALSE,
  scale = FALSE,
  design = NULL,
  niter = 100,
  CI.perc = c(0.025, 0.975),
  suppressProgressBar = TRUE
)

Arguments

data

the data-set;

center

whether to center the data (default to FALSE), see expo.scale;

scale

whether to scale the data ( default to FALSE), see expo.scale;

design

whether there is a design;

niter

number of bootstrap iterations;

CI.perc

vector of the two levels of confidence for the left and right limits of the bootstrap confidence interval (default to c(0.025, 0.0975));

suppressProgressBar

whether to show a progress bar or not (default to FALSE).

Value

list containing four objects

Examples

Boot4Eigs(matrix(rnorm(100), 10, 10))
#> $BootMatEigs
#>        Dimension 1 Dimension 2 Dimension 3  Dimension 4   Dimension 5
#>   [1,]    45.92214    17.45821   10.425026 8.635300e+00  6.060239e+00
#>   [2,]    51.73615    16.94392   12.383692 8.936240e+00  9.346769e-16
#>   [3,]    36.97643    23.55321   19.556309 5.548324e+00  4.365728e+00
#>   [4,]    36.55655    19.64505   14.983354 9.841692e+00  6.391428e+00
#>   [5,]    44.37303    19.95144   17.861598 7.813938e+00  3.171836e-15
#>   [6,]    33.41267    20.17024   14.962481 9.497972e+00  6.760403e+00
#>   [7,]    42.00062    19.41704   14.014989 9.348158e+00  5.219188e+00
#>   [8,]    51.59780    30.55858    7.843611 5.891788e-15 -9.913785e-16
#>   [9,]    52.79751    14.56512   10.321119 9.412744e+00  2.903503e+00
#>  [10,]    33.02853    23.14639   17.752383 1.023420e+01  5.838501e+00
#>  [11,]    33.51194    25.64100   17.331951 8.155213e+00  5.359890e+00
#>  [12,]    38.99384    18.82236   12.870974 1.191139e+01  6.606412e+00
#>  [13,]    33.54536    28.79162   13.803300 9.192487e+00  3.380879e+00
#>  [14,]    34.86943    25.44183   16.334045 1.055808e+01  2.796613e+00
#>  [15,]    52.56898    24.86430    8.939847 3.626874e+00  1.610304e-15
#>  [16,]    36.68308    22.72653   15.450018 1.062532e+01  4.515054e+00
#>  [17,]    50.47298    17.94995   12.389899 5.629194e+00  2.950099e+00
#>  [18,]    42.43313    21.42684   17.883534 8.256490e+00  1.652204e-15
#>  [19,]    45.88411    17.06448   10.608062 6.786347e+00  5.730661e+00
#>  [20,]    35.32254    17.27733   12.302306 1.065611e+01  6.241044e+00
#>  [21,]    41.95051    17.45202   14.945336 1.098605e+01  2.888392e+00
#>  [22,]    37.51828    15.68808   14.702810 1.026738e+01  6.507063e+00
#>  [23,]    34.53191    27.02369   12.770155 9.778210e+00  5.896033e+00
#>  [24,]    42.38929    26.97031   14.167579 6.472820e+00  7.442587e-15
#>  [25,]    37.55364    20.37296   13.551559 1.008178e+01  6.865936e+00
#>  [26,]    31.63586    23.31197   13.974683 1.083875e+01  5.872347e+00
#>  [27,]    43.80712    24.62673    9.625723 8.031486e+00  3.908938e+00
#>  [28,]    45.25409    24.16339   12.960563 7.621961e+00  4.045247e-15
#>  [29,]    37.07997    22.26969   14.451658 8.686932e+00  5.501555e+00
#>  [30,]    42.56744    23.53753   14.845877 6.721075e+00  2.328080e+00
#>  [31,]    55.66844    17.23946   10.903000 3.695773e+00  2.493328e+00
#>  [32,]    47.25103    16.68788   12.678161 9.122319e+00  3.587475e+00
#>  [33,]    37.67675    19.06340   18.211299 9.695884e+00  4.209528e+00
#>  [34,]    37.81661    23.04549   16.601178 9.488042e+00  3.048673e+00
#>  [35,]    39.13420    19.07122   14.598446 9.647030e+00  4.948191e+00
#>  [36,]    41.89410    17.54471   13.042944 1.122063e+01  4.611098e+00
#>  [37,]    36.11023    20.28308   14.243775 1.020259e+01  4.343350e+00
#>  [38,]    47.23816    16.56093   14.067865 8.125860e+00  2.466366e+00
#>  [39,]    34.00515    29.37203   16.945390 7.375940e+00  2.301494e+00
#>  [40,]    42.84227    23.46645   16.868950 6.822331e+00  2.879855e-15
#>  [41,]    36.04653    18.68094   14.188964 9.647280e+00  6.623428e+00
#>  [42,]    49.89621    17.52968   12.497640 7.723398e+00  2.353073e+00
#>  [43,]    34.07477    19.88538   14.072308 8.510425e+00  6.480033e+00
#>  [44,]    31.97609    21.18685   14.768897 1.202774e+01  7.705826e+00
#>  [45,]    39.68387    26.92956   17.636452 5.750120e+00  3.047643e-15
#>  [46,]    40.67135    20.28708   15.108087 7.800523e+00  4.908727e+00
#>  [47,]    39.08187    22.70676   16.024344 9.167123e+00  3.019903e+00
#>  [48,]    35.07969    32.42242   15.822223 6.675660e+00  1.661707e-16
#>  [49,]    41.33071    19.04894   13.405567 1.184731e+01  2.843700e+00
#>  [50,]    50.73448    20.52720   13.297126 5.441185e+00  3.136419e-15
#>  [51,]    51.85912    20.05734   12.187253 5.896286e+00  1.457704e-14
#>  [52,]    36.89892    22.25684   17.021373 9.965796e+00  3.857070e+00
#>  [53,]    36.31705    23.12206   13.887157 1.140885e+01  3.141542e+00
#>  [54,]    31.79084    17.19688   16.324163 1.471940e+01  5.661317e+00
#>  [55,]    39.44260    19.40971   13.149602 8.759349e+00  6.623946e+00
#>  [56,]    56.41873    18.95589   10.443765 4.181617e+00  5.905369e-15
#>  [57,]    34.34918    26.61735   18.398260 1.063521e+01  2.283803e-14
#>  [58,]    26.57356    24.17138   18.763188 1.025148e+01  6.608848e+00
#>  [59,]    37.61911    22.46153   16.845828 8.124580e+00  2.929033e+00
#>  [60,]    45.52055    16.45068   10.827108 9.060609e+00  6.155938e+00
#>  [61,]    39.98180    16.82413   13.614442 1.097681e+01  6.824770e+00
#>  [62,]    40.14811    18.72767   13.127399 1.137290e+01  6.233842e+00
#>  [63,]    52.89134    15.15832   10.908222 7.520426e+00  2.625990e+00
#>  [64,]    31.43373    20.38062   15.920645 1.142230e+01  6.192224e+00
#>  [65,]    41.85414    19.75892   14.107034 1.130247e+01  2.977444e+00
#>  [66,]    30.10733    21.53713   19.099395 1.313850e+01  5.336530e+00
#>  [67,]    38.65831    27.73259   14.968529 8.640572e+00  3.932131e-15
#>  [68,]    31.25808    22.89436   18.741703 1.352889e+01  3.576966e+00
#>  [69,]    41.82436    20.36218   13.148534 8.902289e+00  5.335265e+00
#>  [70,]    37.65715    21.94968   16.909075 1.348409e+01  2.863265e-15
#>  [71,]    38.88585    27.69600   12.943089 7.158898e+00  3.316159e+00
#>  [72,]    34.34193    29.51367   12.007133 9.972594e+00  4.164667e+00
#>  [73,]    49.49259    15.43216   13.034957 9.494906e+00  2.545390e+00
#>  [74,]    35.81592    21.52038   17.086922 1.079087e+01  4.785901e+00
#>  [75,]    35.93216    24.34197   12.329451 8.267766e+00  4.616762e+00
#>  [76,]    26.57472    23.21920   16.157942 1.492914e+01  5.522626e+00
#>  [77,]    29.31413    20.53647   17.523481 1.517012e+01  5.589366e+00
#>  [78,]    52.11457    15.06049   11.199303 6.958286e+00  4.667348e+00
#>  [79,]    56.79230    18.25099    9.376465 5.580241e+00  3.171030e-15
#>  [80,]    36.74655    24.33911   22.405924 6.508411e+00  9.022772e-15
#>  [81,]    38.64107    17.18124   15.987147 1.101454e+01  7.176001e+00
#>  [82,]    33.80926    21.83991   13.626010 8.757437e+00  6.073213e+00
#>  [83,]    35.23175    26.94139   17.313141 8.199713e+00  2.314004e+00
#>  [84,]    33.49429    20.11429   12.449792 1.075839e+01  8.095879e+00
#>  [85,]    36.94935    19.34817   14.625883 1.042515e+01  4.858873e+00
#>  [86,]    36.75968    18.38857   15.160760 1.180156e+01  5.102930e+00
#>  [87,]    41.79393    19.19844   14.234858 1.091213e+01  3.860630e+00
#>  [88,]    31.42532    28.65598   14.374599 9.031616e+00  6.512490e+00
#>  [89,]    29.82086    26.94947   17.195898 9.765276e+00  5.829646e+00
#>  [90,]    34.74491    20.90741   13.138895 8.263147e+00  6.870929e+00
#>  [91,]    37.04038    18.20011   15.602267 1.090049e+01  6.017064e+00
#>  [92,]    28.43137    19.76210   16.316659 1.341280e+01  6.919699e+00
#>  [93,]    39.31915    23.76607   18.055995 7.040462e+00  1.818320e+00
#>  [94,]    48.18579    17.06713   13.356689 6.762620e+00  3.420632e+00
#>  [95,]    53.45729    15.77767    9.765557 6.465539e+00  2.781211e+00
#>  [96,]    39.11530    20.02481   15.078424 1.037837e+01  4.136043e+00
#>  [97,]    36.29709    21.93211   19.398932 7.790789e+00  3.228954e+00
#>  [98,]    34.23553    26.85151   15.152983 1.196356e+01  1.796419e+00
#>  [99,]    35.52922    23.70611   13.653328 1.025579e+01  5.946373e+00
#> [100,]    36.39742    16.27985   13.639473 1.055414e+01  6.931485e+00
#>          Dimension 6   Dimension 7   Dimension 8   Dimension 9  Dimension 10
#>   [1,]  1.499090e+00  1.947012e-15 -1.953162e-15 -2.759711e-15 -5.577338e-15
#>   [2,] -7.173401e-16 -8.075758e-16 -1.314644e-15 -2.405622e-15 -1.336289e-14
#>   [3,]  4.554316e-15  1.051899e-15  2.585769e-16 -3.925135e-16 -9.788816e-16
#>   [4,]  2.581926e+00  8.371760e-15  1.949830e-15  8.962596e-16 -2.051728e-15
#>   [5,]  3.769280e-16 -1.034893e-16 -4.157225e-16 -6.659109e-16 -1.100294e-15
#>   [6,]  3.566703e+00  1.629527e+00  3.876354e-15 -1.906945e-15 -2.252026e-15
#>   [7,]  3.618823e-15 -4.216263e-16 -1.080532e-15 -1.498436e-15 -5.646901e-15
#>   [8,] -1.028909e-15 -1.399683e-15 -1.709838e-15 -2.058461e-15 -3.624512e-15
#>   [9,]  7.219121e-15  3.531464e-15  1.056881e-15  7.852805e-16 -7.059372e-16
#>  [10,]  5.497782e-15  2.206666e-15 -1.024886e-15 -2.555366e-15 -3.725640e-15
#>  [11,]  5.516277e-15  1.366066e-15  1.008874e-15 -4.263643e-16 -3.478418e-15
#>  [12,]  7.950243e-01 -2.208417e-16 -4.021202e-16 -1.582119e-15 -1.230523e-14
#>  [13,]  1.286361e+00  3.923926e-15  1.812430e-15 -1.712498e-15 -3.676419e-15
#>  [14,]  2.540673e-15  1.330396e-15  2.410946e-16 -7.344257e-16 -3.746750e-15
#>  [15,]  1.247025e-15  1.079117e-15 -8.786912e-16 -3.037913e-15 -3.517054e-15
#>  [16,]  5.860285e-15  1.002994e-15  4.988088e-16  4.000487e-16 -1.715672e-15
#>  [17,]  6.078822e-01 -5.311956e-16 -6.982205e-16 -2.103576e-15 -6.564649e-15
#>  [18,]  2.022903e-16 -3.250187e-17 -1.317570e-15 -3.144720e-15 -1.038393e-14
#>  [19,]  3.100417e+00  8.259248e-01  2.364583e-15 -1.476869e-16 -3.727791e-15
#>  [20,]  4.799417e+00  1.736644e+00  1.664609e+00  3.001158e-16 -3.907299e-15
#>  [21,]  1.777695e+00  2.412007e-16  1.108127e-16 -1.111180e-15 -4.422913e-15
#>  [22,]  4.775772e+00  5.406170e-01  5.077748e-15  2.274433e-15 -1.111390e-15
#>  [23,]  3.551467e-15  6.588880e-16 -5.367823e-16 -1.126255e-15 -1.413306e-15
#>  [24,]  1.721570e-15  4.544394e-16 -6.303340e-16 -1.767427e-15 -2.248888e-15
#>  [25,]  1.574133e+00  7.261357e-15  1.812604e-15 -3.658718e-16 -4.995909e-16
#>  [26,]  2.876500e+00  1.489893e+00  7.452398e-15  1.856515e-15  3.710592e-16
#>  [27,]  3.841927e-15  5.821643e-16 -1.515389e-15 -1.776173e-15 -3.115402e-15
#>  [28,]  3.027580e-15  1.800851e-15 -1.237207e-15 -1.593317e-15 -4.273283e-15
#>  [29,]  2.010190e+00  3.676743e-15  6.482223e-16 -7.160553e-16 -3.726173e-15
#>  [30,]  2.675513e-15  2.238105e-15  2.055551e-15  2.546685e-16 -8.865420e-15
#>  [31,]  2.882711e-15  1.705555e-15  1.461350e-15  1.194010e-15 -7.726922e-15
#>  [32,]  6.731334e-01  2.389993e-15 -6.149629e-17 -2.745842e-16 -4.071581e-15
#>  [33,]  1.143145e+00  1.739666e-15  9.897976e-19 -1.386830e-16 -1.863440e-16
#>  [34,]  2.271382e-15  3.080089e-16 -1.143324e-15 -2.022769e-15 -3.290957e-15
#>  [35,]  2.600914e+00  2.287110e-15 -1.828979e-17 -1.310665e-15 -2.199130e-15
#>  [36,]  1.686511e+00  3.194197e-15  3.700828e-16 -9.595484e-16 -3.889280e-15
#>  [37,]  3.073282e+00  1.743701e+00  9.172961e-16 -1.249156e-15 -1.300023e-15
#>  [38,]  1.540809e+00  1.660978e-15  1.354681e-16  1.185845e-16 -1.489519e-15
#>  [39,]  3.170987e-15  2.836716e-16 -1.643670e-15 -1.800132e-15 -3.118855e-15
#>  [40,]  2.078878e-15  2.017456e-15 -1.173227e-16 -1.653550e-15 -2.876069e-15
#>  [41,]  3.123896e+00  1.688964e+00  3.972676e-15  2.526552e-15 -1.454762e-15
#>  [42,]  2.106310e-15  1.327699e-15  3.122185e-16 -1.108211e-15 -1.848563e-15
#>  [43,]  5.553961e+00  1.423116e+00 -5.571865e-16 -9.671954e-16 -5.784604e-15
#>  [44,]  2.334593e+00  5.424883e-15  1.639048e-15  5.039911e-16  3.319147e-16
#>  [45,]  2.191697e-15  1.611886e-15  7.462533e-16  1.324459e-17 -2.769356e-15
#>  [46,]  1.224237e+00  2.524335e-15 -1.486445e-16 -3.496166e-16 -4.120515e-15
#>  [47,]  8.034875e-16  7.781905e-16  1.455926e-16 -1.401439e-15 -4.413544e-15
#>  [48,] -7.303727e-16 -1.438608e-15 -1.541854e-15 -2.014355e-15 -7.906825e-15
#>  [49,]  1.523782e+00  1.948394e-15 -9.630168e-16 -1.170112e-15 -5.381614e-15
#>  [50,]  1.821700e-15  1.411054e-15  3.246199e-16  7.003598e-18 -5.962129e-15
#>  [51,]  1.987553e-15  3.453030e-16  2.419186e-16 -8.500379e-16 -2.415587e-15
#>  [52,]  2.766348e-15  2.369929e-15 -1.571208e-15 -1.741092e-15 -2.845120e-15
#>  [53,]  2.123341e+00  8.104714e-15  2.011775e-15  1.971587e-15 -6.670886e-16
#>  [54,]  3.925334e+00  3.820683e-01  9.076946e-16 -7.733278e-16 -2.335297e-15
#>  [55,]  2.614792e+00  2.210866e-15  7.087585e-16  6.417556e-16 -5.732611e-15
#>  [56,]  2.400807e-15  6.006084e-16 -1.901338e-16 -4.182650e-16 -4.878404e-16
#>  [57,]  1.749066e-14  8.993331e-16  5.694941e-16 -4.628396e-16 -1.988796e-15
#>  [58,]  3.631544e+00  3.349472e-15  2.439503e-15  7.551520e-16 -9.051054e-15
#>  [59,]  2.019922e+00  7.866352e-15  2.793602e-15  1.255774e-15 -2.721795e-16
#>  [60,]  1.985124e+00  2.019844e-15  1.264596e-15  5.886638e-17 -3.920411e-15
#>  [61,]  1.778050e+00  1.071688e-15  1.034976e-15 -6.868588e-16 -5.039939e-15
#>  [62,]  3.900773e-01  5.906488e-15  5.544058e-16 -6.747352e-16 -7.702676e-16
#>  [63,]  8.957054e-01  1.706709e-15  6.547972e-16  5.248424e-16 -1.112506e-15
#>  [64,]  2.963270e+00  1.687215e+00  2.066047e-15  8.315170e-16 -2.787908e-15
#>  [65,] -4.606731e-18 -8.323281e-16 -1.069902e-15 -1.579341e-15 -1.006523e-14
#>  [66,]  7.811145e-01  4.197476e-15 -2.467371e-15 -2.537335e-15 -2.852011e-15
#>  [67,]  3.442313e-15  3.408104e-15  2.228685e-15  1.679746e-15  1.373180e-15
#>  [68,]  8.064406e-15  5.685696e-16  5.271213e-16  3.426666e-16 -3.064120e-15
#>  [69,]  4.273725e-01  4.722398e-16  3.559770e-16 -2.066858e-15 -2.124629e-15
#>  [70,]  1.053201e-15  9.756316e-16 -2.841800e-16 -1.237817e-15 -1.834223e-15
#>  [71,]  2.545876e-15  7.446048e-16 -1.867807e-16 -5.123131e-16 -4.076192e-15
#>  [72,]  1.465307e-15  7.702277e-16 -1.634972e-15 -1.945381e-15 -2.307137e-15
#>  [73,]  3.890881e-15  2.745419e-15  1.926612e-15 -1.609806e-15 -2.944708e-15
#>  [74,]  4.612690e-15  2.391097e-15  9.124799e-16  8.488585e-16 -1.425756e-15
#>  [75,]  4.511887e+00  1.773081e-15 -4.504529e-16 -1.857614e-15 -2.351072e-15
#>  [76,]  3.243809e+00  3.525685e-01  1.837611e-15 -4.476600e-16 -1.393521e-15
#>  [77,]  1.866441e+00  4.259482e-15  3.755891e-15  5.836382e-16 -4.994310e-16
#>  [78,]  1.527625e-15 -4.008906e-16 -4.828677e-16 -6.353636e-16 -3.351800e-15
#>  [79,]  1.099368e-15  3.533213e-16 -1.003149e-15 -1.942745e-15 -3.361270e-15
#>  [80,]  1.372056e-15 -7.835524e-16 -2.146084e-15 -2.444083e-15 -3.933744e-15
#>  [81,]  8.149654e-15  1.374063e-16 -5.197730e-16 -2.225680e-15 -2.286562e-15
#>  [82,]  3.451615e+00  2.209796e+00  2.327624e-01  4.440892e-16 -5.027641e-16
#>  [83,]  4.754199e-15  1.965251e-15  1.392608e-15 -1.856881e-15 -2.482740e-15
#>  [84,]  4.726981e+00  3.603798e-01  3.224978e-15  2.899151e-15 -1.650219e-15
#>  [85,]  3.467150e+00  3.254228e-01  1.568141e-15 -5.273066e-16 -3.490800e-15
#>  [86,]  2.786499e+00  1.348701e-15  7.226269e-16 -2.175231e-17 -2.752271e-15
#>  [87,]  6.396101e-15  2.427772e-16 -9.784969e-16 -1.885464e-15 -3.007810e-15
#>  [88,]  7.115319e-15  1.148408e-15  7.603378e-17 -2.412832e-16 -1.185499e-15
#>  [89,]  4.388516e-01  5.524538e-16 -9.736098e-16 -3.069108e-15 -3.790888e-15
#>  [90,]  4.408587e+00  1.666120e+00 -1.800133e-16 -8.396427e-16 -6.305973e-15
#>  [91,]  2.239685e+00  1.579667e-15  1.468436e-15  7.252659e-16 -5.891420e-15
#>  [92,]  5.157377e+00  2.946371e-15  2.503971e-15  1.603166e-15  7.495113e-16
#>  [93,]  3.120368e-15  1.900663e-15  1.204254e-15 -2.829071e-16 -3.066555e-15
#>  [94,]  1.207144e+00  3.819715e-16  3.005933e-16 -2.320144e-15 -5.825579e-15
#>  [95,]  1.752734e+00 -2.943004e-16 -5.365429e-16 -5.767817e-16 -1.015636e-14
#>  [96,]  1.267057e+00  5.430594e-15 -2.360881e-16 -2.559532e-15 -3.541167e-15
#>  [97,]  1.352123e+00  7.656759e-16 -1.360309e-15 -1.470551e-15 -1.093308e-14
#>  [98,]  2.002548e-15  9.729641e-16 -8.903026e-16 -1.554127e-15 -1.805376e-15
#>  [99,]  9.091785e-01  7.839627e-15  5.330008e-16  3.969171e-16 -5.019669e-16
#> [100,]  4.280517e+00  1.917110e+00  4.730218e-15  8.833456e-16  7.242059e-16
#> 
#> $BootMatEigsCI
#>                      LowCI       HighCI
#> Dimension 1   2.657472e+01 5.566844e+01
#> Dimension 2   1.506049e+01 2.951367e+01
#> Dimension 3   8.939847e+00 1.939893e+01
#> Dimension 4   3.626874e+00 1.471940e+01
#> Dimension 5   1.661707e-16 7.176001e+00
#> Dimension 6  -7.303727e-16 4.799417e+00
#> Dimension 7  -1.399683e-15 1.743701e+00
#> Dimension 8  -2.146084e-15 7.452398e-15
#> Dimension 9  -3.069108e-15 2.274433e-15
#> Dimension 10 -1.230523e-14 7.242059e-16
#> 
#> $FixedEigs
#>  [1] 3.366217e+01 1.531314e+01 1.459126e+01 1.218047e+01 8.103911e+00
#>  [6] 5.862069e+00 1.564668e+00 1.551069e+00 8.591714e-01 3.433889e-04
#> 
#> $BootstrappedEstEigs
#>   Dimension 1   Dimension 2   Dimension 3   Dimension 4   Dimension 5 
#>  3.964481e+01  2.131199e+01  1.449582e+01  9.167168e+00  3.898119e+00 
#>   Dimension 6   Dimension 7   Dimension 8   Dimension 9  Dimension 10 
#>  1.263327e+00  1.997907e-01  1.897371e-02 -6.828615e-16 -3.421549e-15 
#>