Boostrap for eigenvalues.

Boostrap for eigenvalues.

Usage

Boot4Eigs(
  data,
  center = FALSE,
  scale = FALSE,
  design = NULL,
  niter = 100,
  CI.perc = c(0.025, 0.975),
  suppressProgressBar = TRUE
)

Arguments

data

the data-set;

center

whether to center the data (default to FALSE), see expo.scale;

scale

whether to scale the data ( default to FALSE), see expo.scale;

design

whether there is a design;

niter

number of bootstrap iterations;

CI.perc

vector of the two levels of confidence for the left and right limits of the bootstrap confidence interval (default to c(0.025, 0.0975));

suppressProgressBar

whether to show a progress bar or not (default to TRUE).

Value

list containing four objects

Examples

Boot4Eigs(matrix(rnorm(100), 10, 10))
#> $BootMatEigs
#>        Dimension 1 Dimension 2 Dimension 3 Dimension 4 Dimension 5  Dimension 6
#>   [1,]    49.29892    23.21040    13.79264    9.722883   7.0589862 4.513704e+00
#>   [2,]    34.82888    28.13221    22.93578   14.975943   9.6449060 2.454669e+00
#>   [3,]    51.52945    33.56847    18.87343   11.587388   8.2415664 2.103799e+00
#>   [4,]    47.74503    33.90597    18.60761   10.686629   5.5711393 2.516557e+00
#>   [5,]    51.16771    34.29879    22.45618   11.157762   2.0331562 1.447602e-14
#>   [6,]    37.78427    26.71384    16.94544    6.533132   3.5157182 1.294913e-15
#>   [7,]    35.95322    30.49146    21.43603   10.695549   8.6088950 3.885131e+00
#>   [8,]    41.24549    22.35241    13.09550    9.726651   5.6622865 1.732714e+00
#>   [9,]    52.63455    38.78743    14.99098   11.000852   7.5199761 3.836800e+00
#>  [10,]    52.97880    28.76513    20.39169    5.931515   3.0437006 3.926259e-15
#>  [11,]    43.37035    25.57171    15.06254    6.820205   5.4477749 1.568256e+00
#>  [12,]    47.78703    16.72573    12.57525    9.237200   7.0334599 4.827604e+00
#>  [13,]    40.89567    16.73294    12.48080   11.842346   7.6310652 4.632831e+00
#>  [14,]    36.77963    24.36312    18.09096   11.705443   9.8855977 7.084019e+00
#>  [15,]    51.99347    30.05585    19.65155   10.612859   3.0090053 1.457330e+00
#>  [16,]    33.22198    22.27023    16.24608   14.219803  13.3950966 3.604201e+00
#>  [17,]    40.21500    28.09446    24.69719   10.009817   9.3505480 4.286795e+00
#>  [18,]    44.04500    23.17545    21.88041   13.247177   9.3739434 6.123809e+00
#>  [19,]    47.15794    23.64471    21.55615   13.455727  10.6364723 4.138711e-15
#>  [20,]    27.18382    23.61411    19.54843   18.342507  11.7276104 5.287261e+00
#>  [21,]    52.37026    31.48393    16.89074   10.924080   4.8526448 2.523123e+00
#>  [22,]    35.54154    28.48078    18.50165   10.349875   6.8629778 1.698969e+00
#>  [23,]    45.03202    34.24432    15.22087    9.842536   6.2455757 4.481337e-15
#>  [24,]    52.54726    32.26918    12.36220   10.590378   9.1273826 5.167368e+00
#>  [25,]    33.06554    28.29394    19.82039   17.695748   7.4053211 1.787217e+00
#>  [26,]    39.10245    30.59168    18.20187    7.898247   4.3399943 3.251803e-01
#>  [27,]    37.42050    28.53859    23.72583   14.845915   4.9789513 3.668533e+00
#>  [28,]    45.55039    26.48533    12.94461    7.703493   3.4410462 4.945414e-16
#>  [29,]    43.68197    27.81466    21.91408   15.908258  10.9266781 1.388571e-14
#>  [30,]    41.07508    37.82332    19.57914   12.807668   6.3640074 2.079356e+00
#>  [31,]    35.20423    19.80902    18.65714   13.671678   8.1920681 2.510851e+00
#>  [32,]    43.34958    26.40390    13.71215   11.851129   8.2637610 5.130876e+00
#>  [33,]    31.62038    24.67835    20.38996   14.659736   3.7487752 1.523223e+00
#>  [34,]    46.31601    32.14356    18.38156    6.482900   4.3195050 1.417465e+00
#>  [35,]    39.77908    28.42835    10.50367    8.177039   3.7617916 1.889780e+00
#>  [36,]    37.14344    23.65301    15.78413   13.411859   6.8722980 2.604405e+00
#>  [37,]    45.51273    26.38525    11.11202    6.494853   3.0142128 1.183669e-15
#>  [38,]    51.12537    24.44084    12.28468   10.301886   6.1668038 2.432109e+00
#>  [39,]    38.14942    28.91576    17.65908   14.127471   9.0367280 2.408207e+00
#>  [40,]    37.60678    22.81703    19.30402   15.396190   4.4861138 2.382096e+00
#>  [41,]    39.17381    33.49649    18.97183   10.640237   4.7751544 2.302201e+00
#>  [42,]    37.40342    30.56961    15.74679   12.800056   5.5350198 2.084574e+00
#>  [43,]    54.99507    21.36405    12.88038    6.709822   3.3551483 3.249843e-01
#>  [44,]    40.23574    30.18264    18.19123   12.872660   5.3546758 2.964329e+00
#>  [45,]    46.02880    27.28319    21.80136    9.615687   5.4243503 1.621797e+00
#>  [46,]    29.56889    25.20183    16.40690    6.716821   5.0762899 1.752557e+00
#>  [47,]    54.14290    29.46083    18.67609   11.306542   6.3682788 1.730161e+00
#>  [48,]    35.75660    23.83882    17.62190   11.130159   9.3270871 4.758655e+00
#>  [49,]    32.59933    21.66431    19.16462   12.629823   7.7454864 4.014614e+00
#>  [50,]    32.46876    24.45082    20.94079   13.772153   6.6036958 3.332833e+00
#>  [51,]    44.06151    24.96750    14.15263   10.690392   5.0985036 3.315934e-15
#>  [52,]    33.44391    27.47646    19.67928    8.759684   3.2919223 2.311950e+00
#>  [53,]    41.92142    33.97982    12.50051    6.021676   1.8319077 2.529152e-15
#>  [54,]    51.83418    26.25000    20.52804    8.932518   5.9648099 2.413499e+00
#>  [55,]    48.76566    20.85230    17.91255    9.330214   3.0517732 1.894557e+00
#>  [56,]    43.79065    33.53697    12.97198    5.703987   0.3320913 1.407253e-15
#>  [57,]    37.47111    24.56233    19.78349    8.256053   4.8318573 2.613587e+00
#>  [58,]    39.25238    27.46421    17.85792   10.105802   6.4741287 3.843174e+00
#>  [59,]    47.96991    20.24278    17.22969   11.169184  10.4590799 4.127130e+00
#>  [60,]    43.72476    25.47659    18.83690   14.337158   4.6694283 1.301446e+00
#>  [61,]    36.45279    25.17950    18.09334   15.875044  10.3543237 7.143058e+00
#>  [62,]    36.96783    31.11384    21.21243    9.370201   2.8440727 2.774091e+00
#>  [63,]    37.44801    31.66648    16.68280    6.684982   3.8377735 4.169124e-17
#>  [64,]    55.91548    15.86041    13.95665   10.322187   4.6744577 6.762185e-01
#>  [65,]    49.70174    20.49020    15.99700   10.927898   9.7542173 1.821083e+00
#>  [66,]    38.52326    32.71139    15.95613    6.699965   4.9385287 2.617487e-15
#>  [67,]    30.89213    23.40933    18.78878   11.817833   8.8010962 3.256117e+00
#>  [68,]    49.22911    30.16751    15.33762   10.232459   4.4770425 6.841893e-15
#>  [69,]    42.01442    23.25139    18.15124   10.467516   8.8620796 3.566707e+00
#>  [70,]    46.99446    30.42347    17.49430   12.239559   5.3663191 2.074593e+00
#>  [71,]    32.71199    22.43157    19.30840   16.599505   8.8369464 6.032700e+00
#>  [72,]    52.02149    17.00875    12.39513   11.568226   4.1688917 1.932637e+00
#>  [73,]    37.88441    25.97869    20.19728   13.536057   6.8661406 3.773115e+00
#>  [74,]    36.50843    18.85183    17.71576   11.854522  10.0647120 6.462828e+00
#>  [75,]    36.84293    26.92381    21.88119   10.775889   6.9274206 2.436562e+00
#>  [76,]    35.64680    21.01596    18.17102   13.917486   7.3968432 4.013801e+00
#>  [77,]    42.34280    25.78806    20.06037   16.913839   5.7822543 2.873241e+00
#>  [78,]    33.55510    26.93986    25.25759   18.206033   4.4872132 4.246682e+00
#>  [79,]    41.06465    21.20133    19.65569   11.941594   6.1595365 2.304038e+00
#>  [80,]    28.68819    26.37170    20.14173    8.917581   5.0981772 6.117996e-01
#>  [81,]    37.02997    27.72967    20.45250   11.947476   8.3717703 6.645227e+00
#>  [82,]    37.28583    31.99001    14.04898   12.971193  10.0512324 1.742011e+00
#>  [83,]    31.66663    19.37345    15.14909    8.375627   5.8023301 2.815851e+00
#>  [84,]    48.80713    23.33975    16.46091    7.715511   1.2295524 3.611663e-15
#>  [85,]    43.11511    27.59929    14.30330    9.392276   5.2475527 1.634627e+00
#>  [86,]    35.34063    23.03292    10.77683    9.937136   5.6900602 4.960555e+00
#>  [87,]    49.26211    25.79303    13.06471   10.721947   5.3162858 1.567656e+00
#>  [88,]    46.61196    26.41943    18.78729    7.873592   1.7822201 5.475878e-15
#>  [89,]    38.01428    24.14670    15.60929   14.031114   2.3932076 1.282269e+00
#>  [90,]    31.73707    19.49436    13.21223   10.368998   7.5762420 2.331879e-01
#>  [91,]    65.83874    28.45393    19.83302    9.958331   3.2157063 8.663934e-15
#>  [92,]    37.28864    26.12047    22.05509   11.166009   6.7138636 2.318485e+00
#>  [93,]    33.41939    24.76253    24.12157   11.025624   8.1050867 3.150052e+00
#>  [94,]    40.18274    22.54340    14.12425   10.515614   4.0816668 2.371037e-15
#>  [95,]    29.32007    21.45392    14.25234    9.996110   8.1819872 5.659663e+00
#>  [96,]    48.15136    26.87477    11.92334    9.516877   6.7479357 2.078718e+00
#>  [97,]    32.78058    22.55615    16.62259   15.296262  11.0507930 4.027307e+00
#>  [98,]    40.33265    21.78453    18.55316   16.100095   6.8614050 3.540277e+00
#>  [99,]    63.19312    14.32054    12.02269    6.605166   4.1820692 3.864806e-01
#> [100,]    35.49256    27.20478    17.73963   11.468932   9.3431218 4.142053e+00
#>          Dimension 7   Dimension 8   Dimension 9
#>   [1,]  1.814424e-01  3.835522e-15  1.527667e-15
#>   [2,]  5.291941e-15  2.972906e-15  9.379896e-17
#>   [3,]  1.391682e+00  4.320726e-15 -7.400625e-17
#>   [4,]  1.236790e-15  1.169381e-15 -3.715816e-16
#>   [5,] -4.161154e-16 -7.362699e-16 -2.229420e-15
#>   [6,]  2.725251e-16 -2.339772e-16 -5.929024e-16
#>   [7,]  1.001357e+00  1.823231e-01  1.998401e-15
#>   [8,]  6.852349e-15  3.312191e-15 -1.766994e-15
#>   [9,]  4.968056e-15  8.353135e-16  1.629694e-16
#>  [10,]  3.810111e-15  1.732536e-15  1.603667e-15
#>  [11,]  2.276261e-15  1.682726e-15  1.032428e-15
#>  [12,]  1.400673e+00  4.215387e-15 -4.079268e-16
#>  [13,]  2.480934e-01  2.892691e-15 -6.722451e-16
#>  [14,]  2.420065e+00  8.025775e-01  1.412612e-01
#>  [15,]  5.797023e-15  5.577019e-16  4.098545e-16
#>  [16,]  2.841782e+00  1.200587e+00  5.715960e-15
#>  [17,]  1.629488e+00  5.239569e-15 -5.016734e-16
#>  [18,]  4.096329e-15 -2.912723e-16 -4.003138e-15
#>  [19,]  1.147026e-15  9.422769e-16 -2.835055e-16
#>  [20,]  1.899200e+00  3.014248e-01  5.646803e-15
#>  [21,]  1.233026e+00  2.885166e-16 -5.623451e-16
#>  [22,]  3.292662e-15  3.208287e-15 -8.314549e-16
#>  [23,] -4.266053e-16 -1.666909e-15 -3.276928e-15
#>  [24,]  1.018031e+00  1.053656e-14  9.893777e-16
#>  [25,]  3.041538e-15  1.499540e-15  1.257602e-15
#>  [26,]  1.710580e-15  9.577674e-16 -1.571567e-16
#>  [27,]  1.487883e+00  1.806519e-15  1.267035e-15
#>  [28,] -2.913293e-16 -7.341727e-16 -3.112038e-15
#>  [29,]  2.190258e-15  2.066353e-15  1.318198e-15
#>  [30,]  1.181500e-14  3.433238e-15  1.239789e-15
#>  [31,]  3.038155e-16 -7.626367e-16 -2.125901e-15
#>  [32,]  1.057306e+00  4.855093e-15  1.625269e-15
#>  [33,]  3.631315e-15  2.111801e-15 -1.311447e-15
#>  [34,]  1.130966e-15  7.395583e-16 -3.096484e-16
#>  [35,]  2.754267e-15 -5.666061e-16 -1.169304e-15
#>  [36,]  4.888785e-01  3.649434e-15  1.984205e-15
#>  [37,]  8.296249e-16  4.418246e-16  6.417273e-17
#>  [38,]  3.006879e-15  2.454961e-15  7.577690e-16
#>  [39,]  1.635433e+00  5.969512e-15  3.187275e-15
#>  [40,]  1.381968e+00  3.845662e-15 -1.511186e-16
#>  [41,]  2.071880e-15  1.513243e-15  2.292480e-16
#>  [42,]  4.569526e-15 -3.216951e-17 -6.750160e-16
#>  [43,] -4.825593e-16 -7.303362e-16 -1.537721e-15
#>  [44,]  4.979503e-15  1.715178e-15 -2.697612e-15
#>  [45,] -7.212297e-17 -1.478829e-15 -1.569674e-15
#>  [46,]  3.412841e-15 -4.310547e-17 -2.102685e-15
#>  [47,]  2.149626e-15  9.599133e-16  7.758874e-17
#>  [48,]  1.008819e+00  1.733770e-01  2.081668e-15
#>  [49,]  1.570820e+00  1.984817e-01  3.334465e-15
#>  [50,]  1.502590e+00  3.213761e-16 -8.945372e-16
#>  [51,]  1.842437e-15  1.050001e-15 -3.212025e-17
#>  [52,]  3.571498e-01  1.157326e-15  5.265665e-16
#>  [53,] -1.647932e-16 -2.103398e-16 -2.847907e-15
#>  [54,]  5.275252e-15  1.117747e-15  2.122397e-16
#>  [55,]  6.375866e-15  5.299139e-15 -1.016731e-16
#>  [56,]  7.213392e-16 -4.515168e-16 -8.157853e-16
#>  [57,]  1.297453e+00  8.848561e-16 -2.742334e-16
#>  [58,]  1.900971e-15  1.683278e-15  1.039293e-15
#>  [59,]  1.512904e+00  5.893222e-01 -1.863472e-16
#>  [60,]  3.419293e-15  1.568108e-16 -1.799747e-15
#>  [61,]  2.684009e+00  1.484323e+00  2.497957e-01
#>  [62,]  4.710128e-15 -8.276892e-17 -2.194880e-15
#>  [63,] -1.159783e-19 -3.178303e-16 -8.793560e-16
#>  [64,]  7.393541e-15  1.033021e-16 -5.822362e-16
#>  [65,]  1.577370e-15  1.057346e-15 -1.109784e-16
#>  [66,] -1.165902e-16 -1.943030e-16 -3.764313e-15
#>  [67,]  2.794159e+00  8.874658e-16  2.418069e-16
#>  [68,]  3.840891e-15  2.298001e-15 -4.493166e-16
#>  [69,]  1.496709e+00  4.955716e-15  8.350267e-16
#>  [70,]  7.537906e-15  3.071830e-15 -4.968434e-16
#>  [71,]  2.370838e+00  1.319848e+00 -8.008638e-16
#>  [72,]  2.619806e-01  6.745664e-15  3.020906e-15
#>  [73,]  1.583594e-01  8.443475e-16 -2.772101e-15
#>  [74,]  2.514356e+00  2.648687e-15  7.341135e-16
#>  [75,]  1.253594e+00  4.664789e-15  7.145587e-16
#>  [76,]  1.521220e+00  3.216535e-01  5.852366e-15
#>  [77,]  1.051945e-15  1.828960e-16 -6.653032e-16
#>  [78,]  9.567120e-15 -4.077334e-16 -2.280809e-15
#>  [79,]  2.023920e+00  3.102245e-01  3.089818e-15
#>  [80,]  1.455896e-15  7.478869e-16  1.781149e-17
#>  [81,]  1.912330e+00  3.060779e-01  1.085913e-15
#>  [82,]  1.927054e-15  9.815472e-16  2.183167e-16
#>  [83,]  5.984376e-01  8.448072e-17 -5.730222e-16
#>  [84,]  8.808350e-16  2.091110e-16 -3.439690e-16
#>  [85,]  4.631254e-15  1.396882e-15  6.976360e-16
#>  [86,]  5.275675e-15  5.101371e-15  5.774892e-16
#>  [87,]  1.270673e-15 -4.310282e-16 -2.307471e-15
#>  [88,]  5.965376e-16 -7.747369e-16 -8.011579e-16
#>  [89,]  1.311557e-15 -1.135269e-16 -4.001495e-16
#>  [90,]  1.974210e-15  1.700978e-15 -2.890083e-16
#>  [91,]  5.104518e-15 -4.525472e-16 -1.696405e-15
#>  [92,]  1.325499e+00  5.715783e-16 -1.285268e-15
#>  [93,]  9.222153e-01  6.044428e-16  5.952385e-16
#>  [94,]  2.139476e-15 -4.007766e-17 -2.839024e-15
#>  [95,]  1.530659e+00  1.353334e-15  5.141423e-17
#>  [96,]  2.806653e-15  1.387254e-16 -1.280567e-15
#>  [97,]  9.812966e-01  5.231565e-15  1.120631e-15
#>  [98,]  1.365591e+00  1.089612e+00  5.886035e-15
#>  [99,]  3.081707e-15 -5.721408e-16 -8.897952e-16
#> [100,]  1.026800e+00  2.547704e-15 -3.150495e-16
#> 
#> $BootMatEigsCI
#>                     LowCI       HighCI
#> Dimension 1  2.868819e+01 5.591548e+01
#> Dimension 2  1.586041e+01 3.429879e+01
#> Dimension 3  1.077683e+01 2.412157e+01
#> Dimension 4  5.931515e+00 1.769575e+01
#> Dimension 5  1.229552e+00 1.105079e+01
#> Dimension 6  4.945414e-16 6.645227e+00
#> Dimension 7 -4.266053e-16 2.684009e+00
#> Dimension 8 -1.478829e-15 1.200587e+00
#> Dimension 9 -3.764313e-15 5.886035e-15
#> 
#> $FixedEigs
#> [1] 26.0248696 21.9734833 17.4744535 16.5652771 10.6763967  7.3267567  3.1769855
#> [8]  1.8085971  0.5806198
#> 
#> $BootstrappedEstEigs
#>  Dimension 1  Dimension 2  Dimension 3  Dimension 4  Dimension 5  Dimension 6 
#> 41.598955860 26.122754615 17.436865947 11.100052230  6.280103648  2.450399770 
#>  Dimension 7  Dimension 8  Dimension 9 
#>  0.553080182  0.082798318  0.003910568 
#> 
#> $BootMatTaus
#>        Dimension 1 Dimension 2 Dimension 3 Dimension 4 Dimension 5  Dimension 6
#>   [1,]   0.4574076   0.2153518   0.1279715  0.09021132 0.065495026 4.187926e-02
#>   [2,]   0.3082955   0.2490185   0.2030211  0.13256287 0.085374017 2.172804e-02
#>   [3,]   0.4048009   0.2637045   0.1482644  0.09102727 0.064743435 1.652686e-02
#>   [4,]   0.4011077   0.2848453   0.1563232  0.08977876 0.046803346 2.114169e-02
#>   [5,]   0.4224770   0.2831952   0.1854142  0.09212642 0.016787183 1.195243e-16
#>   [6,]   0.4129772   0.2919788   0.1852114  0.07140629 0.038426342 1.415323e-17
#>   [7,]   0.3202846   0.2716293   0.1909601  0.09527993 0.076691235 3.461019e-02
#>   [8,]   0.4396468   0.2382604   0.1395885  0.10367900 0.060355839 1.846947e-02
#>   [9,]   0.4087467   0.3012134   0.1164162  0.08542985 0.058398239 2.979562e-02
#>  [10,]   0.4768104   0.2588868   0.1835257  0.05338377 0.027393373 3.533642e-17
#>  [11,]   0.4432745   0.2613603   0.1539494  0.06970714 0.055679972 1.602865e-02
#>  [12,]   0.4798523   0.1679510   0.1262741  0.09275513 0.070626324 4.847627e-02
#>  [13,]   0.4329245   0.1771361   0.1321227  0.12536392 0.080783005 4.904348e-02
#>  [14,]   0.3305360   0.2189497   0.1625823  0.10519603 0.088841201 6.366360e-02
#>  [15,]   0.4452256   0.2573714   0.1682783  0.09087903 0.025766429 1.247927e-02
#>  [16,]   0.3104865   0.2081335   0.1518328  0.13289565 0.125188095 3.368420e-02
#>  [17,]   0.3399888   0.2375184   0.2087969  0.08462578 0.079052141 3.624176e-02
#>  [18,]   0.3737511   0.1966591   0.1856699  0.11241112 0.079544152 5.196460e-02
#>  [19,]   0.4049595   0.2030443   0.1851092  0.11554841 0.091338611 3.554036e-17
#>  [20,]   0.2519251   0.2188429   0.1811644  0.16998856 0.108685230 4.899951e-02
#>  [21,]   0.4354108   0.2617601   0.1404310  0.09082374 0.040345307 2.097747e-02
#>  [22,]   0.3503846   0.2807764   0.1823976  0.10203376 0.067658345 1.674921e-02
#>  [23,]   0.4072152   0.3096642   0.1376391  0.08900401 0.056477444 4.052380e-17
#>  [24,]   0.4269296   0.2621767   0.1004389  0.08604341 0.074157049 4.198320e-02
#>  [25,]   0.3059693   0.2618157   0.1834064  0.16374618 0.068524541 1.653787e-02
#>  [26,]   0.3892363   0.3045178   0.1811863  0.07862127 0.043201468 3.236932e-03
#>  [27,]   0.3263429   0.2488840   0.2069122  0.12947071 0.043421261 3.199315e-02
#>  [28,]   0.4738668   0.2755305   0.1346645  0.08014048 0.035797667 5.144780e-18
#>  [29,]   0.3632728   0.2313153   0.1822443  0.13229800 0.090869634 1.154779e-16
#>  [30,]   0.3430683   0.3159089   0.1635294  0.10697253 0.053153623 1.736725e-02
#>  [31,]   0.3590620   0.2020401   0.1902916  0.13944292 0.083554183 2.560917e-02
#>  [32,]   0.3949175   0.2405412   0.1249186  0.10796456 0.075283401 4.674261e-02
#>  [33,]   0.3272639   0.2554154   0.2110316  0.15172504 0.038798997 1.576502e-02
#>  [34,]   0.4246798   0.2947301   0.1685439  0.05944288 0.039606324 1.299699e-02
#>  [35,]   0.4298595   0.3072016   0.1135045  0.08836248 0.040650565 2.042129e-02
#>  [36,]   0.3715904   0.2366294   0.1579076  0.13417492 0.068751843 2.605499e-02
#>  [37,]   0.4919281   0.2851872   0.1201052  0.07020015 0.032579369 1.279379e-17
#>  [38,]   0.4789186   0.2289504   0.1150772  0.09650326 0.057767741 2.278286e-02
#>  [39,]   0.3408265   0.2583331   0.1577660  0.12621465 0.080734015 2.151489e-02
#>  [40,]   0.3637927   0.2207227   0.1867393  0.14893649 0.043396843 2.304343e-02
#>  [41,]   0.3582106   0.3062964   0.1734809  0.09729575 0.043664652 2.105163e-02
#>  [42,]   0.3591666   0.2935449   0.1512087  0.12291263 0.053150070 2.001713e-02
#>  [43,]   0.5519961   0.2144351   0.1292828  0.06734777 0.033676268 3.261930e-03
#>  [44,]   0.3664414   0.2748842   0.1656742  0.11723598 0.048766973 2.699722e-02
#>  [45,]   0.4117980   0.2440898   0.1950465  0.08602704 0.048529116 1.450946e-02
#>  [46,]   0.3490054   0.2974605   0.1936528  0.07927951 0.059916110 2.068566e-02
#>  [47,]   0.4449438   0.2421077   0.1534793  0.09291663 0.052334214 1.421838e-02
#>  [48,]   0.3450896   0.2300702   0.1700703  0.10741798 0.090016398 4.592612e-02
#>  [49,]   0.3273436   0.2175405   0.1924401  0.12682139 0.077775698 4.031243e-02
#>  [50,]   0.3150116   0.2372216   0.2031673  0.13361729 0.064068990 3.233511e-02
#>  [51,]   0.4451983   0.2522721   0.1429984  0.10801591 0.051515369 3.350425e-17
#>  [52,]   0.3508580   0.2882538   0.2064541  0.09189730 0.034535353 2.425453e-02
#>  [53,]   0.4355231   0.3530175   0.1298682  0.06255940 0.019031754 2.627545e-17
#>  [54,]   0.4471430   0.2264433   0.1770834  0.07705559 0.051454909 2.081984e-02
#>  [55,]   0.4790008   0.2048218   0.1759461  0.09164604 0.029976048 1.860929e-02
#>  [56,]   0.4545632   0.3481262   0.1346539  0.05920950 0.003447231 1.460781e-17
#>  [57,]   0.3792013   0.2485666   0.2002056  0.08354985 0.048897575 2.644906e-02
#>  [58,]   0.3738406   0.2615699   0.1700793  0.09624792 0.061659769 3.660248e-02
#>  [59,]   0.4233885   0.1786653   0.1520714  0.09858062 0.092313146 3.642656e-02
#>  [60,]   0.4035649   0.2351404   0.1738583  0.13232718 0.043097264 1.201191e-02
#>  [61,]   0.3101938   0.2142641   0.1539647  0.13508814 0.088109764 6.078361e-02
#>  [62,]   0.3544971   0.2983612   0.2034132  0.08985405 0.027272781 2.660171e-02
#>  [63,]   0.3887873   0.3287632   0.1732017  0.06940385 0.039843974 4.328407e-19
#>  [64,]   0.5514053   0.1564060   0.1376322  0.10179128 0.046096730 6.668466e-03
#>  [65,]   0.4572708   0.1885159   0.1471771  0.10053991 0.089741702 1.675451e-02
#>  [66,]   0.3897960   0.3309889   0.1614514  0.06779332 0.049970303 2.648494e-17
#>  [67,]   0.3096662   0.2346578   0.1883409  0.11846329 0.088223180 3.263969e-02
#>  [68,]   0.4498120   0.2756440   0.1401416  0.09349514 0.040907248 6.251515e-17
#>  [69,]   0.3897078   0.2156699   0.1683632  0.09709220 0.082200862 3.308325e-02
#>  [70,]   0.4100999   0.2654922   0.1526651  0.10680924 0.046829501 1.810406e-02
#>  [71,]   0.2984349   0.2046456   0.1761526  0.15143903 0.080620392 5.503696e-02
#>  [72,]   0.5235810   0.1711880   0.1247533  0.11643080 0.041958669 1.945142e-02
#>  [73,]   0.3495064   0.2396689   0.1863320  0.12487823 0.063344255 3.480924e-02
#>  [74,]   0.3511356   0.1813157   0.1703890  0.11401601 0.096801735 6.215905e-02
#>  [75,]   0.3441933   0.2515271   0.2044180  0.10067029 0.064717209 2.276280e-02
#>  [76,]   0.3494620   0.2060292   0.1781389  0.13643953 0.072514665 3.934915e-02
#>  [77,]   0.3722099   0.2266871   0.1763385  0.14867927 0.050828282 2.525691e-02
#>  [78,]   0.2977581   0.2390565   0.2241284  0.16155500 0.039818215 3.768381e-02
#>  [79,]   0.3923587   0.2025715   0.1878034  0.11409785 0.058852265 2.201430e-02
#>  [80,]   0.3193638   0.2935761   0.2242225  0.09927266 0.056754138 6.810700e-03
#>  [81,]   0.3237026   0.2424028   0.1787884  0.10444053 0.073183002 5.809018e-02
#>  [82,]   0.3449541   0.2959592   0.1299757  0.12000447 0.092990116 1.611641e-02
#>  [83,]   0.3779672   0.2312380   0.1808169  0.09996998 0.069255569 3.360949e-02
#>  [84,]   0.5003148   0.2392523   0.1687384  0.07909057 0.012603961 3.702262e-17
#>  [85,]   0.4256510   0.2724722   0.1412084  0.09272462 0.051806114 1.613774e-02
#>  [86,]   0.3938195   0.2566681   0.1200920  0.11073483 0.063407388 5.527812e-02
#>  [87,]   0.4659424   0.2439617   0.1235717  0.10141283 0.050283742 1.482757e-02
#>  [88,]   0.4593466   0.2603554   0.1851429  0.07759183 0.017563232 5.396309e-17
#>  [89,]   0.3981518   0.2529063   0.1634877  0.14695826 0.025065839 1.343015e-02
#>  [90,]   0.3841233   0.2359461   0.1599116  0.12549913 0.091697549 2.822344e-03
#>  [91,]   0.5171946   0.2235192   0.1557979  0.07822744 0.025260905 6.805933e-17
#>  [92,]   0.3485308   0.2441438   0.2061453  0.10436687 0.062753395 2.167050e-02
#>  [93,]   0.3167520   0.2347016   0.2286264  0.10450188 0.076820757 2.985648e-02
#>  [94,]   0.4394069   0.2465169   0.1544517  0.11499050 0.044633903 2.592779e-17
#>  [95,]   0.3243559   0.2373359   0.1576678  0.11058286 0.090513959 6.261053e-02
#>  [96,]   0.4573083   0.2552379   0.1132396  0.09038471 0.064087221 1.974222e-02
#>  [97,]   0.3172877   0.2183241   0.1608923  0.14805465 0.106962159 3.898086e-02
#>  [98,]   0.3679069   0.1987144   0.1692385  0.14686208 0.062588462 3.229375e-02
#>  [99,]   0.6274757   0.1421957   0.1193792  0.06558596 0.041525834 3.837557e-03
#> [100,]   0.3335207   0.2556411   0.1666978  0.10777262 0.087796550 3.892254e-02
#>          Dimension 7   Dimension 8   Dimension 9
#>   [1,]  1.683468e-03  3.558693e-17  1.417407e-17
#>   [2,]  4.684279e-17  2.631533e-17  8.302823e-19
#>   [3,]  1.093266e-02  3.394241e-17 -5.813723e-19
#>   [4,]  1.039032e-17  9.824013e-18 -3.121670e-18
#>   [5,] -3.435744e-18 -6.079168e-18 -1.840768e-17
#>   [6,]  2.978664e-18 -2.557340e-18 -6.480346e-18
#>   [7,]  8.920456e-03  1.624202e-03  1.780250e-17
#>   [8,]  7.304104e-17  3.530554e-17 -1.883486e-17
#>   [9,]  3.858067e-17  6.486834e-18  1.265580e-18
#>  [10,]  3.429108e-17  1.559286e-17  1.443304e-17
#>  [11,]  2.326494e-17  1.719860e-17  1.055212e-17
#>  [12,]  1.406483e-02  4.232871e-17 -4.096188e-18
#>  [13,]  2.626334e-03  3.062224e-17 -7.116435e-18
#>  [14,]  2.174896e-02  7.212710e-03  1.269505e-03
#>  [15,]  4.964052e-17  4.775660e-18  3.509628e-18
#>  [16,]  2.655877e-02  1.122047e-02  5.342030e-17
#>  [17,]  1.377614e-02  4.429678e-17 -4.241287e-18
#>  [18,]  3.476008e-17 -2.471639e-18 -3.396929e-17
#>  [19,]  9.849860e-18  8.091617e-18 -2.434548e-18
#>  [20,]  1.760077e-02  2.793444e-03  5.233155e-17
#>  [21,]  1.025149e-02  2.398752e-18 -4.675385e-18
#>  [22,]  3.246056e-17  3.162874e-17 -8.196859e-18
#>  [23,] -3.857703e-18 -1.507351e-17 -2.963258e-17
#>  [24,]  8.271174e-03  8.560613e-17  8.038376e-18
#>  [25,]  2.814462e-17  1.387588e-17  1.163712e-17
#>  [26,]  1.702758e-17  9.533873e-18 -1.564380e-18
#>  [27,]  1.297577e-02  1.575458e-17  1.104977e-17
#>  [28,] -3.030738e-18 -7.637697e-18 -3.237495e-17
#>  [29,]  1.821487e-17  1.718443e-17  1.096254e-17
#>  [30,]  9.868158e-17  2.867518e-17  1.035500e-17
#>  [31,]  3.098736e-18 -7.778437e-18 -2.168292e-17
#>  [32,]  9.632124e-03  4.423021e-17  1.480631e-17
#>  [33,]  3.758331e-17  2.185667e-17 -1.357319e-17
#>  [34,]  1.037003e-17  6.781144e-18 -2.839222e-18
#>  [35,]  2.976308e-17 -6.122842e-18 -1.263570e-17
#>  [36,]  4.890838e-03  3.650967e-17  1.985039e-17
#>  [37,]  8.967069e-18  4.775498e-18  6.936163e-19
#>  [38,]  2.816704e-17  2.299693e-17  7.098427e-18
#>  [39,]  1.461094e-02  5.333154e-17  2.847507e-17
#>  [40,]  1.336860e-02  3.720137e-17 -1.461860e-18
#>  [41,]  1.894555e-17  1.383730e-17  2.096274e-18
#>  [42,]  4.387891e-17 -3.089080e-19 -6.481846e-18
#>  [43,] -4.843540e-18 -7.330525e-18 -1.543440e-17
#>  [44,]  4.535014e-17  1.562074e-17 -2.456813e-17
#>  [45,] -6.452503e-19 -1.323039e-17 -1.404313e-17
#>  [46,]  4.028220e-17 -5.087795e-19 -2.481826e-17
#>  [47,]  1.766552e-17  7.888522e-18  6.376206e-19
#>  [48,]  9.736189e-03  1.673274e-03  2.009033e-17
#>  [49,]  1.577327e-02  1.993038e-03  3.348277e-17
#>  [50,]  1.457812e-02  3.117988e-18 -8.678790e-18
#>  [51,]  1.861602e-17  1.060923e-17 -3.245436e-19
#>  [52,]  3.746836e-03  1.214143e-17  5.524177e-18
#>  [53,] -1.712042e-18 -2.185228e-18 -2.958701e-17
#>  [54,]  4.550650e-17  9.642146e-18  1.830867e-18
#>  [55,]  6.262696e-17  5.205080e-17 -9.986839e-19
#>  [56,]  7.487768e-18 -4.686912e-18 -8.468154e-18
#>  [57,]  1.313000e-02  8.954593e-18 -2.775195e-18
#>  [58,]  1.810490e-17  1.603158e-17  9.898249e-18
#>  [59,]  1.335308e-02  5.201432e-03 -1.644723e-18
#>  [60,]  3.155893e-17  1.447311e-18 -1.661106e-17
#>  [61,]  2.283949e-02  1.263079e-02  2.125628e-03
#>  [62,]  4.516702e-17 -7.936993e-19 -2.104745e-17
#>  [63,] -1.204093e-21 -3.299731e-18 -9.129522e-18
#>  [64,]  7.291071e-17  1.018705e-18 -5.741669e-18
#>  [65,]  1.451227e-17  9.727901e-18 -1.021034e-18
#>  [66,] -1.179713e-18 -1.966047e-18 -3.808905e-17
#>  [67,]  2.800897e-02  8.896058e-18  2.423900e-18
#>  [68,]  3.509466e-17  2.099709e-17 -4.105457e-18
#>  [69,]  1.388283e-02  4.596711e-17  7.745351e-18
#>  [70,]  6.577998e-17  2.680650e-17 -4.335734e-18
#>  [71,]  2.162940e-02  1.204111e-02 -7.306365e-18
#>  [72,]  2.636758e-03  6.789312e-17  3.040454e-17
#>  [73,]  1.460960e-03  7.789611e-18 -2.557429e-17
#>  [74,]  2.418291e-02  2.547490e-17  7.060656e-18
#>  [75,]  1.171130e-02  4.357930e-17  6.675536e-18
#>  [76,]  1.491322e-02  3.153317e-03  5.737344e-17
#>  [77,]  9.247011e-18  1.607728e-18 -5.848276e-18
#>  [78,]  8.489582e-17 -3.618106e-18 -2.023923e-17
#>  [79,]  1.933786e-02  2.964089e-03  2.952216e-17
#>  [80,]  1.620739e-17  8.325657e-18  1.982818e-19
#>  [81,]  1.671690e-02  2.675623e-03  9.492661e-18
#>  [82,]  1.782836e-17  9.080896e-18  2.019782e-18
#>  [83,]  7.142843e-03  1.008347e-18 -6.839490e-18
#>  [84,]  9.029310e-18  2.143566e-18 -3.525976e-18
#>  [85,]  4.572175e-17  1.379062e-17  6.887365e-18
#>  [86,]  5.878967e-17  5.684731e-17  6.435271e-18
#>  [87,]  1.201858e-17 -4.076852e-18 -2.182506e-17
#>  [88,]  5.878695e-18 -7.634794e-18 -7.895165e-18
#>  [89,]  1.373691e-17 -1.189052e-18 -4.191062e-18
#>  [90,]  2.389447e-17  2.058745e-17 -3.497955e-18
#>  [91,]  4.009842e-17 -3.554974e-18 -1.332607e-17
#>  [92,]  1.238923e-02  5.342450e-18 -1.201319e-17
#>  [93,]  8.740841e-03  5.728965e-18  5.641725e-18
#>  [94,]  2.339563e-17 -4.382578e-19 -3.104534e-17
#>  [95,]  1.693305e-02  1.497138e-17  5.687745e-19
#>  [96,]  2.665564e-17  1.317517e-18 -1.216194e-17
#>  [97,]  9.498106e-03  5.063704e-17  1.084675e-17
#>  [98,]  1.245667e-02  9.939241e-03  5.369132e-17
#>  [99,]  3.059979e-17 -5.681069e-18 -8.835217e-18
#> [100,]  9.648754e-03  2.394056e-17 -2.960494e-18
#> 
#> $BootMatTausCI
#>                     LowCI       HighCI
#> Dimension 1  2.977581e-01 5.514053e-01
#> Dimension 2  1.564060e-01 3.309889e-01
#> Dimension 3  1.132396e-01 2.241284e-01
#> Dimension 4  5.920950e-02 1.615550e-01
#> Dimension 5  1.260396e-02 1.069622e-01
#> Dimension 6  5.144780e-18 6.215905e-02
#> Dimension 7 -3.857703e-18 2.418291e-02
#> Dimension 8 -1.323039e-17 1.122047e-02
#> Dimension 9 -3.396929e-17 5.737344e-17
#> 
#> $BootstrappedEstInertia
#>   [1] 107.77897 112.97238 127.29578 119.03293 121.11360  91.49240 112.25396
#>   [8]  93.81506 128.77060 111.11084  97.84083  99.58695  94.46375 111.27267
#>  [15] 116.78006 106.99976 118.28330 117.84579 116.45100 107.90436 120.27780
#>  [22] 101.43579 110.58531 123.08179 108.06816 100.45942 114.66621  96.12487
#>  [29] 120.24565 119.72857  98.04498 109.76870  96.62041 109.06099  92.53971
#>  [36]  99.95802  92.51907 106.75169 111.93210 103.37420 109.35973 104.13946
#>  [43]  99.62946 109.80127 111.77517  84.72329 121.68481 103.61542  99.58749
#>  [50] 103.07164  98.97054  95.32036  96.25533 115.92305 101.80706  96.33567
#>  [57]  98.81589 104.99762 113.30000 108.34628 117.51619 104.28246  96.32005
#>  [64] 101.40541 108.69214  98.82927  99.75945 109.44375 107.81006 114.59270
#>  [71] 109.61180  99.35710 108.39405 103.97243 107.04140 102.00479 113.76057
#>  [78] 112.69248 104.66099  89.82917 114.39501 108.08926  83.78142  97.55286
#>  [85] 101.29215  89.73813 105.72574 101.47450  95.47686  82.62208 127.29972
#>  [92] 106.98805 105.50647  91.44768  90.39476 105.29300 103.31498 109.62731
#>  [99] 100.71006 106.41787
#>