Boostrap for eigenvalues.

Boostrap for eigenvalues.

Usage

Boot4Eigs(
  data,
  center = FALSE,
  scale = FALSE,
  design = NULL,
  niter = 100,
  CI.perc = c(0.025, 0.975),
  suppressProgressBar = TRUE
)

Arguments

data

the data-set;

center

whether to center the data (default to FALSE), see expo.scale;

scale

whether to scale the data ( default to FALSE), see expo.scale;

design

whether there is a design;

niter

number of bootstrap iterations;

CI.perc

vector of the two levels of confidence for the left and right limits of the bootstrap confidence interval (default to c(0.025, 0.0975));

suppressProgressBar

whether to show a progress bar or not (default to FALSE).

Value

list containing four objects

Examples

Boot4Eigs(matrix(rnorm(100), 10, 10))
#> $BootMatEigs
#>        Dimension 1 Dimension 2 Dimension 3 Dimension 4  Dimension 5
#>   [1,]    44.06723    21.74548   14.505565    6.563274 3.282080e+00
#>   [2,]    50.09345    32.95664   20.560621   12.034252 7.934937e+00
#>   [3,]    51.09181    41.30430   17.155428    5.468851 2.942751e+00
#>   [4,]    63.46305    25.68530   11.549240    9.403404 6.211268e+00
#>   [5,]    51.41958    18.00105   13.111365    8.438893 5.913104e+00
#>   [6,]    53.12108    33.30455   18.230051    8.941719 6.477069e+00
#>   [7,]    58.94157    39.37396   12.850130   11.521983 3.948541e+00
#>   [8,]    48.26425    31.25713   17.681865   11.414807 7.006330e+00
#>   [9,]    41.99725    22.10218   16.406027    7.250552 2.999821e+00
#>  [10,]    59.65131    21.76969    7.520578    5.381126 3.405619e+00
#>  [11,]    59.19076    43.08749   19.984736    6.885926 1.133979e-15
#>  [12,]    60.77979    22.49230   15.974266    6.742622 5.872523e+00
#>  [13,]    39.39320    23.71807   13.135799    7.928703 1.781376e+00
#>  [14,]    90.53068    33.60664   16.120300    6.846382 1.454289e+00
#>  [15,]    45.03592    26.97654   17.441886    6.884917 5.195608e+00
#>  [16,]    46.34322    29.70258   18.772492    7.021695 4.890344e+00
#>  [17,]    79.19331    20.86746   12.733834    6.688316 6.191143e+00
#>  [18,]    44.92963    19.42021   15.942744    7.588579 5.163833e+00
#>  [19,]    40.75428    35.19743   15.350203   11.883508 5.606357e+00
#>  [20,]    37.29412    25.55567   11.004805    7.044197 6.400342e+00
#>  [21,]    59.23855    34.50331   15.199252   14.668647 6.989058e+00
#>  [22,]    47.43990    33.59715   12.443174    7.700318 5.216206e+00
#>  [23,]    37.60123    29.35145   24.986104   10.003657 3.669039e+00
#>  [24,]    77.52656    40.45179   21.864274    8.060445 5.912087e+00
#>  [25,]    51.97033    18.67906   11.226655    5.281317 3.542419e+00
#>  [26,]    62.81220    42.41430   16.407082    7.440707 6.838362e+00
#>  [27,]    66.86587    38.39911   21.772441   12.024260 3.930744e+00
#>  [28,]    34.40479    25.74122   24.380162    9.905075 7.928436e+00
#>  [29,]    47.83008    27.65837   19.172358   10.524612 4.735209e+00
#>  [30,]    64.46847    36.95998   11.747714    9.755610 3.817109e+00
#>  [31,]    36.91553    30.50071   20.222197   13.376634 5.737754e+00
#>  [32,]    53.45021    37.21415   11.684421    7.877392 6.239035e+00
#>  [33,]    51.04470    29.24174   18.899246   10.336146 7.439953e+00
#>  [34,]    40.07412    34.05728   18.193613    7.276232 6.908959e+00
#>  [35,]    74.41990    39.59036   15.878738    6.844697 3.475484e+00
#>  [36,]    42.37935    23.71085   15.049255    8.672078 5.683361e+00
#>  [37,]    51.50636    44.94769   13.047821    8.200064 4.818563e+00
#>  [38,]    50.37540    29.18519   13.050007    7.503507 6.613022e+00
#>  [39,]    67.54829    33.67906   15.823145    9.664143 5.864871e+00
#>  [40,]    65.01171    30.75330   18.735931    8.600922 6.111604e+00
#>  [41,]    64.48809    28.20912   21.394612   11.067122 6.127878e+00
#>  [42,]    55.59601    28.22188   17.330186   14.807714 6.725820e+00
#>  [43,]    48.76169    34.48328   14.397597   10.879271 5.770139e+00
#>  [44,]    51.99048    23.07716   14.252517    6.794083 3.494806e+00
#>  [45,]    60.88628    25.83496   15.062220   11.715980 3.268422e+00
#>  [46,]    36.92374    28.11489   23.664659    8.253634 6.994400e+00
#>  [47,]    49.67808    27.62990   15.946644    8.734776 7.635784e+00
#>  [48,]    60.70312    38.60374   12.320845    9.425468 5.997299e+00
#>  [49,]   102.76624    26.29175   10.780426    4.714596 1.930059e+00
#>  [50,]    61.69324    35.53690   21.307682    9.283581 5.124234e+00
#>  [51,]    41.32492    19.91188   17.304746    9.885038 5.183459e+00
#>  [52,]    37.57426    34.78161   15.857138    8.181557 6.897313e+00
#>  [53,]    42.29157    24.71708   20.279868    3.798904 3.228969e+00
#>  [54,]    49.65591    33.73487   27.843792   10.124855 5.864373e+00
#>  [55,]    46.54579    45.30285   22.592264   13.670392 6.239961e+00
#>  [56,]    46.67039    28.20011   15.852004   10.227341 7.626501e+00
#>  [57,]    45.33637    22.26562   12.361301    7.134946 3.609918e+00
#>  [58,]    62.79787    22.75243   16.377628    9.198768 4.775428e+00
#>  [59,]    60.61855    24.87956   15.971802   13.491879 3.129803e+00
#>  [60,]    62.67745    19.47249   12.777672    8.432576 4.776752e+00
#>  [61,]    67.91891    37.78594   19.928653    9.348536 2.826986e+00
#>  [62,]    42.80416    15.86013   14.563742    6.611528 4.183074e+00
#>  [63,]    45.63548    31.30453   26.647166    8.439260 2.885439e+00
#>  [64,]    41.46905    33.64644   25.242832    9.363978 6.723718e+00
#>  [65,]    40.99755    33.72527   14.106445    9.678776 3.992430e+00
#>  [66,]    80.94953    29.19504   13.278147    9.266426 6.939863e+00
#>  [67,]    43.59695    27.12247   16.199893    9.832099 6.629503e+00
#>  [68,]    48.40548    35.69926   21.988030    9.123378 2.198963e+00
#>  [69,]    53.68913    27.75963   20.553464    9.822147 4.077000e+00
#>  [70,]    44.74918    26.78479   16.956398   10.525012 5.331795e+00
#>  [71,]    55.73259    21.09627    9.711866    7.119744 2.222278e+00
#>  [72,]    64.20791    35.74012   23.353238    8.452278 6.618777e+00
#>  [73,]    48.12712    24.67449   23.455317   16.350396 6.622793e+00
#>  [74,]    42.45127    25.26465   14.997419    6.609774 3.898013e+00
#>  [75,]    66.20097    39.65966   18.720992    7.670116 5.654948e+00
#>  [76,]    39.71478    33.10781   13.134276    7.312451 5.145098e+00
#>  [77,]    68.86863    43.77731   16.505094    9.442970 6.241251e+00
#>  [78,]    46.60367    21.72854    9.017241    7.580489 6.696136e+00
#>  [79,]    59.91980    26.09652   20.758557    4.803890 1.921222e+00
#>  [80,]    66.81665    36.18353   16.753997   10.063129 6.811972e+00
#>  [81,]    65.81273    43.73137   14.099744   11.168534 6.672444e+00
#>  [82,]    55.47877    19.69355   14.558980    7.904881 6.171799e+00
#>  [83,]    50.30469    29.13297   18.156591    8.817128 6.088590e+00
#>  [84,]    60.91668    22.46369   14.393708   11.209892 6.317840e+00
#>  [85,]    46.21429    23.76573   17.028206   11.564265 2.792791e+00
#>  [86,]    47.58823    42.79370   18.228996    3.053400 2.133226e+00
#>  [87,]    48.74750    37.01652   14.908827    9.233451 7.755588e+00
#>  [88,]    46.88745    25.67994   14.166775    6.654530 3.870415e+00
#>  [89,]    42.48777    22.72116   16.083462    7.887422 6.207127e+00
#>  [90,]    49.74264    20.92399   15.278168    8.431150 5.188201e+00
#>  [91,]    51.67514    27.92989   22.795366   12.594259 7.111377e+00
#>  [92,]    41.64535    26.09969   13.282548    7.867677 3.721895e+00
#>  [93,]    60.35109    34.81320    6.907691    5.474237 4.321511e+00
#>  [94,]    46.15504    23.09523   12.392770   11.222064 4.832409e+00
#>  [95,]    59.73298    28.47427   16.536975    6.527417 5.578744e+00
#>  [96,]    45.81957    32.81665   18.018791   13.438090 4.828074e+00
#>  [97,]    51.68502    19.29745   14.696756   10.710195 4.281757e+00
#>  [98,]    36.65611    28.69208   24.156144    6.765556 4.559525e+00
#>  [99,]    50.69967    29.12688   19.757712   12.051429 7.676806e+00
#> [100,]    44.85364    24.36331   15.061050    4.640081 2.077749e+00
#>          Dimension 6   Dimension 7   Dimension 8   Dimension 9
#>   [1,]  1.752121e-15 -9.638639e-16 -1.878597e-15 -2.305233e-15
#>   [2,]  3.637142e+00  2.368306e+00  1.508826e+00  3.894129e-15
#>   [3,]  5.794734e-15  1.567103e-15  1.032538e-15 -1.318654e-15
#>   [4,]  2.592487e+00  1.276904e+00  2.043345e-15  1.093159e-15
#>   [5,]  9.399289e-01  1.579124e-15  1.149962e-15 -3.082971e-17
#>   [6,]  4.213715e+00  1.360583e+00  2.275227e-15  1.255289e-15
#>   [7,]  3.001934e+00  8.493820e-01  9.933071e-16  8.957799e-16
#>   [8,]  4.554173e+00  3.258473e+00  1.330683e+00  1.479904e-15
#>   [9,]  1.320075e+00  1.678966e-15  5.486271e-16 -1.100684e-15
#>  [10,]  2.084624e+00  5.208782e-15  6.438918e-16  5.510910e-16
#>  [11,] -6.939134e-16 -8.951739e-16 -2.596320e-15 -5.270411e-15
#>  [12,]  1.680109e+00  1.466077e-15  1.338563e-15 -1.039706e-15
#>  [13,]  4.749135e-15  4.817705e-16  2.787708e-16  1.934811e-16
#>  [14,]  5.328873e-15  2.592581e-15  1.997022e-15 -2.746394e-16
#>  [15,]  3.031763e+00  1.384781e+00  2.993500e-15  2.973597e-16
#>  [16,]  2.335865e+00  5.394603e-15 -2.962471e-16 -1.818123e-15
#>  [17,]  2.859385e+00  5.096817e-15  1.304056e-15 -2.408521e-15
#>  [18,]  2.859497e+00  2.041634e+00  2.692121e-15  1.556835e-16
#>  [19,]  3.741877e+00  2.262511e+00  9.883475e-01  3.270939e-15
#>  [20,]  4.143606e+00  1.470669e+00  1.021689e-15 -2.399886e-15
#>  [21,]  3.345895e+00  2.235220e+00  7.182711e-01 -6.661338e-16
#>  [22,]  4.299492e+00  1.402240e+00  1.091075e+00  1.650078e-15
#>  [23,]  2.753085e+00  1.462314e+00  2.002969e-15  2.505173e-16
#>  [24,]  1.948823e+00  8.521679e-15 -3.228442e-17 -3.323142e-15
#>  [25,]  8.505313e-01  2.357840e-16 -1.572560e-16 -4.732252e-16
#>  [26,]  7.713424e-01  7.529116e-15  4.868939e-15 -1.408224e-15
#>  [27,]  6.490256e-01  1.554913e-14  6.874809e-15 -1.529996e-15
#>  [28,]  3.211621e+00  1.285986e+00 -1.415574e-15 -2.559282e-15
#>  [29,]  2.750863e+00  1.768781e-15  1.162638e-15  5.255464e-16
#>  [30,]  2.393716e+00  9.776550e-01 -8.712316e-16 -2.731326e-15
#>  [31,]  3.975309e+00  2.874210e+00  5.678418e-01  6.710803e-15
#>  [32,]  1.482409e+00  3.730311e-15  9.946248e-16 -2.508214e-16
#>  [33,]  2.164268e+00  4.690215e-01  1.136121e-16 -2.246344e-16
#>  [34,]  5.879206e+00  3.197313e+00  6.200930e-01  5.863235e-16
#>  [35,]  2.328940e+00  3.270142e-16  1.675300e-16 -3.680535e-15
#>  [36,]  2.994988e+00  2.147126e-15  8.315101e-16 -1.100902e-15
#>  [37,]  1.916058e+00  1.160190e+00  5.155860e-15  9.279220e-16
#>  [38,]  7.185921e-01  4.771146e-15 -7.023662e-16 -1.118219e-15
#>  [39,]  3.449317e+00  1.403934e+00  5.647925e-15  3.185646e-17
#>  [40,]  4.657851e+00  8.421102e-15  7.486563e-16 -1.181964e-15
#>  [41,]  1.216464e-15 -1.282297e-15 -1.696138e-15 -3.069957e-15
#>  [42,]  2.561869e+00  1.813909e+00  1.004359e+00  1.219262e-15
#>  [43,]  4.763563e+00  2.297426e+00  9.840458e-01 -1.233578e-15
#>  [44,]  3.786766e-15  1.064580e-15  1.367516e-16 -4.911528e-16
#>  [45,]  2.322521e+00  4.891377e-15  2.740197e-15 -4.907552e-17
#>  [46,]  3.578837e+00  1.205801e+00  3.799056e-15  1.289297e-15
#>  [47,]  3.777639e+00  1.109727e+00  9.020006e-16  1.798304e-16
#>  [48,]  4.516682e+00  9.782575e-01  3.214198e-16 -2.287617e-16
#>  [49,]  2.063347e-14  5.597033e-16 -1.355170e-15 -1.425609e-15
#>  [50,]  1.370917e+00 -4.464653e-16 -1.662952e-15 -4.642549e-15
#>  [51,]  2.577010e+00  1.692569e+00  1.256201e-15  9.050605e-16
#>  [52,]  6.155455e+00  3.189310e+00  1.297130e+00  6.182648e-01
#>  [53,]  4.567986e-15  1.083053e-15  5.420429e-16  3.116500e-16
#>  [54,]  2.048035e+00  1.883183e-15  5.367290e-16 -9.832328e-16
#>  [55,]  8.487429e-01  4.037900e-15  8.306533e-16 -1.318206e-15
#>  [56,]  6.059123e+00  9.012340e-01  1.925150e-15  1.627564e-15
#>  [57,]  4.698130e-15  1.470225e-15  1.134893e-15  1.100387e-15
#>  [58,]  2.638664e+00  2.350273e+00 -1.205594e-15 -2.902738e-15
#>  [59,]  2.326983e+00  3.114720e-15  8.696001e-16  5.614470e-16
#>  [60,]  2.618679e+00 -1.040043e-15 -1.302092e-15 -4.611727e-15
#>  [61,]  2.110115e+00  8.012767e-15  9.673098e-16  1.237889e-16
#>  [62,]  2.332674e-15  1.022428e-15  1.860382e-16 -1.466652e-15
#>  [63,] -1.768078e-17 -7.363031e-16 -1.458137e-15 -1.686483e-15
#>  [64,]  4.490954e+00  3.549250e-15  3.254911e-15 -1.271745e-15
#>  [65,]  1.687999e+00  5.708600e-01  7.735135e-16  7.118750e-16
#>  [66,]  4.143722e-01  8.119952e-15  3.531724e-15  1.186092e-15
#>  [67,]  3.222874e+00  1.387247e+00  4.578213e-01 -8.110277e-16
#>  [68,]  5.092636e-15  2.312927e-15  1.385922e-15 -3.607298e-16
#>  [69,]  1.507924e-15  5.661960e-17 -1.919125e-16 -3.297778e-15
#>  [70,]  2.347853e+00  1.949877e-15 -6.084609e-16 -1.095855e-15
#>  [71,]  6.399143e-01  5.119288e-15  3.423196e-16  2.372877e-16
#>  [72,]  1.018230e+00 -1.816471e-15 -2.343457e-15 -3.991538e-15
#>  [73,]  5.451789e+00  1.070472e+00  1.943784e-01 -3.330669e-15
#>  [74,]  1.373472e+00  2.804591e-15  1.065849e-15  7.181793e-16
#>  [75,]  1.541426e-15  1.446443e-15 -2.031331e-15 -2.907473e-15
#>  [76,]  4.323332e+00  1.861636e+00  1.002020e+00 -9.384039e-17
#>  [77,]  4.388397e+00 -6.475958e-16 -1.991406e-15 -2.214760e-15
#>  [78,]  8.833216e-01  3.139925e-15  1.583275e-15 -6.084798e-16
#>  [79,]  1.881260e-15  8.132939e-16  7.476020e-16  4.887396e-16
#>  [80,]  4.954680e+00  5.553144e-15  4.163537e-15  9.164959e-16
#>  [81,]  8.509363e-01  2.416344e-15  1.524152e-15  1.341688e-15
#>  [82,]  1.266963e+00  2.449667e-15  8.330946e-16 -1.806750e-16
#>  [83,]  4.174528e+00  2.417690e+00  5.599945e-15  1.368987e-15
#>  [84,]  3.081141e+00  2.102144e+00  4.437651e-15  2.705512e-15
#>  [85,]  2.105556e+00  1.821034e-15 -6.011704e-17 -1.609313e-15
#>  [86,]  3.204241e-15  2.138312e-15 -2.790933e-16 -3.729621e-16
#>  [87,]  6.448479e+00  2.418847e+00  3.750945e-01  0.000000e+00
#>  [88,]  2.331587e+00  1.243636e+00  1.696763e-15  1.400364e-15
#>  [89,]  2.356325e+00  1.452355e+00 -2.322345e-16 -4.030681e-16
#>  [90,]  1.955483e+00  6.441338e-01  9.703669e-15 -2.300527e-16
#>  [91,]  3.232487e+00  2.337653e+00  7.683808e-01  3.463240e-15
#>  [92,]  8.750203e-01 -9.904070e-17 -1.206637e-15 -2.253955e-15
#>  [93,]  5.880818e-15  3.616031e-15  1.056405e-15  9.951829e-16
#>  [94,]  3.350702e+00  1.001517e+00  2.826682e-15  1.046610e-15
#>  [95,]  1.809467e+00  4.627378e-01  2.274977e-15  9.394195e-17
#>  [96,]  3.598487e+00  1.687196e+00  3.031034e-15  9.456814e-16
#>  [97,]  2.984017e+00  1.103167e+00  7.081509e-15  1.337224e-15
#>  [98,]  1.893770e+00  6.812114e-15  3.585200e-16 -1.714859e-15
#>  [99,]  4.907237e+00  2.236412e+00  4.359740e-01  1.527539e-15
#> [100,]  1.044092e-14  1.077489e-15 -1.960947e-16 -3.207059e-16
#> 
#> $BootMatEigsCI
#>                     LowCI       HighCI
#> Dimension 1  3.665611e+01 8.094953e+01
#> Dimension 2  1.800105e+01 4.377731e+01
#> Dimension 3  7.520578e+00 2.524283e+01
#> Dimension 4  3.798904e+00 1.466865e+01
#> Dimension 5  1.454289e+00 7.755588e+00
#> Dimension 6 -1.768078e-17 6.059123e+00
#> Dimension 7 -1.282297e-15 3.189310e+00
#> Dimension 8 -2.343457e-15 1.297130e+00
#> Dimension 9 -4.642549e-15 3.894129e-15
#> 
#> $FixedEigs
#> [1] 36.016716 34.543250 15.660149 12.717042  7.006519  6.674324  3.012660
#> [8]  2.103300  1.021405
#> 
#> $BootstrappedEstEigs
#>  Dimension 1  Dimension 2  Dimension 3  Dimension 4  Dimension 5  Dimension 6 
#> 53.097322677 29.696264929 16.658731544  8.901046851  5.083590533  2.282317236 
#>  Dimension 7  Dimension 8  Dimension 9 
#>  0.722775352  0.133443412  0.006182648 
#> 
#> $BootMatTaus
#>        Dimension 1 Dimension 2 Dimension 3 Dimension 4  Dimension 5
#>   [1,]   0.4887473   0.2411780  0.16088045  0.07279292 3.640138e-02
#>   [2,]   0.3821180   0.2513967  0.15683855  0.09179852 6.052852e-02
#>   [3,]   0.4331167   0.3501458  0.14543040  0.04636067 2.494636e-02
#>   [4,]   0.5280594   0.2137206  0.09609820  0.07824326 5.168233e-02
#>   [5,]   0.5256340   0.1840148  0.13403026  0.08626616 6.044641e-02
#>   [6,]   0.4227744   0.2650607  0.14508739  0.07116440 5.154900e-02
#>   [7,]   0.4517028   0.3017451  0.09847786  0.08829952 3.025992e-02
#>   [8,]   0.3868329   0.2505226  0.14171828  0.09148847 5.615500e-02
#>   [9,]   0.4561155   0.2400431  0.17817936  0.07874538 3.257987e-02
#>  [10,]   0.5976310   0.2181049  0.07534671  0.05391211 3.412001e-02
#>  [11,]   0.4583140   0.3336265  0.15474180  0.05331773 8.780402e-18
#>  [12,]   0.5353085   0.1980974  0.14069084  0.05938458 5.172133e-02
#>  [13,]   0.4582888   0.2759290  0.15281799  0.09224018 2.072400e-02
#>  [14,]   0.6093950   0.2262185  0.10851161  0.04608549 9.789349e-03
#>  [15,]   0.4250620   0.2546124  0.16462156  0.06498183 4.903764e-02
#>  [16,]   0.4249091   0.2723353  0.17212016  0.06438012 4.483831e-02
#>  [17,]   0.6161300   0.1623504  0.09907019  0.05203561 4.816756e-02
#>  [18,]   0.4587178   0.1982744  0.16277054  0.07747707 5.272116e-02
#>  [19,]   0.3519838   0.3039908  0.13257562  0.10263470 4.842061e-02
#>  [20,]   0.4013858   0.2750482  0.11844151  0.07581464 6.888501e-02
#>  [21,]   0.4327197   0.2520363  0.11102595  0.10715004 5.105296e-02
#>  [22,]   0.4191191   0.2968220  0.10993218  0.06803029 4.608382e-02
#>  [23,]   0.3423682   0.2672519  0.22750446  0.09108569 3.340748e-02
#>  [24,]   0.4977182   0.2596993  0.14036797  0.05174781 3.795542e-02
#>  [25,]   0.5676696   0.2040305  0.12262826  0.05768759 3.869368e-02
#>  [26,]   0.4595432   0.3103092  0.12003660  0.05443729 5.003045e-02
#>  [27,]   0.4655054   0.2673261  0.15157491  0.08371024 2.736497e-02
#>  [28,]   0.3219695   0.2408935  0.22815628  0.09269442 7.419649e-02
#>  [29,]   0.4245092   0.2454780  0.17016156  0.09340971 4.202668e-02
#>  [30,]   0.4954530   0.2840448  0.09028352  0.07497380 2.933524e-02
#>  [31,]   0.3233377   0.2671513  0.17712328  0.11716399 5.025615e-02
#>  [32,]   0.4531690   0.3155142  0.09906449  0.06678721 5.289666e-02
#>  [33,]   0.4268127   0.2445062  0.15802697  0.08642619 6.220953e-02
#>  [34,]   0.3448517   0.2930747  0.15656236  0.06261451 5.945399e-02
#>  [35,]   0.5221052   0.2777528  0.11139994  0.04802012 2.438284e-02
#>  [36,]   0.4302914   0.2407440  0.15280001  0.08805045 5.770502e-02
#>  [37,]   0.4100931   0.3578730  0.10388662  0.06528882 3.836535e-02
#>  [38,]   0.4688451   0.2716273  0.12145674  0.06983532 6.154756e-02
#>  [39,]   0.4915007   0.2450585  0.11513372  0.07031906 4.267447e-02
#>  [40,]   0.4856284   0.2297228  0.13995478  0.06424769 4.565283e-02
#>  [41,]   0.4912000   0.2148663  0.16296084  0.08429728 4.667550e-02
#>  [42,]   0.4341344   0.2203771  0.13532680  0.11562949 5.252013e-02
#>  [43,]   0.3985849   0.2818712  0.11768798  0.08892869 4.716592e-02
#>  [44,]   0.5219454   0.2316774  0.14308456  0.06820749 3.508522e-02
#>  [45,]   0.5112611   0.2169358  0.12647721  0.09837889 2.744489e-02
#>  [46,]   0.3395725   0.2585611  0.21763416  0.07590528 6.432463e-02
#>  [47,]   0.4338221   0.2412827  0.13925673  0.07627789 6.668076e-02
#>  [48,]   0.4579798   0.2912492  0.09295565  0.07111123 4.524713e-02
#>  [49,]   0.7015571   0.1794866  0.07359503  0.03218526 1.317598e-02
#>  [50,]   0.4593123   0.2645757  0.15863779  0.06911717 3.815043e-02
#>  [51,]   0.4222014   0.2034324  0.17679621  0.10099179 5.295749e-02
#>  [52,]   0.3280104   0.3036315  0.13842738  0.07142218 6.021117e-02
#>  [53,]   0.4484011   0.2620656  0.21501955  0.04027830 3.423550e-02
#>  [54,]   0.3841201   0.2609607  0.21538948  0.07832221 4.536466e-02
#>  [55,]   0.3442736   0.3350802  0.16710255  0.10111237 4.615356e-02
#>  [56,]   0.4039443   0.2440792  0.13720319  0.08852028 6.600934e-02
#>  [57,]   0.4998048   0.2454643  0.13627552  0.07865826 3.979706e-02
#>  [58,]   0.5194583   0.1882060  0.13547427  0.07609138 3.950191e-02
#>  [59,]   0.5033987   0.2066090  0.13263570  0.11204151 2.599103e-02
#>  [60,]   0.5659076   0.1758150  0.11536816  0.07613678 4.312876e-02
#>  [61,]   0.4854154   0.2700555  0.14242978  0.06681385 2.020443e-02
#>  [62,]   0.5094361   0.1887602  0.17333119  0.07868747 4.978508e-02
#>  [63,]   0.3971346   0.2724220  0.23189219  0.07344115 2.511001e-02
#>  [64,]   0.3428980   0.2782147  0.20872717  0.07742858 5.559688e-02
#>  [65,]   0.3913499   0.3219309  0.13465574  0.09239059 3.811050e-02
#>  [66,]   0.5780318   0.2084714  0.09481453  0.06616826 4.955509e-02
#>  [67,]   0.4020047   0.2500945  0.14937819  0.09066116 6.113023e-02
#>  [68,]   0.4122594   0.3040432  0.18726745  0.07770190 1.872811e-02
#>  [69,]   0.4632312   0.2395108  0.17733582  0.08474574 3.517646e-02
#>  [70,]   0.4194120   0.2510407  0.15892398  0.09864576 4.997229e-02
#>  [71,]   0.5774042   0.2185628  0.10061747  0.07376241 2.302338e-02
#>  [72,]   0.4606332   0.2564027  0.16753817  0.06063738 4.748369e-02
#>  [73,]   0.3821227   0.1959121  0.18623200  0.12981990 5.258407e-02
#>  [74,]   0.4487706   0.2670834  0.15854413  0.06987475 4.120757e-02
#>  [75,]   0.4800418   0.2875833  0.13575116  0.05561816 4.100561e-02
#>  [76,]   0.3760819   0.3135168  0.12437597  0.06924578 4.872187e-02
#>  [77,]   0.4615129   0.2933671  0.11060642  0.06328065 4.182481e-02
#>  [78,]   0.5037722   0.2348793  0.09747378  0.08194291 7.238331e-02
#>  [79,]   0.5279278   0.2299253  0.18289479  0.04232502 1.692706e-02
#>  [80,]   0.4719225   0.2555624  0.11833259  0.07107535 4.811260e-02
#>  [81,]   0.4623767   0.3072409  0.09905975  0.07846611 4.687820e-02
#>  [82,]   0.5279924   0.1874238  0.13855806  0.07523089 5.873711e-02
#>  [83,]   0.4224013   0.2446254  0.15245830  0.07403616 5.112502e-02
#>  [84,]   0.5055952   0.1864437  0.11946464  0.09303966 5.243669e-02
#>  [85,]   0.4466407   0.2296853  0.16457009  0.11176352 2.699109e-02
#>  [86,]   0.4181832   0.3760512  0.16018795  0.02683186 1.874580e-02
#>  [87,]   0.3841280   0.2916885  0.11748086  0.07275916 6.111367e-02
#>  [88,]   0.4649949   0.2546746  0.14049555  0.06599468 3.838390e-02
#>  [89,]   0.4283231   0.2290540  0.16213883  0.07951381 6.257460e-02
#>  [90,]   0.4868912   0.2048083  0.14954586  0.08252584 5.078319e-02
#>  [91,]   0.4023148   0.2174471  0.17747242  0.09805210 5.536534e-02
#>  [92,]   0.4454421   0.2791644  0.14207121  0.08415331 3.980969e-02
#>  [93,]   0.5394861   0.3111997  0.06174874  0.04893491 3.863054e-02
#>  [94,]   0.4522799   0.2263135  0.12143853  0.10996661 4.735347e-02
#>  [95,]   0.5014412   0.2390334  0.13882316  0.05479579 4.683196e-02
#>  [96,]   0.3811727   0.2730015  0.14989818  0.11179137 4.016471e-02
#>  [97,]   0.4933737   0.1842092  0.14029195  0.10223713 4.087269e-02
#>  [98,]   0.3568436   0.2793146  0.23515765  0.06586202 4.438652e-02
#>  [99,]   0.3995494   0.2295405  0.15570480  0.09497382 6.049868e-02
#> [100,]   0.4929197   0.2677410  0.16551362  0.05099224 2.283346e-02
#>          Dimension 6   Dimension 7   Dimension 8   Dimension 9
#>   [1,]  1.943268e-17 -1.069016e-17 -2.083541e-17 -2.556721e-17
#>   [2,]  2.774450e-02  1.806568e-02  1.150948e-02  2.970482e-17
#>   [3,]  4.912326e-17  1.328468e-17  8.753058e-18 -1.117853e-17
#>   [4,]  2.157141e-02  1.062479e-02  1.700214e-17  9.095892e-18
#>   [5,]  9.608376e-03  1.614251e-17  1.175543e-17 -3.151551e-19
#>   [6,]  3.353566e-02  1.082846e-02  1.810784e-17  9.990458e-18
#>   [7,]  2.300553e-02  6.509298e-03  7.612278e-18  6.864872e-18
#>   [8,]  3.650122e-02  2.611631e-02  1.066528e-02  1.186127e-17
#>   [9,]  1.433681e-02  1.823458e-17  5.958422e-18 -1.195409e-17
#>  [10,]  2.088531e-02  5.218544e-17  6.450985e-18  5.521237e-18
#>  [11,] -5.372971e-18 -6.931331e-18 -2.010330e-17 -4.080879e-17
#>  [12,]  1.479730e-02  1.291224e-17  1.178918e-17 -9.157045e-18
#>  [13,]  5.525003e-17  5.604775e-18  3.243137e-18  2.250902e-18
#>  [14,]  3.587059e-17  1.745161e-17  1.344268e-17 -1.848698e-18
#>  [15,]  2.861466e-02  1.306997e-02  2.825352e-17  2.806567e-18
#>  [16,]  2.141695e-02  4.946173e-17 -2.716214e-18 -1.666990e-17
#>  [17,]  2.224623e-02  3.965362e-17  1.014565e-17 -1.873847e-17
#>  [18,]  2.919459e-02  2.084446e-02  2.748574e-17  1.589481e-18
#>  [19,]  3.231760e-02  1.954070e-02  8.536094e-03  2.825023e-17
#>  [20,]  4.459643e-02  1.582838e-02  1.099614e-17 -2.582927e-17
#>  [21,]  2.444075e-02  1.632761e-02  5.246753e-03 -4.865906e-18
#>  [22,]  3.798489e-02  1.238843e-02  9.639362e-03  1.457800e-17
#>  [23,]  2.506749e-02  1.331472e-02  1.823752e-17  2.281020e-18
#>  [24,]  1.251138e-02  5.470892e-17 -2.072650e-19 -2.133447e-17
#>  [25,]  9.290316e-03  2.575458e-18 -1.717700e-18 -5.169018e-18
#>  [26,]  5.643253e-03  5.508411e-17  3.562187e-17 -1.030277e-17
#>  [27,]  4.518373e-03  1.082496e-16  4.786090e-17 -1.065150e-17
#>  [28,]  3.005523e-02  1.203461e-02 -1.324733e-17 -2.395047e-17
#>  [29,]  2.441490e-02  1.569857e-17  1.031883e-17  4.664413e-18
#>  [30,]  1.839618e-02  7.513473e-03 -6.695588e-18 -2.099078e-17
#>  [31,]  3.481915e-02  2.517478e-02  4.973643e-03  5.877894e-17
#>  [32,]  1.256837e-02  3.162685e-17  8.432768e-18 -2.126549e-18
#>  [33,]  1.809663e-02  3.921746e-03  9.499729e-19 -1.878291e-18
#>  [34,]  5.059261e-02  2.751399e-02  5.336116e-03  5.045517e-18
#>  [35,]  1.633907e-02  2.294223e-18  1.175335e-18 -2.582141e-17
#>  [36,]  3.040910e-02  2.180048e-17  8.442594e-18 -1.117782e-17
#>  [37,]  1.525564e-02  9.237417e-03  4.105090e-17  7.388105e-18
#>  [38,]  6.687954e-03  4.440518e-17 -6.536939e-18 -1.040729e-17
#>  [39,]  2.509821e-02  1.021542e-02  4.109592e-17  2.317967e-19
#>  [40,]  3.479349e-02  6.290445e-17  5.592358e-18 -8.829108e-18
#>  [41,]  9.265698e-18 -9.767144e-18 -1.291933e-17 -2.338359e-17
#>  [42,]  2.000495e-02  1.416433e-02  7.842768e-03  9.520891e-18
#>  [43,]  3.893804e-02  1.877948e-02  8.043729e-03 -1.008344e-17
#>  [44,]  3.801628e-17  1.068759e-17  1.372884e-18 -4.930805e-18
#>  [45,]  1.950217e-02  4.107281e-17  2.300939e-17 -4.120863e-19
#>  [46,]  3.291310e-02  1.108926e-02  3.493836e-17  1.185714e-17
#>  [47,]  3.298886e-02  9.690875e-03  7.876870e-18  1.570399e-18
#>  [48,]  3.407649e-02  7.380546e-03  2.424979e-18 -1.725912e-18
#>  [49,]  1.408591e-16  3.820942e-18 -9.251377e-18 -9.732241e-18
#>  [50,]  1.020661e-02 -3.323978e-18 -1.238084e-17 -3.456423e-17
#>  [51,]  2.632836e-02  1.729236e-02  1.283414e-17  9.246670e-18
#>  [52,]  5.373501e-02  2.784158e-02  1.132350e-02  5.397240e-03
#>  [53,]  4.843258e-17  1.148319e-17  5.747070e-18  3.304303e-18
#>  [54,]  1.584286e-02  1.456762e-17  4.151941e-18 -7.605932e-18
#>  [55,]  6.277684e-03  2.986612e-17  6.143885e-18 -9.750046e-18
#>  [56,]  5.244327e-02  7.800413e-03  1.666267e-17  1.408699e-17
#>  [57,]  5.179390e-17  1.620830e-17  1.251148e-17  1.213107e-17
#>  [58,]  2.182680e-02  1.944125e-02 -9.972567e-18 -2.401119e-17
#>  [59,]  1.932412e-02  2.586578e-17  7.221478e-18  4.662461e-18
#>  [60,]  2.364375e-02 -9.390432e-18 -1.175644e-17 -4.163876e-17
#>  [61,]  1.508096e-02  5.726712e-17  6.913348e-18  8.847174e-19
#>  [62,]  2.776245e-17  1.216849e-17  2.214144e-18 -1.745544e-17
#>  [63,] -1.538639e-19 -6.407546e-18 -1.268917e-17 -1.467632e-17
#>  [64,]  3.713466e-02  2.934793e-17  2.691411e-17 -1.051577e-17
#>  [65,]  1.611311e-02  5.449253e-03  7.383720e-18  6.795337e-18
#>  [66,]  2.958885e-03  5.798169e-17  2.521879e-17  8.469463e-18
#>  [67,]  2.971792e-02  1.279171e-02  4.221542e-03 -7.478435e-18
#>  [68,]  4.337292e-17  1.969872e-17  1.180361e-17 -3.072260e-18
#>  [69,]  1.301041e-17  4.885154e-19 -1.655826e-18 -2.845332e-17
#>  [70,]  2.200527e-02  1.827523e-17 -5.702805e-18 -1.027091e-17
#>  [71,]  6.629680e-03  5.303717e-17  3.546521e-18  2.458362e-18
#>  [72,]  7.304868e-03 -1.303152e-17 -1.681216e-17 -2.863564e-17
#>  [73,]  4.328646e-02  8.499403e-03  1.543338e-03 -2.644506e-17
#>  [74,]  1.451956e-02  2.964854e-17  1.126755e-17  7.592181e-18
#>  [75,]  1.117731e-17  1.048856e-17 -1.472975e-17 -2.108290e-17
#>  [76,]  4.094010e-02  1.762889e-02  9.488699e-03 -8.886283e-19
#>  [77,]  2.940819e-02 -4.339766e-18 -1.334511e-17 -1.484189e-17
#>  [78,]  9.548452e-03  3.394169e-17  1.711474e-17 -6.577491e-18
#>  [79,]  1.657498e-17  7.165586e-18  6.586803e-18  4.306076e-18
#>  [80,]  3.499464e-02  3.922156e-17  2.940684e-17  6.473162e-18
#>  [81,]  5.978373e-03  1.697637e-17  1.070814e-17  9.426222e-18
#>  [82,]  1.205771e-02  2.331352e-17  7.928575e-18 -1.719487e-18
#>  [83,]  3.505291e-02  2.030099e-02  4.702194e-17  1.149519e-17
#>  [84,]  2.557280e-02  1.744734e-02  3.683154e-17  2.245516e-17
#>  [85,]  2.034927e-02  1.759949e-17 -5.810047e-19 -1.555330e-17
#>  [86,]  2.815738e-17  1.879049e-17 -2.452542e-18 -3.277417e-18
#>  [87,]  5.081372e-02  1.906040e-02  2.955727e-03  0.000000e+00
#>  [88,]  2.312294e-02  1.233346e-02  1.682724e-17  1.388777e-17
#>  [89,]  2.375432e-02  1.464132e-02 -2.341177e-18 -4.063366e-18
#>  [90,]  1.914067e-02  6.304915e-03  9.498152e-17 -2.251804e-18
#>  [91,]  2.516640e-02  1.819970e-02  5.982198e-03  2.696292e-17
#>  [92,]  9.359288e-03 -1.059347e-18 -1.290629e-17 -2.410849e-17
#>  [93,]  5.256939e-17  3.232417e-17  9.443341e-18  8.896068e-18
#>  [94,]  3.283401e-02  9.814007e-03  2.769907e-17  1.025588e-17
#>  [95,]  1.518996e-02  3.884551e-03  1.909778e-17  7.886157e-19
#>  [96,]  2.993579e-02  1.403577e-02  2.521515e-17  7.867117e-18
#>  [97,]  2.848476e-02  1.053059e-02  6.759851e-17  1.276484e-17
#>  [98,]  1.843566e-02  6.631525e-17  3.490157e-18 -1.669398e-17
#>  [99,]  3.867251e-02  1.762451e-02  3.435785e-03  1.203809e-17
#> [100,]  1.147407e-16  1.184108e-17 -2.154985e-18 -3.524402e-18
#> 
#> $BootMatTausCI
#>                     LowCI       HighCI
#> Dimension 1  3.233377e-01 6.093950e-01
#> Dimension 2  1.758150e-01 3.501458e-01
#> Dimension 3  7.359503e-02 2.281563e-01
#> Dimension 4  3.218526e-02 1.156295e-01
#> Dimension 5  9.789349e-03 6.888501e-02
#> Dimension 6 -1.538639e-19 5.081372e-02
#> Dimension 7 -1.069016e-17 2.611631e-02
#> Dimension 8 -2.010330e-17 1.066528e-02
#> Dimension 9 -4.080879e-17 2.970482e-17
#> 
#> $BootstrappedEstInertia
#>   [1]  90.16363 131.09418 117.96314 120.18165  97.82391 125.64877 130.48750
#>   [8] 124.76770  92.07591  99.81295 129.14891 113.54162  85.95715 148.55829
#>  [15] 105.95141 109.06620 128.53345  97.94613 115.78451  92.91341 136.89820
#>  [22] 113.18955 109.82688 155.76398  91.55031 136.68400 143.64145 106.85729
#>  [29] 112.67150 130.12025 114.17019 117.94761 119.59507 116.20681 142.53812
#>  [36]  98.48988 125.59675 107.44572 137.43276 133.87131 131.28683 128.06174
#>  [43] 122.33702  99.60905 119.09039 108.73596 114.51256 132.54541 146.48307
#>  [50] 134.31656  97.87962 114.55204  94.31639 129.27183 135.19999 115.53670
#>  [57]  90.70816 120.89105 120.41858 110.75562 139.91914  84.02263 114.91188
#>  [64] 120.93697 104.75933 140.04338 108.44885 117.41512 115.90137 106.69503
#>  [71]  96.52265 139.39055 125.94676  94.59460 137.90668 105.60140 149.22366
#>  [78]  92.50940 113.49999 141.58396 142.33576 105.07494 119.09218 120.48509
#>  [85] 103.47084 113.79755 126.90431 100.83433  99.19562 102.16376 128.44456
#>  [92]  93.49219 111.86772 102.04974 119.12259 120.20686 104.75837 102.72319
#>  [99] 126.89212  90.99584
#>