Boostrap for eigenvalues.

Boostrap for eigenvalues.

Usage

Boot4Eigs(
  data,
  center = FALSE,
  scale = FALSE,
  design = NULL,
  niter = 100,
  CI.perc = c(0.025, 0.975),
  suppressProgressBar = TRUE
)

Arguments

data

the data-set;

center

whether to center the data (default to FALSE), see expo.scale;

scale

whether to scale the data ( default to FALSE), see expo.scale;

design

whether there is a design;

niter

number of bootstrap iterations;

CI.perc

vector of the two levels of confidence for the left and right limits of the bootstrap confidence interval (default to c(0.025, 0.0975));

suppressProgressBar

whether to show a progress bar or not (default to FALSE).

Value

list containing four objects

Examples

Boot4Eigs(matrix(rnorm(100), 10, 10))
#> $BootMatEigs
#>        Dimension 1 Dimension 2 Dimension 3 Dimension 4 Dimension 5
#>   [1,]    47.63818    27.68070    9.485068    7.396727   5.5181847
#>   [2,]    56.61219    21.28375    9.701157    6.445930   4.7714365
#>   [3,]    62.52224    26.68252   10.919956    5.974453   4.3522930
#>   [4,]    40.44990    14.23746   10.494414    7.479196   4.2038325
#>   [5,]    52.53174    20.78363   13.285785    9.489407   6.2624764
#>   [6,]    58.83538    23.87959   19.325540    5.594558   2.4655543
#>   [7,]    44.14442    27.21249   10.025220    7.246332   5.7892801
#>   [8,]    47.15138    41.41510   16.898555    9.151762   1.7559913
#>   [9,]    38.92251    17.87378   13.656724    7.741906   4.8630817
#>  [10,]    60.08590    24.75858   16.951193    9.398999   6.2728188
#>  [11,]    41.83698    20.01323    9.886207    7.805913   5.1624518
#>  [12,]    37.63315    14.10489   11.043148    6.396401   4.9013585
#>  [13,]    54.77962    18.91875    8.198706    7.334476   6.0608294
#>  [14,]    45.39142    16.98339   10.139144    7.672728   3.9847638
#>  [15,]    37.99731    24.80004   13.272409    5.423770   3.5868159
#>  [16,]    53.73262    14.75583   12.937827    6.138600   5.5990582
#>  [17,]    50.34685    25.20111   10.671564    9.025714   6.6668152
#>  [18,]    47.83473    25.51118    9.923239    8.331400   7.7136756
#>  [19,]    46.41103    23.03398   10.646323    9.512756   5.2385297
#>  [20,]    51.72882    16.31763    9.426608    9.228881   5.7567766
#>  [21,]    66.45072    44.39215   10.024262    7.733590   5.5049736
#>  [22,]    53.62411    17.49009   10.120070    8.658044   3.9726405
#>  [23,]    70.93200    14.19824   12.075475    3.238346   2.5600665
#>  [24,]    42.63007    15.90623    8.714508    7.929585   5.2550835
#>  [25,]    46.69652    20.45814   14.296891    8.905513   5.9206670
#>  [26,]    78.59799    39.19443   12.131342    5.466010   0.9740186
#>  [27,]    53.37862    24.58942   14.966801    5.508448   3.0722692
#>  [28,]    34.11208    19.55030   13.588086    9.103180   6.1797070
#>  [29,]    45.17995    17.90243   11.178385    4.569316   3.1220088
#>  [30,]    79.95488    30.96803   10.227208    5.254397   3.4780071
#>  [31,]    42.28527    15.49175   13.831006    5.541826   3.1808368
#>  [32,]    60.08232    22.33936   17.367143    9.096047   6.1855910
#>  [33,]    42.57998    16.30939    8.784012    8.061538   2.7852867
#>  [34,]    42.22710    19.43689    8.497334    6.094357   4.6880649
#>  [35,]    51.98585    21.85650    9.660989    7.324449   4.6203061
#>  [36,]    53.88593    27.41573   11.494799   10.239634   6.6999570
#>  [37,]    46.03804    14.70941    9.510504    7.678058   4.6890153
#>  [38,]    72.99664    21.16882   12.589643    8.269894   5.1770176
#>  [39,]    75.40943    27.18954   12.567012    6.675887   2.4606114
#>  [40,]    39.40219    23.74762   12.484562    8.819477   4.1208120
#>  [41,]    47.83896    17.09256    8.882069    8.445991   5.5003764
#>  [42,]    49.11380    16.34199    8.620726    6.314134   4.3936234
#>  [43,]    66.83491    17.43973    9.183980    7.557388   4.2448198
#>  [44,]    68.59420    24.18653   10.346669    9.364387   4.6828606
#>  [45,]    29.57950    14.22698    9.682403    6.573548   4.4788376
#>  [46,]    68.85943    43.47102    8.221434    6.035539   3.6334785
#>  [47,]    35.89253    16.13837   10.143339    7.239994   4.7466310
#>  [48,]    51.32778    19.30467   12.582868    8.708355   5.8326458
#>  [49,]    63.64793    37.02155   11.883212    7.318159   5.3438362
#>  [50,]    51.98045    19.50202   11.170418    5.872880   3.0995210
#>  [51,]    62.44376    28.17764   13.462001    8.092578   3.0786227
#>  [52,]    50.83414    19.61064   12.404795    6.869826   1.0752566
#>  [53,]    68.19546    22.35253    9.458672    6.209220   4.7800562
#>  [54,]    50.68432    23.72716   11.572704    8.644518   7.2424015
#>  [55,]    57.04708    28.67766   14.300856    6.062632   4.9530420
#>  [56,]    55.79900    30.18383   10.186088    8.870459   4.9583100
#>  [57,]    51.38740    17.79009   12.970194    9.894327   4.4020846
#>  [58,]    35.44119    11.77807   11.433408    6.514207   4.3339314
#>  [59,]    84.85009    34.51491    7.063472    4.145544   2.2983141
#>  [60,]    37.29165    26.81061   11.172432    7.578167   3.5038427
#>  [61,]    41.14990    16.43433   11.657573    7.271045   5.1513314
#>  [62,]    89.31025    33.34640   10.060832    6.942697   4.3398296
#>  [63,]    61.82483    19.77718   14.246059   10.281643   4.9337025
#>  [64,]    66.64918    32.47113   13.019789    7.800937   2.6527645
#>  [65,]    59.84103    24.82516   12.905619    6.119895   2.1788679
#>  [66,]    60.96709    27.00841    9.705137    7.024972   5.3582781
#>  [67,]    58.99935    20.48024   10.434107    7.913278   3.9105504
#>  [68,]    39.11334    13.66719   10.419996    7.409435   4.1513523
#>  [69,]    45.51055    35.15562   11.613794    8.796304   5.5676723
#>  [70,]    52.74632    17.84927   10.607309    9.076587   6.0560193
#>  [71,]    35.33641    14.16711   13.636932    4.444268   3.8908185
#>  [72,]    49.38858    16.47027   12.500560    6.410729   1.6708312
#>  [73,]    44.00176    18.51278    8.634365    5.342699   3.3031674
#>  [74,]    47.90343    25.02864   10.251833    9.059800   5.4850150
#>  [75,]    71.95136    31.31712   15.267735    5.011500   2.3248644
#>  [76,]    73.25042    25.35363   11.186333    7.465136   5.4653136
#>  [77,]    74.09464    15.62841   12.044618    5.460931   4.1532566
#>  [78,]    51.60765    31.50800    9.034275    6.329462   5.0298439
#>  [79,]    61.05542    22.91832   17.043832   10.205250   3.7810562
#>  [80,]    53.10322    34.56947   16.230179   10.060622   3.6625298
#>  [81,]    56.02512    18.83601   11.896227    8.298719   5.5843368
#>  [82,]    50.49187    19.39706    8.555944    6.723687   4.5870563
#>  [83,]    62.63834    22.40423   12.938642    5.910931   3.8329311
#>  [84,]    37.94383    19.08302   11.419789    7.133012   4.4411398
#>  [85,]    58.74236    16.05200    8.853243    6.427010   2.1436536
#>  [86,]    55.47497    32.59791    9.581904    9.250823   4.3285258
#>  [87,]    63.24259    25.09484    9.834494    7.121564   2.7584259
#>  [88,]    70.50832    30.06305   14.429510    8.726387   4.9842972
#>  [89,]    51.25752    22.02313   14.423932    8.534318   5.0804904
#>  [90,]    65.18729    21.45064   15.785888    9.923008   2.3699505
#>  [91,]    75.14984    29.09436   15.624243    9.032304   2.3378253
#>  [92,]    52.09900    22.63884   15.240543   10.316581   4.9993271
#>  [93,]    63.91831    15.94936   13.849319    6.134347   5.2088375
#>  [94,]    64.58440    22.03386   10.258286    6.455527   4.5107724
#>  [95,]    50.39110    22.79322   12.095158    7.548368   4.5737149
#>  [96,]    67.03839    30.89963   15.454065    9.377895   4.9247440
#>  [97,]    46.81967    15.09992   10.430207    8.934183   5.9051946
#>  [98,]    66.76614    33.45295    9.809459    8.184293   5.5730812
#>  [99,]    65.97486    25.60592    9.575071    7.101327   6.2671273
#> [100,]    52.91326    18.97099   12.340305    7.263790   4.1312176
#>          Dimension 6   Dimension 7   Dimension 8   Dimension 9
#>   [1,]  3.420969e+00  2.705588e+00  6.708958e-01  6.333923e-16
#>   [2,]  8.371336e-01  7.275113e-15  1.079485e-15  7.125619e-16
#>   [3,]  2.194537e+00  6.550584e-01  1.981595e-15 -1.418857e-15
#>   [4,]  3.616484e+00  9.598251e-01  2.479460e-15  7.325922e-17
#>   [5,]  4.451869e+00  3.017792e+00  1.766801e-15  1.377613e-17
#>   [6,]  1.222923e-14  1.722365e-15 -2.681992e-16 -7.001853e-16
#>   [7,]  3.100968e+00  2.276712e+00  5.133640e-01  2.775558e-17
#>   [8,]  2.586860e-15  4.843767e-16  4.470486e-16  1.700291e-17
#>   [9,]  3.371056e+00  2.392238e+00  7.366786e-01 -4.440892e-16
#>  [10,]  3.145158e+00  4.986907e-01 -6.369026e-16 -2.652133e-15
#>  [11,]  3.444125e+00  1.005326e+00  2.569335e-15  8.640896e-17
#>  [12,]  2.767847e+00  1.059513e+00  2.543540e-15  1.374581e-16
#>  [13,]  3.541457e+00  2.374088e+00  4.953391e-01  6.228014e-15
#>  [14,]  8.367718e-01  8.346738e-16  8.453044e-17 -6.138929e-16
#>  [15,]  2.175884e+00  5.799439e-16  5.162538e-16 -1.649559e-15
#>  [16,]  2.854032e+00  1.069499e+00  2.331633e-15 -9.727910e-17
#>  [17,]  3.371701e+00  1.436672e+00  4.098933e-01 -1.139445e-16
#>  [18,]  2.975579e+00  8.179862e-01  2.363645e-15  6.227455e-16
#>  [19,]  3.588487e+00  2.796586e+00  6.510205e-01  1.896968e-01
#>  [20,]  3.296508e+00  1.468237e+00  4.341231e-01 -5.508844e-16
#>  [21,]  1.939597e+00  6.133464e-01  4.480001e-15  4.386658e-15
#>  [22,]  1.581922e+00  6.466089e-01  1.054578e-15 -4.689998e-16
#>  [23,]  2.416794e-15  2.087800e-15 -4.573620e-16 -7.727825e-16
#>  [24,]  3.305979e+00  2.909978e+00  5.016709e-01  3.675493e-15
#>  [25,]  3.209838e+00  2.407805e+00  5.695714e-01  6.028529e-15
#>  [26,]  1.164490e-15 -3.070439e-16 -1.881651e-15 -3.085355e-15
#>  [27,]  9.484977e-01  9.654095e-15  1.861064e-15 -2.684622e-16
#>  [28,]  3.963333e+00  7.852823e-01  3.482164e-15  2.305782e-15
#>  [29,]  1.417573e+00  3.098377e-15 -2.717792e-16 -2.963860e-15
#>  [30,]  2.304161e+00  1.530934e-14  1.345559e-15 -4.002548e-16
#>  [31,]  4.035586e-15  2.356970e-15  8.083274e-16 -1.273088e-15
#>  [32,]  3.036955e+00  4.390962e-01  6.947738e-15  8.854105e-17
#>  [33,]  1.621400e+00  1.415311e-15 -1.018337e-16 -3.064317e-15
#>  [34,]  3.367964e+00  2.186309e+00  1.017662e+00  1.343300e-15
#>  [35,]  5.552782e-01  7.838885e-15  7.082361e-16  5.541877e-16
#>  [36,]  1.529078e+00  3.449519e-15  1.155411e-15 -2.444413e-15
#>  [37,]  3.395518e+00  5.594945e-01  7.161410e-15 -7.427709e-16
#>  [38,]  3.059388e+00  2.153328e+00  5.139461e-15 -2.512797e-16
#>  [39,]  1.425028e+00  2.486419e-15  5.777337e-16 -1.583410e-15
#>  [40,]  2.499222e+00  3.857128e-01  1.899923e-15 -1.108889e-15
#>  [41,]  3.389216e+00  1.213690e+00  1.390778e-15  5.983487e-16
#>  [42,]  3.656384e+00  8.362257e-16  6.122591e-16 -1.631535e-15
#>  [43,]  1.836515e+00  5.892558e-15  1.923584e-15  9.179335e-16
#>  [44,]  2.365495e+00  6.267858e-15  3.561222e-15 -1.341507e-15
#>  [45,]  2.137758e+00  4.183720e-15  4.603571e-16 -6.241876e-16
#>  [46,]  1.894815e+00  5.931094e-01  4.581580e-15 -4.189521e-17
#>  [47,]  2.301192e+00  5.935783e-15  1.709072e-15  1.438432e-16
#>  [48,]  3.388292e+00  2.433898e+00  1.085433e+00  4.095074e-15
#>  [49,]  2.909224e+00  5.201832e-01  3.257937e-01 -2.699757e-15
#>  [50,]  2.246221e+00  9.092183e-01  9.050295e-15  6.780764e-16
#>  [51,]  4.758295e-01  1.215742e-15 -6.857394e-16 -7.863690e-16
#>  [52,]  4.193233e-15  1.050879e-15  5.639420e-16  7.382604e-17
#>  [53,]  2.183514e+00  4.447540e-01  1.040662e-15 -5.259510e-15
#>  [54,]  3.889796e+00  2.365766e+00  2.685735e-01  3.049231e-15
#>  [55,]  1.375666e+00  3.926240e-15  1.244423e-15  8.784504e-17
#>  [56,]  2.242909e+00  2.033005e+00  2.405954e-15  1.148605e-15
#>  [57,]  2.466447e-15  1.541211e-15 -7.193370e-16 -1.925033e-15
#>  [58,]  2.854958e+00  2.049184e-15  8.829222e-16  7.768509e-16
#>  [59,]  3.286941e-15  1.896363e-15 -5.785194e-16 -2.676064e-15
#>  [60,]  1.399423e+00  6.091352e-16 -5.614402e-16 -1.064026e-15
#>  [61,]  3.026159e+00  5.285790e-01  5.865060e-16 -3.638906e-18
#>  [62,]  1.474860e+00  4.160652e-01 -4.574572e-16 -1.690131e-15
#>  [63,]  3.022789e+00  5.447796e-15  3.736680e-15  4.212853e-16
#>  [64,]  2.335921e+00  1.335924e-14  1.875653e-15  4.231026e-16
#>  [65,]  9.820625e-01  1.716013e-14  2.805988e-15  4.543750e-17
#>  [66,]  8.763319e-01  7.216686e-15  1.553857e-15 -2.058165e-15
#>  [67,]  2.979223e+00  1.194038e+00  2.124498e-15  1.998712e-15
#>  [68,]  2.499307e+00  1.022768e+00  7.112065e-15  1.295987e-15
#>  [69,]  1.891418e+00  5.584191e-15  1.149822e-15  5.931186e-16
#>  [70,]  2.620255e+00  1.520740e+00  6.915155e-01  1.387779e-15
#>  [71,]  1.867855e+00  3.990248e-15  5.626928e-16 -1.575700e-16
#>  [72,]  9.024999e-01  8.858646e-16  8.475954e-17 -2.343920e-15
#>  [73,]  1.034066e+00  4.923665e-15  2.004786e-15  2.381424e-16
#>  [74,]  2.608598e+00  5.075522e-01  1.398403e-15 -9.718736e-16
#>  [75,]  7.473052e-15  4.089543e-15  2.288203e-15 -8.789834e-16
#>  [76,]  4.407719e+00  7.984176e-15 -4.784679e-17 -1.597946e-15
#>  [77,]  1.191397e+00  2.197989e-15 -4.215519e-16 -6.971431e-16
#>  [78,]  6.443970e-15  1.767488e-15 -1.977427e-16 -3.962185e-16
#>  [79,]  1.679598e+00  4.181914e-01  2.951969e-15 -1.704533e-16
#>  [80,]  9.033062e-01  3.889257e-15  1.149405e-15 -8.821499e-17
#>  [81,]  3.148464e+00  8.994341e-01  2.880829e-15  2.477073e-15
#>  [82,]  3.226829e+00  1.462128e+00  5.463485e-01  7.781043e-15
#>  [83,]  1.221517e+00  1.060203e-14 -4.017861e-16 -6.544528e-16
#>  [84,]  2.867893e+00  1.074413e+00  2.272325e-15  2.049659e-15
#>  [85,]  1.370394e+00  2.606821e-15  2.007477e-15 -1.712259e-16
#>  [86,]  1.209750e+00  1.176561e-14  4.900595e-15 -5.846939e-16
#>  [87,]  1.858383e+00  8.263248e-15  1.304997e-15  1.002263e-15
#>  [88,]  3.190783e+00  5.460371e-15 -3.138727e-16 -7.359982e-16
#>  [89,]  4.857085e+00  1.290751e+00  7.519682e-15 -8.176010e-16
#>  [90,] -1.515533e-15 -1.666016e-15 -2.553975e-15 -2.821414e-15
#>  [91,]  7.016622e-01  2.179043e-15 -2.595214e-16 -2.516495e-15
#>  [92,]  3.092308e+00  3.226483e-01  2.886580e-15  6.661338e-16
#>  [93,]  2.796133e+00  2.028792e+00 -4.708318e-16 -5.283689e-16
#>  [94,]  3.465187e+00  6.105965e-15  2.771840e-16 -1.748352e-16
#>  [95,]  3.423789e+00  7.033515e-01  4.394353e-15  1.964743e-15
#>  [96,]  2.470583e+00  1.037754e+00  2.157109e-15  6.156872e-16
#>  [97,]  1.701804e+00  4.707874e-01  6.322998e-15  5.072645e-16
#>  [98,]  3.240098e+00  1.505373e+00  3.137922e-01  6.549532e-17
#>  [99,]  1.748084e+00  1.962271e-16 -1.474277e-16 -6.795565e-16
#> [100,]  2.275227e+00  8.357535e-01 -3.249040e-16 -6.812356e-16
#> 
#> $BootMatEigsCI
#>                     LowCI       HighCI
#> Dimension 1  3.411208e+01 7.995488e+01
#> Dimension 2  1.366719e+01 4.141510e+01
#> Dimension 3  8.198706e+00 1.704383e+01
#> Dimension 4  4.145544e+00 1.023963e+01
#> Dimension 5  1.075257e+00 6.699957e+00
#> Dimension 6  1.164490e-15 4.407719e+00
#> Dimension 7 -3.070439e-16 2.796586e+00
#> Dimension 8 -1.881651e-15 7.366786e-01
#> Dimension 9 -3.085355e-15 6.228014e-15
#> 
#> $FixedEigs
#> [1] 46.8262426 23.6659120  9.9143065  8.7560277  5.9582344  4.6749401  2.7978786
#> [8]  2.2867337  0.4116893
#> 
#> $BootstrappedEstEigs
#>  Dimension 1  Dimension 2  Dimension 3  Dimension 4  Dimension 5  Dimension 6 
#> 54.776515327 23.081403242 11.686376656  7.500946252  4.437651705  2.168678466 
#>  Dimension 7  Dimension 8  Dimension 9 
#>  0.653735173  0.092316750  0.001896968 
#> 
#> $BootMatTaus
#>        Dimension 1 Dimension 2 Dimension 3 Dimension 4 Dimension 5
#>   [1,]   0.4557966   0.2648457  0.09075204  0.07077103 0.052797356
#>   [2,]   0.5681012   0.2135816  0.09735075  0.06468466 0.047881186
#>   [3,]   0.5518240   0.2355010  0.09638000  0.05273078 0.038413525
#>   [4,]   0.4966766   0.1748191  0.12885892  0.09183563 0.051618063
#>   [5,]   0.4783322   0.1892472  0.12097485  0.08640661 0.057023514
#>   [6,]   0.5343783   0.2168888  0.17552617  0.05081314 0.022393646
#>   [7,]   0.4400852   0.2712872  0.09994359  0.07224026 0.057714587
#>   [8,]   0.4051753   0.3558830  0.14521054  0.07864177 0.015089363
#>   [9,]   0.4346068   0.1995778  0.15249032  0.08644575 0.054300934
#>  [10,]   0.4961212   0.2044283  0.13996372  0.07760627 0.051793822
#>  [11,]   0.4692652   0.2244787  0.11088882  0.08755516 0.057904734
#>  [12,]   0.4830565   0.1810495  0.14174909  0.08210376 0.062913500
#>  [13,]   0.5386221   0.1860191  0.08061399  0.07211642 0.059593263
#>  [14,]   0.5339651   0.1997852  0.11927253  0.09025867 0.046875046
#>  [15,]   0.4354681   0.2842209  0.15210844  0.06215911 0.041106702
#>  [16,]   0.5534455   0.1519849  0.13325950  0.06322753 0.057670248
#>  [17,]   0.4699589   0.2352379  0.09961292  0.08424986 0.062230892
#>  [18,]   0.4639293   0.2474224  0.09624141  0.08080282 0.074811766
#>  [19,]   0.4547051   0.2256720  0.10430576  0.09319981 0.051323713
#>  [20,]   0.5296959   0.1670903  0.09652715  0.09450245 0.058948589
#>  [21,]   0.4862533   0.3248397  0.07335257  0.05659057 0.040282662
#>  [22,]   0.5580411   0.1820112  0.10531484  0.09010022 0.041341413
#>  [23,]   0.6886326   0.1378415  0.11723292  0.03143899 0.024854017
#>  [24,]   0.4891400   0.1825091  0.09999079  0.09098454 0.060297142
#>  [25,]   0.4557317   0.1996599  0.13952958  0.08691277 0.057782366
#>  [26,]   0.5763846   0.2874255  0.08896308  0.04008403 0.007142795
#>  [27,]   0.5209497   0.2399809  0.14606880  0.05375981 0.029983874
#>  [28,]   0.3908262   0.2239901  0.15568034  0.10429624 0.070801652
#>  [29,]   0.5419231   0.2147355  0.13408216  0.05480790 0.037447781
#>  [30,]   0.6048633   0.2342750  0.07736943  0.03974982 0.026311328
#>  [31,]   0.5263900   0.1928497  0.17217588  0.06898766 0.039596784
#>  [32,]   0.5068249   0.1884438  0.14650066  0.07672977 0.052178598
#>  [33,]   0.5313092   0.2035072  0.10960614  0.10059116 0.034754564
#>  [34,]   0.4825089   0.2220961  0.09709499  0.06963732 0.053568285
#>  [35,]   0.5415003   0.2276639  0.10063176  0.07629366 0.048126496
#>  [36,]   0.4843020   0.2464000  0.10330999  0.09202914 0.060216146
#>  [37,]   0.5317396   0.1698938  0.10984639  0.08868163 0.054158158
#>  [38,]   0.5820420   0.1687905  0.10038409  0.06594037 0.041279184
#>  [39,]   0.5997846   0.2162577  0.09995435  0.05309806 0.019570986
#>  [40,]   0.4308153   0.2596515  0.13650357  0.09643030 0.045056092
#>  [41,]   0.5179458   0.1850588  0.09616494  0.09144358 0.059551817
#>  [42,]   0.5553305   0.1847792  0.09747468  0.07139402 0.049678770
#>  [43,]   0.6240576   0.1628400  0.08575358  0.07056560 0.039635157
#>  [44,]   0.5738173   0.2023298  0.08655393  0.07833676 0.039173961
#>  [45,]   0.4436102   0.2133652  0.14520913  0.09858495 0.067170114
#>  [46,]   0.5188760   0.3275669  0.06195092  0.04547956 0.027379329
#>  [47,]   0.4694162   0.2110638  0.13265848  0.09468741 0.062078259
#>  [48,]   0.4904056   0.1844444  0.12022161  0.08320301 0.055727363
#>  [49,]   0.4935100   0.2870558  0.09213943  0.05674316 0.041434758
#>  [50,]   0.5484285   0.2057593  0.11785537  0.06196281 0.032702017
#>  [51,]   0.5395621   0.2434765  0.11632205  0.06992610 0.026601669
#>  [52,]   0.5598803   0.2159889  0.13662472  0.07566332 0.011842729
#>  [53,]   0.6001843   0.1967233  0.08324522  0.05464698 0.042068995
#>  [54,]   0.4675881   0.2188949  0.10676395  0.07974998 0.066814754
#>  [55,]   0.5074598   0.2551009  0.12721265  0.05392988 0.044059571
#>  [56,]   0.4882930   0.2641365  0.08913771  0.07762474 0.043389810
#>  [57,]   0.5328206   0.1844601  0.13448405  0.10259131 0.045643896
#>  [58,]   0.4898185   0.1627800  0.15801653  0.09003023 0.059897523
#>  [59,]   0.6385836   0.2597600  0.05315984  0.03119945 0.017297160
#>  [60,]   0.4249464   0.3055127  0.12731228  0.08635485 0.039927046
#>  [61,]   0.4828729   0.1928484  0.13679559  0.08532196 0.060448211
#>  [62,]   0.6121714   0.2285707  0.06896133  0.04758827 0.029747083
#>  [63,]   0.5419133   0.1733529  0.12487100  0.09012170 0.043245390
#>  [64,]   0.5334934   0.2599152  0.10421691  0.06244260 0.021234054
#>  [65,]   0.5600332   0.2323308  0.12077961  0.05727417 0.020391336
#>  [66,]   0.5495490   0.2434501  0.08748078  0.06332213 0.048298785
#>  [67,]   0.5570665   0.1933726  0.09851789  0.07471645 0.036923061
#>  [68,]   0.4996378   0.1745860  0.13310610  0.09464888 0.053029801
#>  [69,]   0.4193154   0.3239094  0.10700471  0.08104552 0.051298237
#>  [70,]   0.5213735   0.1764319  0.10484844  0.08971794 0.059861006
#>  [71,]   0.4817940   0.1931614  0.18593265  0.06059534 0.053049337
#>  [72,]   0.5654525   0.1885689  0.14311957  0.07339678 0.019129435
#>  [73,]   0.5443819   0.2290369  0.10682283  0.06609893 0.040866199
#>  [74,]   0.4750210   0.2481895  0.10165944  0.08983898 0.054390621
#>  [75,]   0.5716206   0.2488002  0.12129517  0.03981408 0.018469984
#>  [76,]   0.5761917   0.1994330  0.08799229  0.05872116 0.042990450
#>  [77,]   0.6581904   0.1388288  0.10699359  0.04851002 0.036893810
#>  [78,]   0.4985802   0.3043980  0.08727990  0.06114877 0.048593191
#>  [79,]   0.5213881   0.1957130  0.14554731  0.08714864 0.032288663
#>  [80,]   0.4480175   0.2916533  0.13692965  0.08487876 0.030899776
#>  [81,]   0.5351612   0.1799247  0.11363472  0.07927073 0.053342504
#>  [82,]   0.5315442   0.2041990  0.09007118  0.07078242 0.048289417
#>  [83,]   0.5749454   0.2056442  0.11876132  0.05425530 0.035181742
#>  [84,]   0.4519108   0.2272787  0.13600963  0.08495413 0.052893954
#>  [85,]   0.6276654   0.1715165  0.09459738  0.06867296 0.022905057
#>  [86,]   0.4933569   0.2899038  0.08521498  0.08227057 0.038494985
#>  [87,]   0.5754019   0.2283211  0.08947745  0.06479433 0.025097063
#>  [88,]   0.5345494   0.2279190  0.10939540  0.06615794 0.037787783
#>  [89,]   0.4769595   0.2049288  0.13421702  0.07941321 0.047274784
#>  [90,]   0.5682455   0.1869878  0.13760749  0.08650006 0.020659145
#>  [91,]   0.5695749   0.2205116  0.11841910  0.06845754 0.017718821
#>  [92,]   0.4792508   0.2082513  0.14019546  0.09490068 0.045988058
#>  [93,]   0.5816831   0.1451458  0.12603455  0.05582510 0.047402582
#>  [94,]   0.5802313   0.1979539  0.09216124  0.05799695 0.040525129
#>  [95,]   0.4963237   0.2245002  0.11913043  0.07434714 0.045048494
#>  [96,]   0.5109514   0.2355100  0.11778737  0.07147618 0.037535282
#>  [97,]   0.5239340   0.1689752  0.11671889  0.09997768 0.066081887
#>  [98,]   0.5181888   0.2596368  0.07613369  0.06352036 0.043254088
#>  [99,]   0.5674164   0.2202236  0.08235034  0.06107492 0.053900391
#> [100,]   0.5359360   0.1921492  0.12498974  0.07357186 0.041843359
#>          Dimension 6   Dimension 7   Dimension 8   Dimension 9
#>   [1,]  3.273144e-02  2.588676e-02  6.419054e-03  6.060224e-18
#>   [2,]  8.400604e-03  7.300549e-17  1.083259e-17  7.150532e-18
#>   [3,]  1.936908e-02  5.781574e-03  1.748965e-17 -1.252289e-17
#>   [4,]  4.440612e-02  1.178551e-02  3.044482e-17  8.995361e-19
#>   [5,]  4.053687e-02  2.747876e-02  1.608776e-17  1.254398e-19
#>   [6,]  1.110732e-16  1.564355e-17 -2.435947e-18 -6.359504e-18
#>   [7,]  3.091422e-02  2.269703e-02  5.117836e-03  2.767013e-19
#>   [8,]  2.222908e-17  4.162285e-18  3.841522e-18  1.461072e-19
#>   [9,]  3.764105e-02  2.671161e-02  8.225718e-03 -4.958679e-18
#>  [10,]  2.596914e-02  4.117622e-03 -5.258819e-18 -2.189831e-17
#>  [11,]  3.863110e-02  1.127625e-02  2.881899e-17  9.692077e-19
#>  [12,]  3.552789e-02  1.359984e-02  3.264870e-17  1.764403e-18
#>  [13,]  3.482147e-02  2.334328e-02  4.870435e-03  6.123711e-17
#>  [14,]  9.843423e-03  9.818743e-18  9.943798e-19 -7.221572e-18
#>  [15,]  2.493672e-02  6.646446e-18  5.916527e-18 -1.890477e-17
#>  [16,]  2.939650e-02  1.101583e-02  2.401580e-17 -1.001974e-18
#>  [17,]  3.147290e-02  1.341051e-02  3.826119e-03 -1.063606e-18
#>  [18,]  2.885892e-02  7.933311e-03  2.292402e-17  6.039752e-18
#>  [19,]  3.515766e-02  2.739913e-02  6.378276e-03  1.858526e-03
#>  [20,]  3.375579e-02  1.503454e-02  4.445360e-03 -5.640980e-18
#>  [21,]  1.419301e-02  4.488164e-03  3.278242e-17  3.209939e-17
#>  [22,]  1.646233e-02  6.728956e-03  1.097450e-17 -4.880661e-18
#>  [23,]  2.346307e-17  2.026909e-17 -4.440229e-18 -7.502442e-18
#>  [24,]  3.793300e-02  3.338926e-02  5.756202e-03  4.217283e-17
#>  [25,]  3.132621e-02  2.349882e-02  5.558695e-03  5.883504e-17
#>  [26,]  8.539586e-18 -2.251652e-18 -1.379876e-17 -2.262591e-17
#>  [27,]  9.256882e-03  9.421934e-17  1.816309e-17 -2.620062e-18
#>  [28,]  4.540839e-02  8.997075e-03  3.989557e-17  2.641762e-17
#>  [29,]  1.700346e-02  3.716432e-17 -3.259929e-18 -3.555083e-17
#>  [30,]  1.743111e-02  1.158160e-16  1.017923e-17 -3.027951e-18
#>  [31,]  5.023716e-17  2.934085e-17  1.006250e-17 -1.584809e-17
#>  [32,]  2.561826e-02  3.703999e-03  5.860769e-17  7.468887e-19
#>  [33,]  2.023169e-02  1.766013e-17 -1.270672e-18 -3.823628e-17
#>  [34,]  3.848412e-02  2.498191e-02  1.162834e-02  1.534924e-17
#>  [35,]  5.783945e-03  8.165218e-17  7.377200e-18  5.772585e-18
#>  [36,]  1.374265e-02  3.100270e-17  1.038431e-17 -2.196926e-17
#>  [37,]  3.921825e-02  6.462165e-03  8.271433e-17 -8.579009e-18
#>  [38,]  2.439417e-02  1.716966e-02  4.097973e-17 -2.003590e-18
#>  [39,]  1.133426e-02  1.977626e-17  4.595126e-18 -1.259398e-17
#>  [40,]  2.732597e-02  4.217302e-03  2.077336e-17 -1.212436e-17
#>  [41,]  3.669457e-02  1.314046e-02  1.505776e-17  6.478239e-18
#>  [42,]  4.134279e-02  9.455217e-18  6.922823e-18 -1.844779e-17
#>  [43,]  1.714809e-02  5.502058e-17  1.796108e-17  8.571020e-18
#>  [44,]  1.978829e-02  5.243308e-17  2.979102e-17 -1.122223e-17
#>  [45,]  3.206042e-02  6.274417e-17  6.904077e-18 -9.361080e-18
#>  [46,]  1.427798e-02  4.469253e-03  3.452355e-17 -3.156927e-19
#>  [47,]  3.009587e-02  7.763045e-17  2.235190e-17  1.881237e-18
#>  [48,]  3.237306e-02  2.325441e-02  1.037064e-02  3.912592e-17
#>  [49,]  2.255739e-02  4.033369e-03  2.526122e-03 -2.093323e-17
#>  [50,]  2.369913e-02  9.592861e-03  9.548666e-17  7.154159e-18
#>  [51,]  4.111533e-03  1.050494e-17 -5.925316e-18 -6.794833e-18
#>  [52,]  4.618370e-17  1.157423e-17  6.211180e-18  8.131100e-19
#>  [53,]  1.921698e-02  3.914254e-03  9.158807e-18 -4.628864e-17
#>  [54,]  3.588530e-02  2.182537e-02  2.477724e-03  2.813067e-17
#>  [55,]  1.223718e-02  3.492570e-17  1.106971e-17  7.814218e-19
#>  [56,]  1.962753e-02  1.779068e-02  2.105432e-17  1.005136e-17
#>  [57,]  2.557385e-17  1.598036e-17 -7.458590e-18 -1.996009e-17
#>  [58,]  3.945723e-02  2.832095e-17  1.220251e-17  1.073654e-17
#>  [59,]  2.473759e-17  1.427206e-17 -4.353949e-18 -2.014012e-17
#>  [60,]  1.594673e-02  6.941227e-18 -6.397732e-18 -1.212481e-17
#>  [61,]  3.551041e-02  6.202601e-03  6.882345e-18 -4.270068e-20
#>  [62,]  1.010933e-02  2.851892e-03 -3.135611e-18 -1.158489e-17
#>  [63,]  2.649566e-02  4.775157e-17  3.275313e-17  3.692692e-18
#>  [64,]  1.869788e-02  1.069341e-16  1.501366e-17  3.386725e-18
#>  [65,]  9.190813e-03  1.605963e-16  2.626036e-17  4.252352e-19
#>  [66,]  7.899136e-03  6.505022e-17  1.400626e-17 -1.855202e-17
#>  [67,]  2.812956e-02  1.127400e-02  2.005932e-17  1.887165e-17
#>  [68,]  3.192640e-02  1.306495e-02  9.085025e-17  1.655507e-17
#>  [69,]  1.742675e-02  5.145043e-17  1.059399e-17  5.464750e-18
#>  [70,]  2.590003e-02  1.503182e-02  6.835317e-03  1.371756e-17
#>  [71,]  2.546726e-02  5.440501e-17  7.672031e-18 -2.148387e-18
#>  [72,]  1.033277e-02  1.014231e-17  9.704165e-19 -2.683566e-17
#>  [73,]  1.279329e-02  6.091470e-17  2.480286e-17  2.946255e-18
#>  [74,]  2.586744e-02  5.032999e-03  1.386688e-17 -9.637313e-18
#>  [75,]  5.936998e-17  3.248954e-17  1.817873e-17 -6.983121e-18
#>  [76,]  3.467136e-02  6.280396e-17 -3.763655e-19 -1.256953e-17
#>  [77,]  1.058330e-02  1.952496e-17 -3.744689e-18 -6.192795e-18
#>  [78,]  6.225503e-17  1.707566e-17 -1.910387e-18 -3.827857e-18
#>  [79,]  1.434308e-02  3.571183e-03  2.520860e-17 -1.455601e-18
#>  [80,]  7.620951e-03  3.281261e-17  9.697221e-18 -7.442461e-19
#>  [81,]  3.007465e-02  8.591542e-03  2.751815e-17  2.366141e-17
#>  [82,]  3.396986e-02  1.539230e-02  5.751587e-03  8.191354e-17
#>  [83,]  1.121207e-02  9.731397e-17 -3.687918e-18 -6.007097e-18
#>  [84,]  3.415659e-02  1.279626e-02  2.706338e-17  2.441142e-17
#>  [85,]  1.464273e-02  2.785402e-17  2.145000e-17 -1.829558e-18
#>  [86,]  1.075870e-02  1.046354e-16  4.358258e-17 -5.199873e-18
#>  [87,]  1.690818e-02  7.518174e-17  1.187329e-17  9.118920e-18
#>  [88,]  2.419049e-02  4.139707e-17 -2.379584e-18 -5.579872e-18
#>  [89,]  4.519597e-02  1.201065e-02  6.997185e-17 -7.607909e-18
#>  [90,] -1.321109e-17 -1.452286e-17 -2.226331e-17 -2.459460e-17
#>  [91,]  5.318030e-03  1.651538e-17 -1.966962e-18 -1.907299e-17
#>  [92,]  2.844568e-02  2.967993e-03  2.655321e-17  6.127665e-18
#>  [93,]  2.544597e-02  1.846285e-02 -4.284765e-18 -4.808376e-18
#>  [94,]  3.113151e-02  5.485647e-17  2.490243e-18 -1.570733e-18
#>  [95,]  3.372237e-02  6.927613e-03  4.328188e-17  1.935160e-17
#>  [96,]  1.883022e-02  7.909524e-03  1.644100e-17  4.692628e-18
#>  [97,]  1.904398e-02  5.268331e-03  7.075730e-17  5.676526e-18
#>  [98,]  2.514722e-02  1.168358e-02  2.435421e-03  5.083257e-19
#>  [99,]  1.503439e-02  1.687650e-18 -1.267951e-18 -5.844521e-18
#> [100,]  2.304481e-02  8.464995e-03 -3.290816e-18 -6.899948e-18
#> 
#> $BootMatTausCI
#>                     LowCI       HighCI
#> Dimension 1  4.051753e-01 6.385836e-01
#> Dimension 2  1.388288e-01 3.248397e-01
#> Dimension 3  6.195092e-02 1.721759e-01
#> Dimension 4  3.143899e-02 1.005912e-01
#> Dimension 5  1.184273e-02 6.717011e-02
#> Dimension 6  8.539586e-18 4.440612e-02
#> Dimension 7 -2.251652e-18 2.739913e-02
#> Dimension 8 -1.379876e-17 8.225718e-03
#> Dimension 9 -3.823628e-17 6.123711e-17
#> 
#> $BootstrappedEstInertia
#>   [1] 104.51631  99.65159 113.30106  81.44111 109.82270 110.10062 100.30879
#>   [8] 116.37279  89.55798 121.11133  89.15423  77.90631 101.70326  85.00821
#>  [15]  87.25623  97.08747 107.13032 103.10779 102.06841  97.65758 136.65864
#>  [22]  96.09349 103.00413  87.15311 102.46495 136.36379 102.46405  87.28196
#>  [29]  83.36966 132.18668  80.33068 118.54652  80.14161  87.51568  96.00337
#>  [36] 111.26512  86.58004 125.41473 125.72751  91.45960  92.36286  88.44066
#>  [43] 107.09734 119.54014  66.67902 132.70882  76.46205 104.66395 128.96989
#>  [50]  94.78073 115.73043  90.79466 113.62421 108.39524 112.41694 114.27361
#>  [57]  96.44410  72.35577 132.87234  87.75612  85.21892 145.89093 114.08621
#>  [64] 124.92972 106.85263 110.94023 105.91079  78.28339 108.53535 101.16802
#>  [71]  73.34339  87.34347  80.82884 100.84487 125.87257 127.12855 112.57326
#>  [78] 103.50923 117.10166 118.52933 104.68831  94.99092 108.94660  83.96309
#>  [85]  93.58866 112.44389 109.91030 131.90235 107.46724 114.71678 131.94023
#>  [92] 108.70925 109.88510 111.30803 101.52870 131.20307  89.36177 128.84519
#>  [99] 116.27239  98.73054
#>