Boostrap for eigenvalues.

Boostrap for eigenvalues.

Usage

Boot4Eigs(
  data,
  center = FALSE,
  scale = FALSE,
  design = NULL,
  niter = 100,
  CI.perc = c(0.025, 0.975),
  suppressProgressBar = TRUE
)

Arguments

data

the data-set;

center

whether to center the data (default to FALSE), see expo.scale;

scale

whether to scale the data ( default to FALSE), see expo.scale;

design

whether there is a design;

niter

number of bootstrap iterations;

CI.perc

vector of the two levels of confidence for the left and right limits of the bootstrap confidence interval (default to c(0.025, 0.0975));

suppressProgressBar

whether to show a progress bar or not (default to TRUE).

Value

list containing four objects

Examples

Boot4Eigs(matrix(rnorm(100), 10, 10))
#> $BootMatEigs
#>        Dimension 1 Dimension 2 Dimension 3  Dimension 4  Dimension 5
#>   [1,]    73.39629    14.83223   10.640610 6.094398e+00 3.632523e+00
#>   [2,]    32.73571    12.63783   12.010823 6.031125e+00 2.759980e+00
#>   [3,]    58.99252    22.72974   12.359635 6.766930e+00 3.810466e+00
#>   [4,]    33.76563    28.39962   11.266960 6.073526e+00 5.109349e+00
#>   [5,]    53.99338    16.94494   13.957092 8.709248e+00 6.643681e+00
#>   [6,]    43.48507    30.48161   14.031089 1.035734e+01 3.685273e+00
#>   [7,]    39.37857    16.57186   14.625886 7.745253e+00 5.645787e+00
#>   [8,]    75.18004    18.70508   13.814966 7.906105e+00 4.590680e+00
#>   [9,]    37.21685    24.26445   16.212616 6.095134e+00 3.974241e+00
#>  [10,]    38.37189    16.54959   12.166007 7.969896e+00 1.889502e+00
#>  [11,]    59.56085    22.37232   15.442197 2.302233e+00 1.626971e+00
#>  [12,]    46.50936    28.07309   25.718649 7.100738e+00 4.430732e+00
#>  [13,]    40.50145    19.30201   14.629165 9.902949e+00 4.680941e+00
#>  [14,]    68.63563    18.06296   13.451424 6.487460e+00 2.984711e+00
#>  [15,]    64.49644    22.59384    9.904911 5.535776e+00 3.059203e+00
#>  [16,]    47.98821    20.06425   15.920756 9.717490e+00 6.586153e+00
#>  [17,]    39.80104    28.44205   12.539537 8.990930e+00 5.069240e+00
#>  [18,]    61.83495    18.07360   13.597266 9.781119e+00 6.513137e+00
#>  [19,]    69.55991    15.03436   10.163754 7.960754e+00 4.750059e+00
#>  [20,]    49.47059    16.83741   11.475117 5.794915e+00 5.414303e+00
#>  [21,]    33.89448    17.47738   13.841233 9.481058e+00 5.199747e+00
#>  [22,]    55.74463    28.47188   20.601129 1.061393e+01 5.933417e+00
#>  [23,]    55.39117    34.95899   15.072709 1.240834e+01 2.108183e+00
#>  [24,]    47.32741    17.13946   13.008715 9.691536e+00 6.762083e+00
#>  [25,]    44.47776    16.25462   11.477842 9.010796e+00 6.589532e+00
#>  [26,]    67.04145    27.19799   20.201500 7.090205e+00 5.641832e+00
#>  [27,]    29.78688    19.50294   16.350002 9.805779e+00 8.146490e+00
#>  [28,]    40.53849    27.50257   22.730115 7.905058e+00 4.901898e+00
#>  [29,]    46.88333    21.57771   16.179341 1.062146e+01 7.153306e+00
#>  [30,]    52.53651    15.46953   12.652602 9.125780e+00 4.519671e+00
#>  [31,]    43.63025    27.83625   19.872861 8.094728e+00 4.664407e+00
#>  [32,]    40.23639    18.36601    9.446819 6.995712e+00 3.727272e+00
#>  [33,]    25.71995    20.23745   15.730073 1.029185e+01 6.390250e+00
#>  [34,]    69.92336    22.55275   14.537615 8.884772e+00 4.872173e+00
#>  [35,]    63.03012    17.36510   13.358487 1.171016e+01 3.487197e+00
#>  [36,]    55.23587    17.51475   16.902741 9.739388e+00 6.398871e+00
#>  [37,]    34.52021    21.04377    9.561734 6.506098e+00 3.842092e+00
#>  [38,]    45.16835    22.62274   13.213106 1.067722e+01 4.109357e+00
#>  [39,]    55.00015    17.15763   14.466676 8.863387e+00 5.732427e+00
#>  [40,]    43.52687    29.88623   17.030612 8.505557e+00 3.990968e+00
#>  [41,]    54.91202    23.15297   11.121376 5.409234e+00 3.965488e+00
#>  [42,]    34.42696    26.97837   23.992468 8.956745e+00 4.005496e+00
#>  [43,]    67.50555    25.11556   12.132436 5.168671e+00 4.214070e+00
#>  [44,]    94.94870    15.82515   13.884079 8.891921e+00 4.285547e+00
#>  [45,]    52.18536    27.74465   20.423674 1.023865e+01 5.077342e+00
#>  [46,]    42.15853    29.86838   16.108447 8.763867e+00 6.772446e+00
#>  [47,]    41.36301    16.93748   13.265447 7.328158e+00 3.832160e+00
#>  [48,]    59.19421    19.91724    9.262551 2.551520e+00 1.798341e+00
#>  [49,]    52.40769    14.64434   13.375589 8.295421e+00 4.694763e+00
#>  [50,]    42.09354    28.39829   17.162206 7.371197e+00 4.854419e+00
#>  [51,]    44.71666    32.82389   22.677041 6.842587e+00 3.981566e+00
#>  [52,]    52.17128    24.23559   13.706900 7.556911e+00 1.757565e+00
#>  [53,]    46.72830    33.45627   14.202651 1.090900e+01 6.067854e+00
#>  [54,]    43.10412    22.52998    6.815550 4.237286e-15 2.989164e-15
#>  [55,]    35.05487    26.64725   10.808031 6.216402e+00 2.890443e+00
#>  [56,]    36.16290    29.39303   16.195547 9.431157e+00 8.148522e+00
#>  [57,]    40.61275    21.83712   14.262048 9.858383e+00 5.265830e+00
#>  [58,]    68.71705    31.86769   13.373173 1.043783e+01 4.938142e+00
#>  [59,]    65.52667    21.15854   12.535667 8.402479e+00 1.892652e+00
#>  [60,]    43.47235    19.63898   15.248776 9.181974e+00 4.561511e+00
#>  [61,]    29.43158    12.03790   11.120265 3.629816e+00 3.209777e+00
#>  [62,]    49.51363    14.98658   14.657787 1.120764e+01 4.624666e+00
#>  [63,]    40.85065    28.94891   15.428569 7.716067e+00 4.001421e+00
#>  [64,]    37.12184    15.89130   10.214324 5.961151e+00 3.782180e+00
#>  [65,]    65.53967    24.06807   10.332673 6.442467e+00 3.613487e+00
#>  [66,]    60.11884    25.54349   10.150839 7.107273e+00 7.225921e-15
#>  [67,]    61.28652    27.78591   10.049931 5.749732e+00 5.251705e+00
#>  [68,]    40.30985    29.87270   15.160319 9.379566e+00 6.064670e+00
#>  [69,]    48.46233    29.31681   14.765984 7.374784e+00 5.204136e+00
#>  [70,]    59.95996    21.89721   12.863687 5.926840e+00 3.236325e+00
#>  [71,]    36.43498    18.96109   14.091369 7.333012e+00 4.587542e+00
#>  [72,]    43.88502    15.80085   12.418961 9.231658e+00 6.156024e+00
#>  [73,]    74.25747    22.67742   13.802409 1.240695e+01 3.805572e+00
#>  [74,]    55.94327    28.18766   18.545801 1.058753e+01 3.912075e+00
#>  [75,]    37.91217    15.48193   13.245111 8.324515e+00 5.262252e+00
#>  [76,]    60.45502    20.00406    8.924777 6.117314e+00 4.061360e+00
#>  [77,]    45.41364    16.18932   13.232363 1.142397e+01 6.123193e+00
#>  [78,]    45.00458    28.90294   12.936630 5.282059e+00 4.556547e+00
#>  [79,]    44.48802    18.66860    8.397193 5.728064e+00 3.109670e+00
#>  [80,]    61.71304    16.40483    9.082483 4.861921e+00 4.003882e-15
#>  [81,]    44.85900    19.06396   12.296684 8.617334e+00 3.640466e+00
#>  [82,]    37.07951    26.15851   15.765901 9.037858e+00 3.145603e+00
#>  [83,]    85.21072    23.85034    6.925342 4.972451e+00 1.924480e+00
#>  [84,]    46.97747    20.10621   12.810908 7.735169e+00 4.029672e+00
#>  [85,]    45.27146    22.74267   11.067230 4.861609e+00 3.675273e+00
#>  [86,]    59.70341    21.53073    8.429779 5.711823e+00 5.434634e+00
#>  [87,]    40.56580    17.79626   10.698288 1.002048e+01 3.309607e+00
#>  [88,]    40.53778    32.58887   11.459431 5.255389e+00 3.404603e+00
#>  [89,]    41.96807    28.65261   13.304268 8.520069e+00 5.606746e+00
#>  [90,]    41.77748    13.63462    9.342313 7.493258e+00 4.464819e+00
#>  [91,]    80.13791    18.20182   13.000602 7.793234e+00 5.116029e+00
#>  [92,]    46.60888    25.41747   14.850952 1.069060e+01 4.960164e+00
#>  [93,]    53.90863    17.73128   12.936266 1.082785e+01 7.283336e+00
#>  [94,]    74.86087    19.90788    9.114148 5.586225e+00 3.469236e+00
#>  [95,]    42.40699    30.79789   19.177779 8.598555e+00 3.551241e+00
#>  [96,]    70.84455    18.16820   14.610379 6.626172e+00 4.542700e+00
#>  [97,]    76.00361    17.67514   12.222877 4.489129e+00 3.321429e+00
#>  [98,]    55.78721    24.32585   14.900687 9.831820e+00 5.043339e+00
#>  [99,]    52.94959    28.16809   13.892534 5.265968e+00 4.924552e+00
#> [100,]    76.78810    19.40791   13.955230 4.324473e+00 1.567778e+00
#>         Dimension 6   Dimension 7   Dimension 8   Dimension 9
#>   [1,] 7.880711e-01  2.129468e-15  1.052125e-15 -5.818674e-17
#>   [2,] 1.670993e+00  1.262046e-15  7.996501e-16  5.227061e-17
#>   [3,] 1.776748e+00  3.955244e-01 -5.748249e-17 -6.225291e-16
#>   [4,] 3.948542e+00  1.967980e+00  7.560917e-01  1.784678e-15
#>   [5,] 2.261519e+00  9.487578e-01  8.090793e-03  3.896189e-15
#>   [6,] 1.404304e-15 -2.150176e-15 -2.637084e-15 -2.645348e-15
#>   [7,] 4.046417e+00  2.066641e+00  1.927793e-03 -1.480869e-16
#>   [8,] 2.437911e+00  2.833748e-15 -9.215144e-17 -3.489102e-16
#>   [9,] 9.379293e-01  7.591175e-15 -1.759905e-16 -3.958280e-16
#>  [10,] 9.764312e-01  3.153730e-15 -4.087696e-16 -1.150755e-15
#>  [11,] 5.836360e-15  1.429596e-15 -4.333416e-16 -1.106768e-15
#>  [12,] 2.440635e+00  1.832268e-15  6.864612e-16  2.727629e-17
#>  [13,] 3.115486e+00  1.134483e+00  8.918327e-01  1.696199e-15
#>  [14,] 8.614460e-01  6.174310e-01  4.698659e-16  4.030530e-16
#>  [15,] 6.754108e-01  5.806487e-15  2.347936e-15  1.941905e-15
#>  [16,] 2.919951e+00  1.450345e+00  5.380909e-01  7.953380e-15
#>  [17,] 1.670663e+00  1.391364e-15 -1.440423e-15 -1.656502e-15
#>  [18,] 4.953909e+00  1.754339e+00  5.436213e-01  1.393909e-15
#>  [19,] 2.510293e+00  1.115839e+00  1.968245e-15  9.553848e-17
#>  [20,] 2.723565e+00  1.198869e+00  2.234937e-15 -2.663577e-16
#>  [21,] 3.148534e+00  1.558695e+00  7.299840e-01  1.081948e-15
#>  [22,] 2.656082e+00  1.224846e-14  5.367758e-16  1.624899e-16
#>  [23,] 7.858239e-15  2.416810e-15  1.312796e-15  6.740738e-16
#>  [24,] 4.704778e+00  1.899736e+00  7.869262e-01  4.481397e-01
#>  [25,] 4.210191e+00  1.965172e+00  7.775310e-01  5.464195e-16
#>  [26,] 3.016907e+00  5.387706e-01 -5.335332e-17 -3.017107e-15
#>  [27,] 4.807084e+00  8.479708e-01  2.647386e-15 -3.416071e-17
#>  [28,] 3.437064e+00  1.563926e+00  3.578921e-15 -1.774105e-15
#>  [29,] 4.666805e+00  1.303834e+00  7.105307e-15  9.970233e-16
#>  [30,] 1.630592e+00  5.684584e-01  1.192461e-15  1.129714e-15
#>  [31,] 3.948657e+00  1.829404e+00  5.268061e-15  7.494889e-16
#>  [32,] 3.202744e+00  1.492626e+00  5.901303e-01  5.945621e-15
#>  [33,] 2.859414e+00  1.504460e+00  8.335185e-01 -4.679016e-16
#>  [34,] 3.233304e+00  1.687947e-01 -6.117698e-16 -3.648711e-15
#>  [35,] 1.771111e+00  2.621583e-15 -8.467788e-16 -1.255901e-15
#>  [36,] 4.534376e+00  2.045895e+00  4.837809e-01 -3.566783e-16
#>  [37,] 1.242785e+00  2.138427e-15  1.822523e-15 -4.284428e-16
#>  [38,] 1.893508e+00  5.404162e-01 -7.879955e-16 -9.419575e-16
#>  [39,] 3.059480e+00  1.184520e+00  7.011612e-01  4.292780e-15
#>  [40,] 2.943647e+00  1.234894e-01  1.069138e-14  1.295878e-15
#>  [41,] 1.800948e+00  8.833985e-01 -2.397782e-16 -1.115879e-15
#>  [42,] 2.926986e+00  1.418082e+00  3.489344e-15  2.402119e-15
#>  [43,] 1.799224e+00  5.145627e-01  4.255422e-15  1.276333e-15
#>  [44,] 1.174178e+00  3.121046e-15  1.176181e-15 -6.107389e-16
#>  [45,] 2.594780e+00  1.620333e+00  4.423371e-15 -5.886991e-16
#>  [46,] 1.796629e+00  1.668197e-15 -5.326198e-16 -3.720093e-15
#>  [47,] 1.291497e+00  2.057403e-15 -8.568816e-17 -1.228464e-15
#>  [48,] 6.324845e-15  1.614198e-15 -2.266627e-16 -4.591150e-16
#>  [49,] 2.864917e+00  1.106744e+00  4.695564e-15 -3.732426e-16
#>  [50,] 2.793044e-15 -4.381447e-16 -5.687821e-16 -6.396636e-16
#>  [51,] 2.983276e+00  2.032180e-15  7.421943e-16  3.360434e-16
#>  [52,] 2.156574e-01  7.832784e-15  1.615766e-15  6.475400e-16
#>  [53,] 9.864283e-01  3.089627e-15 -9.098501e-16 -1.535161e-15
#>  [54,] 8.714727e-16 -1.523165e-16 -5.478059e-16 -1.504631e-15
#>  [55,] 9.012119e-01  6.546516e-15 -2.763406e-16 -1.013738e-15
#>  [56,] 2.572369e+00  1.807311e-01  3.082482e-15  2.048830e-15
#>  [57,] 2.626811e+00  2.731697e-15 -5.424608e-17 -4.128850e-16
#>  [58,] 9.041441e-01  1.959776e-14 -3.306910e-16 -2.475104e-15
#>  [59,] 1.637316e+00  7.045848e-15  4.116545e-15  7.854058e-16
#>  [60,] 2.996927e+00  2.532478e+00  1.293402e-15  9.600848e-16
#>  [61,] 1.684938e+00  2.701627e-15  2.188699e-15  1.008408e-15
#>  [62,] 3.823111e+00  1.224512e+00  4.893083e-16 -1.963568e-15
#>  [63,] 3.374876e+00  8.342721e-01  1.188567e-01  2.692291e-15
#>  [64,] 2.595912e+00  3.355152e-15  1.674041e-15  4.036512e-16
#>  [65,] 1.806241e+00  6.052320e-01  8.151528e-15  1.179490e-15
#>  [66,] 2.080452e-15  4.718360e-16  9.215393e-18 -4.340743e-16
#>  [67,] 8.588237e-01  4.918952e-15  2.379643e-16 -4.334659e-17
#>  [68,] 3.228713e+00  1.574354e+00  2.501420e-15  8.289099e-16
#>  [69,] 2.566814e+00  2.232023e-01 -4.559369e-16 -1.426212e-15
#>  [70,] 7.373506e-01  1.665571e-16  1.213961e-16 -2.515338e-15
#>  [71,] 3.739929e+00  2.444081e+00  4.189417e-15 -7.264130e-16
#>  [72,] 1.922238e+00  5.704079e-15 -6.106535e-16 -6.213943e-16
#>  [73,] 5.383558e-15  3.252527e-16  1.292144e-16 -1.777562e-17
#>  [74,] 4.644377e-15  4.272961e-15 -6.883058e-16 -7.806035e-16
#>  [75,] 3.597615e+00  2.443273e+00  9.359523e-15  9.279739e-16
#>  [76,] 1.986384e+00  8.169439e-01  2.822870e-15 -4.703483e-16
#>  [77,] 4.672630e+00  2.367442e+00  1.399778e+00  4.850652e-15
#>  [78,] 3.098141e+00  6.371970e-01  3.037006e-15  6.433883e-16
#>  [79,] 1.316033e+00  1.546700e-15  1.538781e-15  1.063435e-15
#>  [80,] 2.746267e-15  2.553583e-16 -3.860170e-16 -7.506711e-16
#>  [81,] 8.721198e-01  2.000148e-15 -7.435666e-16 -9.812368e-16
#>  [82,] 1.676624e+00  2.865859e-15  1.899779e-15 -4.905397e-16
#>  [83,] 1.145263e+00  2.992371e-15  2.000583e-15 -3.284587e-16
#>  [84,] 8.764996e-01  3.304507e-15  4.961586e-16 -9.004225e-16
#>  [85,] 9.728075e-01  4.441719e-15  2.314319e-15  1.622471e-15
#>  [86,] 2.140663e+00  8.496270e-01  3.814529e-01 -3.478072e-16
#>  [87,] 1.658913e+00  1.696928e-15  1.472687e-15  5.146268e-16
#>  [88,] 7.661107e-01  8.535105e-15  2.693677e-15  3.134752e-16
#>  [89,] 3.210359e+00  1.766118e+00  2.561708e-15  1.527330e-15
#>  [90,] 1.690252e+00  8.612385e-01  7.676756e-15  1.229697e-15
#>  [91,] 3.203755e+00  1.134437e+00  1.914905e-15  3.960696e-16
#>  [92,] 3.234349e+00  2.325620e+00  2.975045e-15  7.943532e-16
#>  [93,] 4.772183e+00  2.124444e+00  4.626203e-01  5.296581e-15
#>  [94,] 8.571758e-01  5.708781e-01  3.457581e-15  6.845641e-16
#>  [95,] 1.719716e+00 -1.093872e-15 -1.239675e-15 -1.713160e-15
#>  [96,] 4.506077e-01  8.049446e-15  2.662526e-15  3.773268e-16
#>  [97,] 1.679112e+00  5.454402e-01  2.186409e-15 -6.264005e-16
#>  [98,] 1.311601e+00  8.552385e-01  3.365152e-15  2.250674e-16
#>  [99,] 1.713891e+00  1.288782e-14 -1.693914e-16 -8.011511e-16
#> [100,] 8.574070e-01  2.005420e-15  8.532715e-17 -2.346821e-15
#> 
#> $BootMatEigsCI
#>                     LowCI       HighCI
#> Dimension 1  2.943158e+01 8.013791e+01
#> Dimension 2  1.263783e+01 3.282389e+01
#> Dimension 3  6.925342e+00 2.273011e+01
#> Dimension 4  2.302233e+00 1.171016e+01
#> Dimension 5  4.003882e-15 7.283336e+00
#> Dimension 6  1.404304e-15 4.772183e+00
#> Dimension 7 -1.093872e-15 2.443273e+00
#> Dimension 8 -1.440423e-15 8.335185e-01
#> Dimension 9 -3.648711e-15 5.945621e-15
#> 
#> $FixedEigs
#> [1] 45.7560512 17.7380503 14.9816883  9.3523034  5.4919979  5.0110012  2.7533950
#> [8]  1.2811960  0.7280213
#> 
#> $BootstrappedEstEigs
#>  Dimension 1  Dimension 2  Dimension 3  Dimension 4  Dimension 5  Dimension 6 
#> 51.142995932 22.208332412 13.709011048  7.832100505  4.370105736  2.095754380 
#>  Dimension 7  Dimension 8  Dimension 9 
#>  0.642462569  0.100053949  0.004481397 
#> 
#> $BootMatTaus
#>        Dimension 1 Dimension 2 Dimension 3  Dimension 4  Dimension 5
#>   [1,]   0.6709958   0.1355976  0.09727747 5.571557e-02 3.320887e-02
#>   [2,]   0.4824969   0.1862711  0.17702947 8.889373e-02 4.067980e-02
#>   [3,]   0.5522012   0.2127624  0.11569273 6.334205e-02 3.566798e-02
#>   [4,]   0.3698815   0.3111002  0.12342255 6.653171e-02 5.596975e-02
#>   [5,]   0.5218430   0.1637719  0.13489451 8.417439e-02 6.421080e-02
#>   [6,]   0.4261555   0.2987211  0.13750526 1.015023e-01 3.611583e-02
#>   [7,]   0.4371397   0.1839635  0.16236129 8.597970e-02 6.267363e-02
#>   [8,]   0.6130401   0.1525267  0.11265128 6.446870e-02 3.743375e-02
#>   [9,]   0.4195754   0.2735526  0.18277784 6.871534e-02 4.480481e-02
#>  [10,]   0.4924314   0.2123830  0.15612796 1.022787e-01 2.424823e-02
#>  [11,]   0.5879384   0.2208422  0.15243337 2.272586e-02 1.606020e-02
#>  [12,]   0.4070015   0.2456664  0.22506281 6.213826e-02 3.877315e-02
#>  [13,]   0.4301420   0.2049953  0.15536774 1.051734e-01 4.971351e-02
#>  [14,]   0.6177766   0.1625813  0.12107377 5.839242e-02 2.686483e-02
#>  [15,]   0.6069363   0.2126168  0.09320901 5.209378e-02 2.878827e-02
#>  [16,]   0.4562257   0.1907516  0.15135922 9.238453e-02 6.261480e-02
#>  [17,]   0.4123885   0.2946952  0.12992527 9.315726e-02 5.252366e-02
#>  [18,]   0.5282693   0.1544066  0.11616438 8.356222e-02 5.564313e-02
#>  [19,]   0.6261302   0.1353289  0.09148708 7.165720e-02 4.275674e-02
#>  [20,]   0.5324298   0.1812135  0.12350154 6.236807e-02 5.827172e-02
#>  [21,]   0.3972113   0.2048184  0.16220618 1.111090e-01 6.093613e-02
#>  [22,]   0.4494771   0.2295729  0.16610992 8.558169e-02 4.784201e-02
#>  [23,]   0.4618263   0.2914722  0.12566939 1.034551e-01 1.757707e-02
#>  [24,]   0.4650484   0.1684157  0.12782619 9.523094e-02 6.644556e-02
#>  [25,]   0.4693557   0.1715284  0.12112098 9.508725e-02 6.953664e-02
#>  [26,]   0.5128290   0.2080492  0.15453001 5.423605e-02 4.315681e-02
#>  [27,]   0.3337572   0.2185274  0.18319916 1.098722e-01 9.128012e-02
#>  [28,]   0.3733543   0.2532952  0.20934149 7.280459e-02 4.514587e-02
#>  [29,]   0.4325598   0.1990824  0.14927550 9.799680e-02 6.599856e-02
#>  [30,]   0.5444021   0.1603008  0.13111078 9.456459e-02 4.683444e-02
#>  [31,]   0.3970842   0.2533411  0.18086535 7.367112e-02 4.245134e-02
#>  [32,]   0.4786758   0.2184929  0.11238493 8.322512e-02 4.434183e-02
#>  [33,]   0.3077765   0.2421705  0.18823317 1.231569e-01 7.646862e-02
#>  [34,]   0.5631135   0.1816239  0.11707571 7.155170e-02 3.923705e-02
#>  [35,]   0.5692638   0.1568348  0.12064871 1.057617e-01 3.149502e-02
#>  [36,]   0.4894381   0.1551960  0.14977307 8.629949e-02 5.669960e-02
#>  [37,]   0.4499700   0.2743050  0.12463696 8.480682e-02 5.008157e-02
#>  [38,]   0.4598472   0.2303162  0.13451918 1.087020e-01 4.183629e-02
#>  [39,]   0.5180608   0.1616122  0.13626540 8.348656e-02 5.399523e-02
#>  [40,]   0.4106023   0.2819260  0.16065498 8.023553e-02 3.764802e-02
#>  [41,]   0.5423654   0.2286816  0.10984570 5.342694e-02 3.916708e-02
#>  [42,]   0.3352020   0.2626780  0.23360542 8.720837e-02 3.899997e-02
#>  [43,]   0.5796952   0.2156766  0.10418574 4.438530e-02 3.618779e-02
#>  [44,]   0.6830371   0.1138422  0.09987858 6.396624e-02 3.082915e-02
#>  [45,]   0.4352959   0.2314276  0.17036086 8.540406e-02 4.235185e-02
#>  [46,]   0.3997271   0.2831977  0.15273260 8.309479e-02 6.421309e-02
#>  [47,]   0.4923128   0.2015941  0.15788862 8.722155e-02 4.561131e-02
#>  [48,]   0.6383924   0.2148016  0.09989394 2.751741e-02 1.939459e-02
#>  [49,]   0.5381248   0.1503689  0.13734123 8.517781e-02 4.820606e-02
#>  [50,]   0.4214426   0.2843251  0.17182884 7.380079e-02 4.860268e-02
#>  [51,]   0.3921653   0.2878657  0.19887777 6.000952e-02 3.491835e-02
#>  [52,]   0.5235773   0.2432220  0.13755884 7.583917e-02 1.763846e-02
#>  [53,]   0.4159153   0.2977848  0.12641377 9.709793e-02 5.400825e-02
#>  [54,]   0.5949528   0.3109743  0.09407292 5.848595e-17 4.125851e-17
#>  [55,]   0.4248137   0.3229257  0.13097752 7.533370e-02 3.502795e-02
#>  [56,]   0.3542456   0.2879291  0.15864883 9.238601e-02 7.982154e-02
#>  [57,]   0.4299331   0.2311713  0.15098035 1.043624e-01 5.574493e-02
#>  [58,]   0.5276266   0.2446880  0.10268255 8.014426e-02 3.791628e-02
#>  [59,]   0.5895161   0.1903545  0.11277815 7.559359e-02 1.702740e-02
#>  [60,]   0.4452629   0.2011510  0.15618467 9.404581e-02 4.672101e-02
#>  [61,]   0.4815827   0.1969736  0.18195856 5.939391e-02 5.252091e-02
#>  [62,]   0.4949486   0.1498090  0.14652228 1.120339e-01 4.622912e-02
#>  [63,]   0.4033691   0.2858485  0.15234538 7.619029e-02 3.951099e-02
#>  [64,]   0.4912459   0.2102950  0.13516963 7.888594e-02 5.005088e-02
#>  [65,]   0.5830524   0.2141138  0.09192128 5.731333e-02 3.214622e-02
#>  [66,]   0.5841292   0.2481867  0.09862802 6.905600e-02 7.020880e-17
#>  [67,]   0.5522172   0.2503627  0.09055410 5.180750e-02 4.732006e-02
#>  [68,]   0.3817576   0.2829117  0.14357699 8.882991e-02 5.743593e-02
#>  [69,]   0.4490826   0.2716681  0.13683096 6.833942e-02 4.822482e-02
#>  [70,]   0.5731139   0.2092996  0.12295467 5.665037e-02 3.093369e-02
#>  [71,]   0.4159624   0.2164706  0.16087507 8.371783e-02 5.237398e-02
#>  [72,]   0.4908029   0.1767142  0.13889163 1.032454e-01 6.884797e-02
#>  [73,]   0.5849356   0.1786329  0.10872334 9.773115e-02 2.997698e-02
#>  [74,]   0.4774281   0.2405576  0.15827258 9.035555e-02 3.338622e-02
#>  [75,]   0.4394754   0.1794655  0.15353649 9.649725e-02 6.099970e-02
#>  [76,]   0.5905780   0.1954173  0.08718509 5.975932e-02 3.967495e-02
#>  [77,]   0.4504323   0.1605727  0.13124436 1.133080e-01 6.073251e-02
#>  [78,]   0.4481720   0.2878261  0.12882768 5.260066e-02 4.537576e-02
#>  [79,]   0.5444785   0.2284807  0.10277128 7.010444e-02 3.805852e-02
#>  [80,]   0.6703401   0.1781928  0.09865586 5.281122e-02 4.349102e-17
#>  [81,]   0.5020618   0.2133638  0.13762444 9.644517e-02 4.074408e-02
#>  [82,]   0.3992883   0.2816862  0.16977407 9.732358e-02 3.387322e-02
#>  [83,]   0.6870248   0.1922971  0.05583666 4.009117e-02 1.551642e-02
#>  [84,]   0.5076674   0.2172800  0.13844253 8.359098e-02 4.354711e-02
#>  [85,]   0.5110162   0.2567152  0.12492493 5.487698e-02 4.148583e-02
#>  [86,]   0.5730677   0.2066644  0.08091388 5.482537e-02 5.216474e-02
#>  [87,]   0.4826427   0.2117358  0.12728580 1.192214e-01 3.937695e-02
#>  [88,]   0.4311971   0.3466452  0.12189304 5.590115e-02 3.621448e-02
#>  [89,]   0.4073453   0.2781044  0.12913225 8.269644e-02 5.441950e-02
#>  [90,]   0.5270677   0.1720153  0.11786329 9.453547e-02 5.632847e-02
#>  [91,]   0.6232156   0.1415517  0.10110293 6.060633e-02 3.978628e-02
#>  [92,]   0.4312122   0.2351552  0.13739682 9.890644e-02 4.589004e-02
#>  [93,]   0.4898709   0.1611251  0.11755261 9.839335e-02 6.618410e-02
#>  [94,]   0.6545704   0.1740710  0.07969253 4.884498e-02 3.033440e-02
#>  [95,]   0.3991165   0.2898566  0.18049306 8.092593e-02 3.342276e-02
#>  [96,]   0.6147427   0.1576517  0.12677932 5.749759e-02 3.941858e-02
#>  [97,]   0.6555610   0.1524550  0.10542713 3.872051e-02 2.864864e-02
#>  [98,]   0.4978523   0.2170870  0.13297566 8.774044e-02 4.500741e-02
#>  [99,]   0.4952511   0.2634634  0.12994044 4.925396e-02 4.606060e-02
#> [100,]   0.6568649   0.1660202  0.11937658 3.699264e-02 1.341117e-02
#>         Dimension 6   Dimension 7   Dimension 8   Dimension 9
#>   [1,] 7.204621e-03  1.946780e-17  9.618626e-18 -5.319488e-19
#>   [2,] 2.462903e-02  1.860150e-17  1.178617e-17  7.704249e-19
#>   [3,] 1.663130e-02  3.702318e-03 -5.380666e-19 -5.827202e-18
#>   [4,] 4.325382e-02  2.155800e-02  8.282515e-03  1.955004e-17
#>   [5,] 2.185745e-02  9.169691e-03  7.819706e-05  3.765644e-17
#>   [6,] 1.376223e-17 -2.107181e-17 -2.584353e-17 -2.592452e-17
#>   [7,] 4.491909e-02  2.294169e-02  2.140034e-05 -1.643906e-18
#>   [8,] 1.987944e-02  2.310722e-17 -7.514299e-19 -2.845116e-18
#>   [9,] 1.057403e-02  8.558141e-17 -1.984082e-18 -4.462487e-18
#>  [10,] 1.253067e-02  4.047223e-17 -5.245794e-18 -1.476778e-17
#>  [11,] 5.761201e-17  1.411186e-17 -4.277612e-18 -1.092516e-17
#>  [12,] 2.135789e-02  1.603410e-17  6.007193e-18  2.386936e-19
#>  [13,] 3.308774e-02  1.204868e-02  9.471629e-03  1.801433e-17
#>  [14,] 7.753715e-03  5.557382e-03  4.229175e-18  3.627805e-18
#>  [15,] 6.355875e-03  5.464127e-17  2.209498e-17  1.827407e-17
#>  [16,] 2.776009e-02  1.378848e-02  5.115650e-03  7.561308e-17
#>  [17,] 1.731016e-02  1.441627e-17 -1.492458e-17 -1.716343e-17
#>  [18,] 4.232232e-02  1.498770e-02  4.644274e-03  1.190846e-17
#>  [19,] 2.259592e-02  1.004401e-02  1.771678e-17  8.599712e-19
#>  [20,] 2.931251e-02  1.290289e-02  2.405362e-17 -2.866688e-18
#>  [21,] 3.689784e-02  1.826643e-02  8.554723e-03  1.267941e-17
#>  [22,] 2.141638e-02  9.876110e-17  4.328101e-18  1.310180e-18
#>  [23,] 6.551842e-17  2.015026e-17  1.094550e-17  5.620120e-18
#>  [24,] 4.623007e-02  1.866718e-02  7.732491e-03  4.403509e-03
#>  [25,] 4.442843e-02  2.073766e-02  8.204968e-03  5.766142e-18
#>  [26,] 2.307763e-02  4.121289e-03 -4.081226e-19 -2.307916e-17
#>  [27,] 5.386261e-02  9.501378e-03  2.966354e-17 -3.827653e-19
#>  [28,] 3.165492e-02  1.440356e-02  3.296141e-17 -1.633928e-17
#>  [29,] 4.305735e-02  1.202957e-02  6.555572e-17  9.198839e-18
#>  [30,] 1.689677e-02  5.890569e-03  1.235670e-17  1.170650e-17
#>  [31,] 3.593721e-02  1.664963e-02  4.794527e-17  6.821191e-18
#>  [32,] 3.810173e-02  1.775716e-02  7.020539e-03  7.073261e-17
#>  [33,] 3.421704e-02  1.800305e-02  9.974259e-03 -5.599122e-18
#>  [34,] 2.603875e-02  1.359354e-03 -4.926763e-18 -2.938415e-17
#>  [35,] 1.599599e-02  2.367713e-17 -7.647780e-18 -1.134281e-17
#>  [36,] 4.017853e-02  1.812842e-02  4.286721e-03 -3.160481e-18
#>  [37,] 1.619967e-02  2.787434e-17  2.375654e-17 -5.584741e-18
#>  [38,] 1.927731e-02  5.501836e-03 -8.022377e-18 -9.589823e-18
#>  [39,] 2.881804e-02  1.115730e-02  6.604420e-03  4.043482e-17
#>  [40,] 2.776832e-02  1.164913e-03  1.008551e-16  1.222441e-17
#>  [41,] 1.778794e-02  8.725316e-03 -2.368287e-18 -1.102152e-17
#>  [42,] 2.849893e-02  1.380732e-02  3.397440e-17  2.338851e-17
#>  [43,] 1.545061e-02  4.418742e-03  3.654289e-17  1.096035e-17
#>  [44,] 8.446740e-03  2.245202e-17  8.461152e-18 -4.393502e-18
#>  [45,] 2.164395e-02  1.351575e-02  3.689685e-17 -4.910541e-18
#>  [46,] 1.703477e-02  1.581705e-17 -5.050046e-18 -3.527215e-17
#>  [47,] 1.537171e-02  2.448772e-17 -1.019882e-18 -1.462148e-17
#>  [48,] 6.821162e-17  1.740866e-17 -2.444492e-18 -4.951422e-18
#>  [49,] 2.941712e-02  1.136410e-02  4.821429e-17 -3.832474e-18
#>  [50,] 2.796409e-17 -4.386726e-18 -5.694675e-18 -6.404343e-18
#>  [51,] 2.616335e-02  1.782223e-17  6.509048e-18  2.947103e-18
#>  [52,] 2.164281e-03  7.860776e-17  1.621541e-17  6.498541e-18
#>  [53,] 8.779918e-03  2.749989e-17 -8.098318e-18 -1.366403e-17
#>  [54,] 1.202867e-17 -2.102378e-18 -7.561195e-18 -2.076795e-17
#>  [55,] 1.092137e-02  7.933420e-17 -3.348844e-18 -1.228502e-17
#>  [56,] 2.519849e-02  1.770412e-03  3.019547e-17  2.006999e-17
#>  [57,] 2.780784e-02  2.891818e-17 -5.742578e-19 -4.370867e-18
#>  [58,] 6.942243e-03  1.504765e-16 -2.539128e-18 -1.900447e-17
#>  [59,] 1.473024e-02  6.338855e-17  3.703483e-17  7.065967e-18
#>  [60,] 3.069584e-02  2.593875e-02  1.324759e-17  9.833611e-18
#>  [61,] 2.757028e-02  4.420616e-17  3.581322e-17  1.650038e-17
#>  [62,] 3.821661e-02  1.224048e-02  4.891228e-18 -1.962824e-17
#>  [63,] 3.332433e-02  8.237803e-03  1.173620e-03  2.658432e-17
#>  [64,] 3.435258e-02  4.439987e-17  2.215315e-17  5.341654e-18
#>  [65,] 1.606864e-02  5.384251e-03  7.251743e-17  1.049295e-17
#>  [66,] 2.021418e-17  4.584473e-18  8.953900e-20 -4.217572e-18
#>  [67,] 7.738362e-03  4.432182e-17  2.144158e-18 -3.905709e-19
#>  [68,] 3.057778e-02  1.491004e-02  2.368990e-17  7.850256e-18
#>  [69,] 2.378572e-02  2.068334e-03 -4.225000e-18 -1.321618e-17
#>  [70,] 7.047801e-03  1.591999e-18  1.160338e-18 -2.404230e-17
#>  [71,] 4.269715e-02  2.790302e-02  4.782876e-17 -8.293143e-18
#>  [72,] 2.149800e-02  6.379349e-17 -6.829449e-18 -6.949573e-18
#>  [73,] 4.240697e-17  2.562057e-18  1.017839e-18 -1.400208e-19
#>  [74,] 3.963579e-17  3.646608e-17 -5.874102e-18 -6.661784e-18
#>  [75,] 4.170332e-02  2.832226e-02  1.084950e-16  1.075702e-17
#>  [76,] 1.940475e-02  7.980629e-03  2.757629e-17 -4.594777e-18
#>  [77,] 4.634518e-02  2.348133e-02  1.388361e-02  4.811089e-17
#>  [78,] 3.085241e-02  6.345440e-03  3.024361e-17  6.407095e-18
#>  [79,] 1.610661e-02  1.892970e-17  1.883278e-17  1.301513e-17
#>  [80,] 2.983053e-17  2.773756e-18 -4.192999e-18 -8.153949e-18
#>  [81,] 9.760762e-03  2.238565e-17 -8.321995e-18 -1.098200e-17
#>  [82,] 1.805462e-02  3.086082e-17  2.045764e-17 -5.282345e-18
#>  [83,] 9.233865e-03  2.412646e-17  1.613002e-17 -2.648250e-18
#>  [84,] 9.471993e-03  3.571053e-17  5.361794e-18 -9.730518e-18
#>  [85,] 1.098088e-02  5.013733e-17  2.612362e-17  1.831416e-17
#>  [86,] 2.054731e-02  8.155209e-03  3.661405e-03 -3.338454e-18
#>  [87,] 1.973737e-02  2.018966e-17  1.752169e-17  6.122912e-18
#>  [88,] 8.149057e-03  9.078722e-17  2.865243e-17  3.334410e-18
#>  [89,] 3.115999e-02  1.714208e-02  2.486414e-17  1.482438e-17
#>  [90,] 2.132434e-02  1.086545e-02  9.685049e-17  1.551394e-17
#>  [91,] 2.491493e-02  8.822274e-03  1.489181e-17  3.080149e-18
#>  [92,] 2.992328e-02  2.151598e-02  2.752428e-17  7.349132e-18
#>  [93,] 4.336511e-02  1.930495e-02  4.203857e-03  4.813035e-17
#>  [94,] 7.494996e-03  4.991659e-03  3.023249e-17  5.985709e-18
#>  [95,] 1.618524e-02 -1.029505e-17 -1.166730e-17 -1.612353e-17
#>  [96,] 3.910079e-03  6.984783e-17  2.310366e-17  3.274196e-18
#>  [97,] 1.448301e-02  4.704637e-03  1.885864e-17 -5.402952e-18
#>  [98,] 1.170490e-02  7.632259e-03  3.003106e-17  2.008531e-18
#>  [99,] 1.603047e-02  1.205431e-16 -1.584361e-18 -7.493372e-18
#> [100,] 7.334477e-03  1.715487e-17  7.299101e-19 -2.007530e-17
#> 
#> $BootMatTausCI
#>                     LowCI       HighCI
#> Dimension 1  3.337572e-01 6.709958e-01
#> Dimension 2  1.353289e-01 3.111002e-01
#> Dimension 3  7.969253e-02 2.093415e-01
#> Dimension 4  2.272586e-02 1.133080e-01
#> Dimension 5  4.349102e-17 7.646862e-02
#> Dimension 6  1.376223e-17 4.623007e-02
#> Dimension 7 -1.029505e-17 2.593875e-02
#> Dimension 8 -1.492458e-17 9.471629e-03
#> Dimension 9 -2.938415e-17 7.073261e-17
#> 
#> $BootstrappedEstInertia
#>   [1] 109.38411  67.84646 106.83156  91.28769 103.46672 102.04038  90.08235
#>   [8] 122.63478  88.70121  77.92331 101.30457 114.27321  94.15832 111.10106
#>  [15] 106.26559 105.18525  96.51346 117.05194 111.09497  92.91477  85.33111
#>  [22] 124.02107 119.93939 101.76878  94.76345 130.72865  89.24715 108.57912
#>  [29] 108.38578  96.50314 109.87655  84.05770  83.56696 124.17276 110.72218
#>  [36] 112.85568  76.71669  98.22470 106.16544 106.00737 101.24544 102.70510
#>  [43] 116.45006 139.00958 119.88478 105.46830  84.01775  92.72386  97.38947
#>  [50]  99.87966 114.02502  99.64390 112.35051  72.44965  82.51821 102.08425
#>  [57]  94.46294 130.23802 111.15333  97.63299  61.11427 100.03794 101.27362
#>  [64]  75.56671 112.40784 102.92044 110.98262 105.59017 107.91406 104.62138
#>  [71]  87.59200  89.41475 126.94982 117.17634  86.26686 102.36586 100.82234
#>  [78] 100.41810  81.70759  92.06228  89.34956  92.86401 124.02859  92.53593
#>  [85]  88.59105 104.18213  84.04935  94.01219 103.02824  79.26398 128.58778
#>  [92] 108.08804 110.04661 114.36641 106.25217 115.24260 115.93673 112.05575
#>  [99] 106.91463 116.90090
#>