Boostrap for eigenvalues.

Boostrap for eigenvalues.

Usage

Boot4Eigs(
  data,
  center = FALSE,
  scale = FALSE,
  design = NULL,
  niter = 100,
  CI.perc = c(0.025, 0.975),
  suppressProgressBar = TRUE
)

Arguments

data

the data-set;

center

whether to center the data (default to FALSE), see expo.scale;

scale

whether to scale the data ( default to FALSE), see expo.scale;

design

whether there is a design;

niter

number of bootstrap iterations;

CI.perc

vector of the two levels of confidence for the left and right limits of the bootstrap confidence interval (default to c(0.025, 0.0975));

suppressProgressBar

whether to show a progress bar or not (default to FALSE).

Value

list containing four objects

Examples

Boot4Eigs(matrix(rnorm(100), 10, 10))
#> $BootMatEigs
#>        Dimension 1 Dimension 2 Dimension 3  Dimension 4  Dimension 5
#>   [1,]    37.70757    22.64811   17.150800 7.501818e+00 4.991701e+00
#>   [2,]    39.86643    23.86417   20.011507 6.257898e+00 4.131062e-15
#>   [3,]    36.76842    22.00733   13.980199 8.236502e+00 5.944128e+00
#>   [4,]    47.28480    25.95947    9.930822 3.644631e+00 3.180273e+00
#>   [5,]    42.47178    23.98813   14.950237 6.175989e+00 2.413857e+00
#>   [6,]    39.53473    18.60535   14.906719 1.205483e+01 4.898375e+00
#>   [7,]    35.33464    16.32678   13.413793 1.282054e+01 7.395882e+00
#>   [8,]    32.73519    24.66722   14.920884 1.101943e+01 3.769863e+00
#>   [9,]    54.95224    22.41676    8.941779 3.689217e+00 3.753428e-15
#>  [10,]    39.57674    21.19651   14.384207 8.401031e+00 4.272761e+00
#>  [11,]    30.61820    23.15441   15.973607 1.125140e+01 6.145071e+00
#>  [12,]    46.30856    25.70492   13.616299 4.370226e+00 3.447943e-15
#>  [13,]    42.42842    25.43593   22.135649 1.239684e-14 1.881237e-15
#>  [14,]    31.34971    18.45439   14.622378 1.215119e+01 8.114105e+00
#>  [15,]    30.62249    21.99967   14.840311 1.186807e+01 6.072915e+00
#>  [16,]    28.39221    19.14950   15.504910 1.162761e+01 9.817668e+00
#>  [17,]    34.89086    20.41271   17.467265 1.205240e+01 4.300108e+00
#>  [18,]    39.59343    25.95473   16.105469 5.990954e+00 2.355421e+00
#>  [19,]    30.18622    18.09474   16.688044 1.384180e+01 1.067080e+01
#>  [20,]    31.50180    23.03063   14.187610 1.326202e+01 4.717754e+00
#>  [21,]    42.88010    31.43179   15.688115 1.121387e-14 4.367535e-15
#>  [22,]    33.89904    29.20984   15.119159 8.892272e+00 2.879687e+00
#>  [23,]    33.66084    20.63519   13.744898 1.053580e+01 6.722617e+00
#>  [24,]    33.40169    23.97308   11.758288 9.017278e+00 6.000497e+00
#>  [25,]    42.18733    22.11892   14.286317 6.303112e+00 3.201811e+00
#>  [26,]    39.72961    19.21688   13.443287 1.083802e+01 5.808514e+00
#>  [27,]    35.08659    19.50237   14.147729 1.228956e+01 5.159209e+00
#>  [28,]    33.89279    25.43243   15.448680 1.102488e+01 3.528285e+00
#>  [29,]    29.66740    26.26291   16.677697 1.032540e+01 4.107143e+00
#>  [30,]    41.19286    23.46550   16.623506 5.965154e+00 2.752985e+00
#>  [31,]    32.28544    18.77075   14.154091 1.170974e+01 8.995622e+00
#>  [32,]    29.68962    20.46981   15.666265 1.124149e+01 9.857527e+00
#>  [33,]    32.13094    27.34380   16.887882 7.589582e+00 4.468314e+00
#>  [34,]    36.68605    19.14450   14.004857 8.340220e+00 6.546533e+00
#>  [35,]    39.90969    21.22121   12.801714 7.149575e+00 6.314029e+00
#>  [36,]    32.57013    28.61896   18.664194 8.367844e+00 1.778875e+00
#>  [37,]    37.41362    24.95035   15.133800 7.060370e+00 5.441856e+00
#>  [38,]    33.78085    18.40410   12.804652 1.034148e+01 8.457365e+00
#>  [39,]    38.57373    18.04987   16.826720 1.119601e+01 3.156120e+00
#>  [40,]    37.91682    19.01813   13.437908 7.953208e+00 6.781182e+00
#>  [41,]    42.27916    15.28131   14.645615 1.205741e+01 4.794184e+00
#>  [42,]    44.82731    22.28599   15.772706 7.113991e+00 4.751880e-15
#>  [43,]    27.53925    20.25061   14.878812 1.143907e+01 8.960320e+00
#>  [44,]    40.64060    17.63803   15.698558 1.058200e+01 5.440816e+00
#>  [45,]    37.51230    33.35986   15.590387 3.537460e+00 1.209232e-15
#>  [46,]    46.24880    27.07356   10.535306 6.142337e+00 3.140014e-15
#>  [47,]    35.35846    25.92660   16.342507 1.039770e+01 1.974732e+00
#>  [48,]    32.36455    26.75447   12.349710 9.511410e+00 5.413049e+00
#>  [49,]    35.03948    24.59047   12.369667 9.088828e+00 5.596063e+00
#>  [50,]    37.23273    25.72783   16.637177 6.857857e+00 3.544401e+00
#>  [51,]    39.61460    27.02044   17.782619 5.582338e+00 3.069570e-15
#>  [52,]    33.80581    22.88940   14.947723 9.292395e+00 7.210782e+00
#>  [53,]    33.04642    24.99827   14.760705 7.717180e+00 5.300473e+00
#>  [54,]    36.69632    26.01008   18.669752 5.671120e+00 2.289488e+00
#>  [55,]    35.21588    24.96972   13.005214 1.053457e+01 6.274605e+00
#>  [56,]    43.03597    25.54470   21.419326 4.683018e-15 3.641957e-15
#>  [57,]    35.49818    21.52870   19.594585 7.180272e+00 5.095368e+00
#>  [58,]    39.01021    22.23072   16.634476 9.630310e+00 2.494287e+00
#>  [59,]    32.40779    25.92076   16.034747 8.096271e+00 6.224205e+00
#>  [60,]    35.98205    17.43457   14.459429 1.038700e+01 9.110522e+00
#>  [61,]    35.45937    29.12739   12.018779 1.078653e+01 2.607939e+00
#>  [62,]    30.39544    25.28956   18.248164 1.284009e+01 3.226749e+00
#>  [63,]    38.47060    21.14120   11.040049 1.000491e+01 5.496162e+00
#>  [64,]    29.66487    20.07269   19.254055 1.452182e+01 4.877263e+00
#>  [65,]    26.77430    20.92387   18.187115 1.137524e+01 8.869853e+00
#>  [66,]    39.31185    20.59631   17.360569 1.064203e+01 1.160529e+00
#>  [67,]    37.03858    24.11928   15.566331 9.982508e+00 3.293299e+00
#>  [68,]    32.56591    25.74179   19.555924 8.457135e+00 3.499051e+00
#>  [69,]    37.32564    26.10849   14.751500 8.581929e+00 3.232431e+00
#>  [70,]    40.76202    23.45950   14.619693 8.599386e+00 2.559400e+00
#>  [71,]    32.86126    22.33830   19.902839 1.032055e+01 4.026885e+00
#>  [72,]    33.81339    23.69026   13.642654 1.177511e+01 4.136442e+00
#>  [73,]    37.46430    19.00898   16.921175 8.173490e+00 6.063116e+00
#>  [74,]    35.66203    20.72669   15.018790 9.061800e+00 5.396662e+00
#>  [75,]    36.85113    30.75415   15.726844 6.667879e+00 4.132884e-15
#>  [76,]    38.73332    25.53741   11.147850 7.696131e+00 6.198824e+00
#>  [77,]    34.85171    20.86508   13.398216 1.046028e+01 6.433487e+00
#>  [78,]    42.18977    23.08357   14.235160 9.398533e+00 1.092966e+00
#>  [79,]    32.24470    19.55800   14.419981 1.275014e+01 7.618830e+00
#>  [80,]    30.55070    25.14639   17.659369 1.099752e+01 3.640346e+00
#>  [81,]    34.30651    24.71941   15.071467 9.227199e+00 5.645816e+00
#>  [82,]    32.29252    26.24543   21.431999 8.601784e+00 1.428273e+00
#>  [83,]    46.80769    22.08814   13.683418 5.110163e+00 2.310581e+00
#>  [84,]    56.02522    19.35081    8.262734 3.613007e+00 2.748225e+00
#>  [85,]    32.93788    22.10187   16.979473 1.029480e+01 7.217867e+00
#>  [86,]    45.52171    27.59081   13.679391 3.208093e+00 1.165880e-15
#>  [87,]    40.67611    22.63392   17.936219 8.753744e+00 4.410483e-16
#>  [88,]    32.65243    18.71867   14.996633 1.111849e+01 7.872518e+00
#>  [89,]    34.95860    19.64768   14.432929 1.138974e+01 8.463275e+00
#>  [90,]    38.84736    25.02339   15.267001 7.013732e+00 3.848518e+00
#>  [91,]    37.27441    22.19542   14.371383 1.158983e+01 4.568957e+00
#>  [92,]    34.16286    23.15544   17.079245 1.096250e+01 3.387654e+00
#>  [93,]    44.78625    19.36908   16.799592 7.056611e+00 1.988467e+00
#>  [94,]    34.17245    18.97470   13.702664 9.699090e+00 8.458837e+00
#>  [95,]    48.26371    22.52345   12.218646 6.994200e+00 3.610125e-15
#>  [96,]    46.39770    29.05984   10.927585 3.614870e+00 3.107304e-15
#>  [97,]    38.92147    24.55240   20.539218 5.986920e+00 6.669532e-15
#>  [98,]    32.27800    25.38801   17.457498 9.626149e+00 5.250337e+00
#>  [99,]    45.23416    24.76222   16.834747 3.168871e+00 6.155138e-15
#> [100,]    41.31826    27.30416   13.876167 7.501408e+00 5.500988e-15
#>          Dimension 6   Dimension 7   Dimension 8   Dimension 9  Dimension 10
#>   [1,]  3.934461e-15  2.571267e-16 -1.131023e-16 -6.391286e-16 -1.745651e-15
#>   [2,]  1.249481e-15  3.275595e-16 -5.032970e-16 -3.105587e-15 -3.913233e-15
#>   [3,]  2.787363e+00  2.760551e-01  1.056763e-15 -8.668981e-16 -1.397288e-15
#>   [4,]  4.140046e-15  2.472391e-16 -6.155691e-16 -6.266640e-16 -1.976312e-15
#>   [5,]  2.278132e-15 -6.131860e-16 -7.449734e-16 -3.811192e-15 -5.414040e-15
#>   [6,]  3.589697e-16  1.782441e-17 -1.731467e-17 -2.290818e-15 -4.834104e-15
#>   [7,]  2.599002e+00  2.109364e+00  3.520275e-15  1.006689e-15 -4.024781e-15
#>   [8,]  2.887408e+00  8.816915e-16 -9.975636e-16 -1.447283e-15 -7.429784e-15
#>   [9,]  2.579275e-15  1.443227e-15  2.102589e-16  1.146769e-16 -3.533684e-15
#>  [10,]  2.168748e+00  2.619489e-15  2.615203e-15 -2.897263e-15 -2.998781e-15
#>  [11,]  2.857315e+00  6.557505e-16  6.441039e-16 -1.286183e-15 -7.432614e-15
#>  [12,]  1.521809e-15  5.639297e-16 -1.102855e-15 -1.475980e-15 -4.987360e-15
#>  [13,] -3.481533e-16 -6.526781e-16 -6.544827e-16 -9.039565e-16 -1.685338e-15
#>  [14,]  4.132627e+00  1.175605e+00  7.861520e-16  3.687079e-16 -1.738059e-15
#>  [15,]  4.596545e+00  5.416846e-15  1.624260e-15  4.779401e-16 -2.453062e-15
#>  [16,]  4.013271e+00  1.341594e+00  1.532343e-01 -2.078018e-15 -2.609024e-15
#>  [17,]  8.766581e-01  3.375377e-16 -6.076484e-16 -3.613461e-15 -7.520288e-15
#>  [18,]  1.034843e-14  2.068499e-15  6.613488e-16 -1.215972e-16 -9.350002e-16
#>  [19,]  5.183960e-01  1.715380e-15  9.456668e-16  5.555997e-16 -1.221459e-15
#>  [20,]  2.751100e+00  5.490742e-01  5.858620e-16 -4.622352e-16 -5.031333e-16
#>  [21,]  2.350365e-15  1.640433e-15  7.582840e-16  9.272185e-17 -2.431914e-15
#>  [22,]  2.126259e-16 -7.705684e-16 -1.206889e-15 -2.004088e-15 -5.184630e-15
#>  [23,]  3.282748e+00  1.417909e+00  4.559572e-16 -9.932154e-16 -4.683367e-15
#>  [24,]  5.484978e+00  3.641923e-01  3.018353e-15  1.217993e-15  8.474382e-16
#>  [25,]  1.902511e+00  6.339061e-15  3.508603e-15  2.756855e-15 -5.007688e-16
#>  [26,]  9.636853e-01  4.497594e-15  1.011056e-15 -3.014466e-15 -3.111100e-15
#>  [27,]  3.814541e+00  7.886691e-15  1.187374e-16 -9.977285e-16 -3.732546e-15
#>  [28,]  6.729334e-01  2.672655e-15  1.620996e-15  6.597493e-16 -2.880174e-15
#>  [29,]  2.959456e+00  3.173890e-16 -7.589672e-17 -1.199840e-15 -4.042055e-15
#>  [30,]  1.093262e-15 -3.902036e-17 -5.482039e-16 -2.458472e-15 -3.221639e-15
#>  [31,]  3.487738e+00  5.966261e-01  3.567439e-15  2.881608e-15 -1.316245e-15
#>  [32,]  2.397160e+00  6.781228e-01  4.298474e-15  2.813523e-15 -2.597002e-15
#>  [33,]  1.579489e+00  5.634501e-15  1.834962e-15 -2.879145e-15 -3.280315e-15
#>  [34,]  5.045251e+00  2.325839e-01  3.011912e-15 -1.768789e-16 -2.418773e-15
#>  [35,]  2.603786e+00  6.627476e-16 -6.530219e-16 -1.060753e-15 -4.711573e-15
#>  [36,]  2.559696e-15 -7.445789e-18 -6.319053e-16 -1.024489e-15 -2.657248e-15
#>  [37,] -4.761831e-16 -5.320886e-16 -1.129202e-15 -4.866221e-15 -9.294941e-15
#>  [38,]  5.840532e+00  3.710267e-01  5.628997e-15 -6.864262e-16 -9.364191e-16
#>  [39,]  2.197546e+00  1.971949e-15 -1.193068e-15 -1.632277e-15 -4.354889e-15
#>  [40,]  3.269824e+00  1.622923e+00  3.305720e-16 -1.175991e-15 -4.402634e-15
#>  [41,]  9.423134e-01  3.222507e-15  9.550896e-16  3.396660e-16 -3.380848e-15
#>  [42,]  1.314458e-15  1.293326e-15  5.232130e-16 -1.385642e-15 -1.856672e-15
#>  [43,]  4.366016e+00  2.565914e+00  1.275135e-15  8.430238e-16 -1.989462e-15
#>  [44,]  3.192002e-15  2.820197e-15  2.432536e-15  1.435919e-15 -3.523179e-15
#>  [45,]  1.036738e-15  5.937202e-16 -1.113723e-16 -7.664081e-16 -2.968795e-15
#>  [46,] -7.300578e-16 -2.076253e-15 -2.513201e-15 -3.378442e-15 -5.115134e-15
#>  [47,]  4.941543e-15  3.000264e-15  2.063304e-15 -7.648117e-17 -7.033053e-15
#>  [48,]  2.702294e+00  9.045178e-01  5.574536e-16  1.704177e-16 -2.982977e-15
#>  [49,]  3.315499e+00  9.487818e-15  1.271893e-15 -4.358135e-16 -1.219983e-15
#>  [50,]  1.756755e-15 -6.347939e-16 -1.196398e-15 -1.898143e-15 -2.194610e-15
#>  [51,]  2.032325e-15  2.023665e-15  4.388469e-18 -6.647521e-16 -2.190036e-15
#>  [52,]  1.853894e+00  5.660441e-15  6.250028e-16 -5.044852e-16 -1.510410e-15
#>  [53,]  2.821650e+00  1.355298e+00  4.750643e-15  1.464633e-16 -4.913524e-16
#>  [54,]  6.632478e-01  2.924894e-15 -2.899487e-16 -4.822774e-16 -2.863655e-15
#>  [55,]  1.681317e-15  1.654436e-15  1.528789e-17 -4.863395e-15 -8.454575e-15
#>  [56,]  7.909952e-16  7.686653e-16 -4.972581e-17 -3.389457e-16 -4.150238e-15
#>  [57,]  1.102897e+00  1.520177e-15  1.487739e-15 -2.789832e-16 -2.970078e-15
#>  [58,]  1.183071e-15  2.186167e-16 -8.996551e-16 -3.124997e-15 -4.044635e-15
#>  [59,]  1.316229e+00  9.036216e-16  7.157093e-16 -1.252869e-15 -4.326198e-15
#>  [60,]  2.626430e+00  5.737471e-15  1.454560e-15  6.776684e-16  2.874552e-16
#>  [61,]  1.901573e-15  1.773722e-15  3.966332e-16  2.602768e-16 -4.406316e-16
#>  [62,]  1.913778e-15 -9.106810e-16 -1.195687e-15 -1.931359e-15 -2.687859e-15
#>  [63,]  3.847075e+00  4.752434e-15 -1.217341e-15 -2.404967e-15 -2.456225e-15
#>  [64,]  1.609307e+00  2.099395e-15  5.724852e-16 -4.549879e-16 -1.015518e-15
#>  [65,]  3.869619e+00  1.970042e-15  8.688401e-16  8.684053e-16 -2.128062e-15
#>  [66,]  9.287070e-01  4.034718e-15  8.825423e-16 -3.563778e-16 -8.031136e-16
#>  [67,]  5.270623e-16  3.498994e-16  3.049189e-16 -9.755161e-16 -5.367376e-15
#>  [68,]  1.801883e-01 -5.441395e-16 -2.362355e-15 -2.552659e-15 -2.906619e-15
#>  [69,]  5.436753e-15  3.311142e-15  1.932228e-15  1.210695e-15 -1.564797e-16
#>  [70,]  4.310644e-15  3.620007e-15 -1.913479e-15 -1.985978e-15 -2.270811e-15
#>  [71,]  5.501649e-01  1.917045e-15 -5.039640e-16 -1.396764e-15 -1.684078e-15
#>  [72,]  2.942149e+00  3.811829e-15  3.217231e-15 -3.984915e-16 -3.817475e-15
#>  [73,]  2.368944e+00  1.227484e-15 -3.798968e-16 -2.055955e-15 -4.339857e-15
#>  [74,]  2.850462e+00  1.283562e+00  5.112646e-15  1.139904e-15  6.837283e-16
#>  [75,]  3.634996e-15  1.748258e-15 -3.151027e-15 -3.421958e-15 -4.322389e-15
#>  [76,]  6.864688e-01  4.284376e-15  1.621895e-15  4.850498e-16 -8.526906e-16
#>  [77,]  3.991228e+00  3.796223e-15 -2.418653e-15 -2.483758e-15 -3.209586e-15
#>  [78,]  7.892600e-15  4.584889e-15  1.694146e-15  7.535301e-16  2.949069e-16
#>  [79,]  2.799470e+00  6.088775e-01  2.181621e-15 -1.776660e-15 -5.189070e-15
#>  [80,]  2.005677e+00 -8.728674e-16 -1.250755e-15 -2.501554e-15 -3.241464e-15
#>  [81,]  1.029603e+00  5.692924e-16 -8.296049e-16 -1.690971e-15 -4.412588e-15
#>  [82,]  4.295594e-15  1.039882e-15 -8.946183e-16 -2.766564e-15 -4.392967e-15
#>  [83,]  3.293137e-15  2.655360e-15  3.814631e-16  3.232569e-16 -5.696543e-15
#>  [84,]  3.411897e-15  1.530108e-15  7.059062e-16 -5.159134e-16 -1.380383e-15
#>  [85,]  4.681137e-01  6.966243e-16  6.224009e-16 -1.927917e-15 -4.662820e-15
#>  [86,]  1.082891e-15  5.726576e-16  3.927357e-16 -4.888045e-16 -5.346648e-15
#>  [87,]  2.731454e-16 -6.751630e-16 -1.729930e-15 -1.989672e-15 -2.007017e-15
#>  [88,]  4.048422e+00  5.928401e-01  6.554309e-15  2.668037e-15 -1.591286e-15
#>  [89,]  1.107772e+00  4.214624e-15  1.558139e-16 -1.121014e-17 -1.653852e-15
#>  [90,]  9.603629e-15  1.971148e-15  1.958085e-15  1.426020e-15 -6.232175e-16
#>  [91,]  6.941352e-15  1.854247e-15 -1.095284e-15 -1.526120e-15 -2.940481e-15
#>  [92,]  1.252297e+00  2.395131e-15 -1.501220e-16 -2.089162e-15 -2.195954e-15
#>  [93,]  2.012396e-15  1.828910e-15 -2.651352e-16 -3.334154e-15 -3.569448e-15
#>  [94,]  3.830291e+00  1.161965e+00  6.608434e-16 -1.218093e-15 -4.259321e-15
#>  [95,]  3.097025e-15  2.820107e-15  5.147234e-16 -2.455431e-15 -3.092929e-15
#>  [96,]  1.679135e-15  1.015377e-15  3.914898e-16 -9.106708e-16 -2.264456e-15
#>  [97,]  1.013248e-15  8.429569e-16 -1.019127e-16 -8.885111e-16 -3.340567e-15
#>  [98,]  7.726182e-15  1.340109e-15  1.319402e-15 -2.301771e-15 -2.761290e-15
#>  [99,]  1.473468e-15  1.443975e-15 -2.866861e-16 -4.833886e-16 -1.514550e-15
#> [100,]  2.227243e-15  1.285364e-15  6.489157e-16 -4.991674e-16 -5.142965e-15
#> 
#> $BootMatEigsCI
#>                      LowCI       HighCI
#> Dimension 1   2.753925e+01 4.826371e+01
#> Dimension 2   1.632678e+01 3.075415e+01
#> Dimension 3   8.941779e+00 2.141933e+01
#> Dimension 4   1.121387e-14 1.326202e+01
#> Dimension 5   1.165880e-15 9.817668e+00
#> Dimension 6  -4.761831e-16 5.045251e+00
#> Dimension 7  -9.106810e-16 1.622923e+00
#> Dimension 8  -2.513201e-15 5.628997e-15
#> Dimension 9  -4.863395e-15 2.756855e-15
#> Dimension 10 -8.454575e-15 2.949069e-16
#> 
#> $FixedEigs
#>  [1] 31.338313390 25.714631559 18.265014598 11.275316823  9.469796383
#>  [6]  5.906678740  2.763535539  2.477379693  0.575917979  0.005361466
#> 
#> $BootstrappedEstEigs
#>   Dimension 1   Dimension 2   Dimension 3   Dimension 4   Dimension 5 
#>  3.712862e+01  2.310442e+01  1.531404e+01  8.708023e+00  4.173617e+00 
#>   Dimension 6   Dimension 7   Dimension 8   Dimension 9  Dimension 10 
#>  1.377670e+00  1.920805e-01  1.532343e-03 -9.334800e-16 -3.050022e-15 
#>